Struct boa::builtins::bigint::BigInt

pub struct BigInt(BigInt);

BigInt implementation.

Implementations

impl BigInt

pub(crate) fn from_string(
string: &str,
context: &mut Context
) -> Result<Self, Value>

This function takes a string and conversts it to BigInt type.

More information:

ECMAScript reference

pub fn from_string_radix(buf: &str, radix: u32) -> Option<Self>

Converts a string to a BigInt with the specified radix.

pub fn to_string_radix(&self, radix: u32) -> String

Convert bigint to string with radix.

pub fn to_f64(&self) -> f64

Converts the BigInt to a f64 type.

Returns f64::INFINITY if the BigInt is too big.

pub(crate) fn from_str(string: &str) -> Option<Self>

impl BigInt

pub(crate) fn same_value_zero(x: &Self, y: &Self) -> bool

Checks for SameValueZero equality.

More information:

ECMAScript reference

pub(crate) fn same_value(x: &Self, y: &Self) -> bool

Checks for SameValue equality.

More information:

ECMAScript reference

pub(crate) fn equal(x: &Self, y: &Self) -> bool

Checks for mathematical equality.

The abstract operation BigInt::equal takes arguments x (a BigInt) and y (a BigInt). It returns true if x and y have the same mathematical integer value and false otherwise.

More information:

ECMAScript reference

impl BigInt

pub fn pow(self, other: &Self) -> Result<Self, String>

pub fn shift_right(self, other: Self) -> Result<Self, String>

pub fn shift_left(self, other: Self) -> Result<Self, String>

pub fn mod_floor(self, other: &Self) -> Self

Floored integer modulo.

Examples

assert_eq!((8).mod_floor(&3), 2);
assert_eq!((8).mod_floor(&-3), -1);

impl BigInt

pub(crate) const LENGTH: usize

The amount of arguments this function object takes.

fn constructor(
_: &Value,
args: &[Value],
context: &mut Context
) -> Result<Value>

BigInt()

The BigInt() constructor is used to create BigInt objects.

More information:

ECMAScript reference

MDN documentation

fn this_bigint_value(value: &Value, context: &mut Context) -> Result<RcBigInt>

The abstract operation thisBigIntValue takes argument value.

The phrase “this BigInt value” within the specification of a method refers to the result returned by calling the abstract operation thisBigIntValue with the this value of the method invocation passed as the argument.

More information:

ECMAScript reference

pub(crate) fn to_string(
this: &Value,
args: &[Value],
context: &mut Context
) -> Result<Value>

BigInt.prototype.toString( [radix] )

The toString() method returns a string representing the specified BigInt object.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn value_of(
this: &Value,
_: &[Value],
context: &mut Context
) -> Result<Value>

BigInt.prototype.valueOf()

The valueOf() method returns the wrapped primitive value of a Number object.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn as_int_n(
_: &Value,
args: &[Value],
context: &mut Context
) -> Result<Value>

BigInt.asIntN()

The BigInt.asIntN() method wraps the value of a BigInt to a signed integer between -2**(width - 1) and 2**(width-1) - 1.

pub(crate) fn as_uint_n(
_: &Value,
args: &[Value],
context: &mut Context
) -> Result<Value>

BigInt.asUintN()

The BigInt.asUintN() method wraps the value of a BigInt to an unsigned integer between 0 and 2**(width) - 1.

fn calculate_as_uint_n(
args: &[Value],
context: &mut Context
) -> Result<(BigInt, u32)>

Helper function to wrap the value of a BigInt to an unsigned integer.

This function expects the same arguments as as_uint_n and wraps the value of a BigInt. Additionally to the wrapped unsigned value it returns the converted bits argument, so it can be reused from the as_int_n method.