Module boa::builtins::console

This module implements the global console object.

The console object can be accessed from any global object.

The specifics of how it works varies from browser to browser, but there is a de facto set of features that are typically provided.

More information:

MDN documentation

WHATWG console specification

Structs

ConsoleState
This is the internal console object state.

Enums

LogMessage
This represents the different types of log messages.

Functions

assert
console.assert(condition, ...data)

clear
console.clear()

count
console.count(label)

count_reset
console.countReset(label)

create_constructor
Create a new console object

debug
console.debug(...data)

dir
console.dir(item, options)

error
console.error(...data)

formatter
This represents the console formatter.

get_arg_at_index
Helper function that returns the argument at a specified index.

group
console.group(...data)

group_end
console.groupEnd(label)

info
console.info(...data)

log
console.log(...data)

logger
Helper function for logging messages.

system_time_in_ms
Returns current system time in ms.

time
console.time(label)

time_end
console.timeEnd(label)

time_log
console.timeLog(label, ...data)

trace
console.trace(...data)

warn
console.warn(...data)