Trait boa::builtins::function::ClosureFunctionSignature

pub trait ClosureFunctionSignature: Fn(&JsValue, &[JsValue], Captures, &mut Context) -> JsResult<JsValue> + DynCopy + DynClone + 'static { }

Trait representing a native built-in closure.

Closures need to have this signature in order to be callable from Javascript, but most of the time the compiler is smart enough to correctly infer the types.