Struct boa::builtins::map::ordered_map::OrderedMap

pub struct OrderedMap<V, S = RandomState> {
    map: IndexMap<MapKey, Option<V>, S>,
    lock: u32,
    empty_count: usize,
}

A newtype wrapping indexmap::IndexMap

Fields
map: IndexMap<MapKey, Option<V>, S>lock: u32empty_count: usize
Implementations

impl<V> OrderedMap<V>

pub fn new() -> Self

pub fn with_capacity(capacity: usize) -> Self

pub fn full_len(&self) -> usize

Return the number of key-value pairs in the map, including empty values.

Computes in O(1) time.

pub fn len(&self) -> usize

Gets the number of key-value pairs in the map, not including empty values.

Computes in O(1) time.

pub fn is_empty(&self) -> bool

Returns true if the map contains no elements.

Computes in O(1) time.

pub fn insert(&mut self, key: JsValue, value: V) -> Option<V>

Insert a key-value pair in the map.

If an equivalent key already exists in the map: the key remains and retains in its place in the order, its corresponding value is updated with value and the older value is returned inside Some(_).

If no equivalent key existed in the map: the new key-value pair is inserted, last in order, and None is returned.

Computes in O(1) time (amortized average).

pub fn remove(&mut self, key: &JsValue) -> Option<V>

Remove the key-value pair equivalent to key and return its value.

Like Vec::remove, the pair is removed by shifting all of the elements that follow it, preserving their relative order. This perturbs the index of all of those elements!

Return None if key is not in map.

Computes in O(n) time (average).

pub fn get(&self, key: &JsValue) -> Option<&V>

Return a reference to the value stored for key, if it is present, else None.

Computes in O(1) time (average).

pub fn get_index(&self, index: usize) -> Option<(&JsValue, &V)>

Get a key-value pair by index Valid indices are 0 <= index < self.full_len() Computes in O(1) time.

pub fn iter(&self) -> impl Iterator<Item = (&JsValue, &V)>

Return an iterator over the key-value pairs of the map, in their order

pub fn contains_key(&self, key: &JsValue) -> bool

Return true if an equivalent to key exists in the map.

Computes in O(1) time (average).

pub(crate) fn lock(&mut self, map: JsObject) -> MapLock

Increases the lock counter and returns a lock object that will decrement the counter when dropped.

This allows objects to be removed from the map during iteration without affecting the indexes until the iteration has completed.

fn unlock(&mut self)

Decreases the lock counter and, if 0, removes all empty entries.