Struct boa::builtins::map::Map

pub(crate) struct Map(OrderedMap<Value, Value>);

Implementations

impl Map

pub(crate) const LENGTH: usize

pub(crate) fn constructor(
new_target: &Value,
args: &[Value],
context: &mut Context
) -> Result<Value>

Create a new map

fn get_species(this: &Value, _: &[Value], _: &mut Context) -> Result<Value>

get Map [ @@species ]

The Map [ @@species ] accessor property returns the Map constructor.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn entries(
this: &Value,
_: &[Value],
context: &mut Context
) -> Result<Value>

Map.prototype.entries()

Returns a new Iterator object that contains the [key, value] pairs for each element in the Map object in insertion order.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn keys(
this: &Value,
_: &[Value],
context: &mut Context
) -> Result<Value>

Map.prototype.keys()

Returns a new Iterator object that contains the keys for each element in the Map object in insertion order.

More information:

ECMAScript reference

MDN documentation

fn set_size(this: &Value, size: usize)

Helper function to set the size property.

pub(crate) fn set(
this: &Value,
args: &[Value],
context: &mut Context
) -> Result<Value>

Map.prototype.set( key, value )

This method associates the value with the key. Returns the map object.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn delete(
this: &Value,
args: &[Value],
context: &mut Context
) -> Result<Value>

Map.prototype.delete( key )

This method removes the element associated with the key, if it exists. Returns true if there was an element, false otherwise.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn get(
this: &Value,
args: &[Value],
context: &mut Context
) -> Result<Value>

Map.prototype.get( key )

This method returns the value associated with the key, or undefined if there is none.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn clear(this: &Value, _: &[Value], _: &mut Context) -> Result<Value>

Map.prototype.clear( )

This method removes all entries from the map.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn has(
this: &Value,
args: &[Value],
context: &mut Context
) -> Result<Value>

Map.prototype.has( key )

This method checks if the map contains an entry with the given key.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn for_each(
this: &Value,
args: &[Value],
context: &mut Context
) -> Result<Value>

Map.prototype.forEach( callbackFn [ , thisArg ] )

This method executes the provided callback function for each key-value pair in the map.

More information:

ECMAScript reference

MDN documentation

fn get_size(map: &Value, context: &mut Context) -> Result<usize>

Helper function to get the size of the map.

pub(crate) fn values(
this: &Value,
_: &[Value],
context: &mut Context
) -> Result<Value>

Map.prototype.values()

Returns a new Iterator object that contains the values for each element in the Map object in insertion order.

More information:

ECMAScript reference

MDN documentation

fn get_key_value(
value: &Value,
context: &mut Context
) -> Result<Option<(Value, Value)>>

Helper function to get a key-value pair from an array.