Struct boa::builtins::math::Math

pub(crate) struct Math;

Javascript Math object.

Implementations

impl Math

pub(crate) fn abs(
_: &Value,
args: &[Value],
context: &mut Context
) -> Result<Value>

Get the absolute value of a number.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn acos(
_: &Value,
args: &[Value],
context: &mut Context
) -> Result<Value>

Get the arccos of a number.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn acosh(
_: &Value,
args: &[Value],
context: &mut Context
) -> Result<Value>

Get the hyperbolic arccos of a number.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn asin(
_: &Value,
args: &[Value],
context: &mut Context
) -> Result<Value>

Get the arcsine of a number.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn asinh(
_: &Value,
args: &[Value],
context: &mut Context
) -> Result<Value>

Get the hyperbolic arcsine of a number.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn atan(
_: &Value,
args: &[Value],
context: &mut Context
) -> Result<Value>

Get the arctangent of a number.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn atanh(
_: &Value,
args: &[Value],
context: &mut Context
) -> Result<Value>

Get the hyperbolic arctangent of a number.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn atan2(
_: &Value,
args: &[Value],
context: &mut Context
) -> Result<Value>

Get the arctangent of a numbers.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn cbrt(
_: &Value,
args: &[Value],
context: &mut Context
) -> Result<Value>

Get the cubic root of a number.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn ceil(
_: &Value,
args: &[Value],
context: &mut Context
) -> Result<Value>

Get lowest integer above a number.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn clz32(
_: &Value,
args: &[Value],
context: &mut Context
) -> Result<Value>

Get the number of leading zeros in the 32 bit representation of a number

More information:

ECMAScript reference

MDN documentation

pub(crate) fn cos(
_: &Value,
args: &[Value],
context: &mut Context
) -> Result<Value>

Get the cosine of a number.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn cosh(
_: &Value,
args: &[Value],
context: &mut Context
) -> Result<Value>

Get the hyperbolic cosine of a number.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn exp(
_: &Value,
args: &[Value],
context: &mut Context
) -> Result<Value>

Get the power to raise the natural logarithm to get the number.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn expm1(
_: &Value,
args: &[Value],
context: &mut Context
) -> Result<Value>

The Math.expm1() function returns e^x - 1, where x is the argument, and e the base of the natural logarithms. The result is computed in a way that is accurate even when the value of x is close 0

More information:

ECMAScript reference

MDN documentation

pub(crate) fn floor(
_: &Value,
args: &[Value],
context: &mut Context
) -> Result<Value>

Get the highest integer below a number.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn fround(
_: &Value,
args: &[Value],
context: &mut Context
) -> Result<Value>

Get the nearest 32-bit single precision float representation of a number.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn hypot(
_: &Value,
args: &[Value],
context: &mut Context
) -> Result<Value>

Get an approximation of the square root of the sum of squares of all arguments.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn imul(
_: &Value,
args: &[Value],
context: &mut Context
) -> Result<Value>

Get the result of the C-like 32-bit multiplication of the two parameters.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn log(
_: &Value,
args: &[Value],
context: &mut Context
) -> Result<Value>

Get the natural logarithm of a number.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn log1p(
_: &Value,
args: &[Value],
context: &mut Context
) -> Result<Value>

Get approximation to the natural logarithm of 1 + x.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn log10(
_: &Value,
args: &[Value],
context: &mut Context
) -> Result<Value>

Get the base 10 logarithm of the number.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn log2(
_: &Value,
args: &[Value],
context: &mut Context
) -> Result<Value>

Get the base 2 logarithm of the number.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn max(
_: &Value,
args: &[Value],
context: &mut Context
) -> Result<Value>

Get the maximum of several numbers.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn min(
_: &Value,
args: &[Value],
context: &mut Context
) -> Result<Value>

Get the minimum of several numbers.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn pow(
_: &Value,
args: &[Value],
context: &mut Context
) -> Result<Value>

Raise a number to a power.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn random(_: &Value, _: &[Value], _: &mut Context) -> Result<Value>

Generate a random floating-point number between 0 and 1.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn round(
_: &Value,
args: &[Value],
context: &mut Context
) -> Result<Value>

Round a number to the nearest integer.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn sign(
_: &Value,
args: &[Value],
context: &mut Context
) -> Result<Value>

Get the sign of a number.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn sin(
_: &Value,
args: &[Value],
context: &mut Context
) -> Result<Value>

Get the sine of a number.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn sinh(
_: &Value,
args: &[Value],
context: &mut Context
) -> Result<Value>

Get the hyperbolic sine of a number.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn sqrt(
_: &Value,
args: &[Value],
context: &mut Context
) -> Result<Value>

Get the square root of a number.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn tan(
_: &Value,
args: &[Value],
context: &mut Context
) -> Result<Value>

Get the tangent of a number.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn tanh(
_: &Value,
args: &[Value],
context: &mut Context
) -> Result<Value>

Get the hyperbolic tangent of a number.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn trunc(
_: &Value,
args: &[Value],
context: &mut Context
) -> Result<Value>

Get the integer part of a number.

More information:

ECMAScript reference

MDN documentation