Module boa::builtins::number

This module implements the global Number object.

The Number JavaScript object is a wrapper object allowing you to work with numerical values. A Number object is created using the Number() constructor. A primitive type object number is created using the Number() function.

The JavaScript Number type is double-precision 64-bit binary format IEEE 754 value. In more recent implementations, JavaScript also supports integers with arbitrary precision using the BigInt type.

More information:

ECMAScript reference

MDN documentation

Constants

BUF_SIZE

Functions

call_number
Number() function.

create
Create a new Number object

equals
The abstract operation Number::equal takes arguments x (a Number) and y (a Number). It performs the following steps when called:

init
Initialise the Number object on the global object.

make_number
Create a new number [[Construct]]

next_after

num_to_exponential
Helper function that formats a float as a ES6-style exponential number string.

num_to_string

same_value
The abstract operation Number::sameValue takes arguments x (a Number) and y (a Number). It performs the following steps when called:

same_value_zero
The abstract operation Number::sameValueZero takes arguments x (a Number) and y (a Number). It performs the following steps when called:

to_exponential
Number.prototype.toExponential( [fractionDigits] )

to_fixed
Number.prototype.toFixed( [digits] )

to_locale_string
Number.prototype.toLocaleString( [locales [, options]] )

to_number
Helper function that converts a Value to a Number.

to_precision
Number.prototype.toPrecision( [precision] )

to_string
Number.prototype.toString( [radix] )

value_of
Number.prototype.toString()