Module boa::builtins::object

This module implements the global Object object.

The Object class represents one of JavaScript's data types.

It is used to store various keyed collections and more complex entities. Objects can be created using the Object() constructor or the object initializer / literal syntax.

More information:

ECMAScript reference

MDN documentation

Modules

gcobject
This module implements the GcObject structure.

internal_methods
This module defines the object internal methods.

iter

Structs

GcObject
Garbage collected Object.

IndexProperties
An iterator over the indexed property entries of an Object

IndexPropertyKeys
An iterator over the index keys (u32) of an Object.

IndexPropertyValues
An iterator over the index values (Property) of an Object.

Iter
An iterator over the property entries of an Object

Keys
An iterator over the keys (PropertyKey) of an Object.

Object
The internal representation of an JavaScript object.

StringProperties
An iterator over the String property entries of an Object

StringPropertyKeys
An iterator over the string keys (RcString) of an Object.

StringPropertyValues
An iterator over the string values (Property) of an Object.

SymbolProperties
An iterator over the Symbol property entries of an Object

SymbolPropertyKeys
An iterator over the keys (RcSymbol) of an Object.

SymbolPropertyValues
An iterator over the Symbol values (Property) of an Object.

Values
An iterator over the values (Property) of an Object.

Enums

ObjectData
Defines the different types of objects.

Statics

PROTOTYPE
Static prototype, usually set on constructors as a key to point to their respective prototype object.

Traits

NativeObject
This trait allows Rust types to be passed around as objects.

Functions

create
Object.create( proto, [propertiesObject] )

define_property
Define a property in an object

get_prototype_of
Get the prototype of an object.

has_own_property
Object.prototype.hasOwnPrototype( property )

init
Initialise the Object object on the global object.

is
Uses the SameValue algorithm to check equality of objects

make_object
Create a new object.

property_is_enumerable

set_prototype_of
Set the prototype of an object.

to_string
Object.prototype.toString()

Type Definitions

Ref
A wrapper type for an immutably borrowed Object.

RefMut
A wrapper type for a mutably borrowed Object.