Struct boa::builtins::object::InternalStateCell

pub struct InternalStateCell {
    state: Rc<dyn Any>,
}

Wrapper around Rc to implement Trace and Finalize.

Fields
state: Rc<dyn Any>

The internal state.

Methods

impl InternalStateCell

pub fn new<T: Any + InternalState>(value: T) -> Self

Create new InternalStateCell from a value.

pub fn downcast_ref<T: Any + InternalState>(&self) -> Option<&T>

Get a reference to the stored value and cast it to T.

pub fn downcast_mut<T: Any + InternalState>(&mut self) -> Option<&mut T>

Get a mutable reference to the stored value and cast it to T.