Struct boa::builtins::object::Object

pub struct Object {
    pub data: ObjectData,
    internal_slots: FxHashMap<String, Value>,
    properties: FxHashMap<RcString, Property>,
    symbol_properties: FxHashMap<u32, Property>,
    state: Option<InternalStateCell>,
    extensible: bool,
}

The internal representation of an JavaScript object.

Fields
data: ObjectData

The type of the object.

internal_slots: FxHashMap<String, Value>

Internal Slots

properties: FxHashMap<RcString, Property>

Properties

symbol_properties: FxHashMap<u32, Property>

Symbol Properties

state: Option<InternalStateCell>

Some rust object that stores internal state

extensible: bool

Whether it can have new properties added to it.

Implementations

impl Object

pub fn has_property(&self, val: &Value) -> bool

Check if object has property.

More information:

ECMAScript reference

pub fn is_extensible(&self) -> bool

Check if it is extensible.

More information:

ECMAScript reference

pub fn prevent_extensions(&mut self) -> bool

Disable extensibility.

More information:

ECMAScript reference

pub fn delete(&mut self, prop_key: &Value) -> bool

Delete property.

pub fn get(&self, val: &Value) -> Value

pub fn set(&mut self, field: Value, val: Value) -> bool

[[Set]] https://tc39.es/ecma262/#sec-ordinary-object-internal-methods-and-internal-slots-set-p-v-receiver

pub fn define_own_property(
&mut self,
property_key: String,
desc: Property
) -> bool

Define an own property.

More information:

ECMAScript reference

pub fn get_own_property(&self, prop: &Value) -> Property

The specification returns a Property Descriptor or Undefined.

These are 2 separate types and we can't do that here.

More information:

ECMAScript reference

pub fn set_prototype_of(&mut self, val: Value) -> bool

Object.setPropertyOf(obj, prototype)

This method sets the prototype (i.e., the internal [[Prototype]] property) of a specified object to another object or null.

More information:

ECMAScript reference

MDN documentation

pub fn get_prototype_of(&self) -> Value

Returns either the prototype or null

More information:

ECMAScript reference

[MDN documentation][mdn]

pub fn get_internal_slot(&self, name: &str) -> Value

Helper function to get an immutable internal slot or Null.

pub fn set_internal_slot(&mut self, name: &str, val: Value)

Helper function to set an internal slot.

pub(crate) fn insert_property<N>(&mut self, name: N, p: Property) where
N: Into<RcString>,

Helper function for property insertion.

pub(crate) fn remove_property(&mut self, name: &str)

Helper function for property removal.

pub(crate) fn insert_field<N>(
&mut self,
name: N,
value: Value
) -> Option<Property> where
N: Into<RcString>,

Inserts a field in the object properties without checking if it's writable.

If a field was already in the object with the same name that a Some is returned with that field, otherwise None is retuned.

pub fn get_field(&self, name: &str) -> Option<&Value>

This function returns an Optional reference value to the objects field.

if it exist Some is returned with a reference to that fields value. Otherwise None is retuned.

impl Object

pub fn new() -> Self

pub fn function(function: Function) -> Self

Return a new ObjectData struct, with kind set to Ordinary

pub fn create(proto: Value) -> Self

ObjectCreate is used to specify the runtime creation of new ordinary objects.

More information:

ECMAScript reference

pub fn boolean(value: bool) -> Self

Return a new Boolean object whose [[BooleanData]] internal slot is set to argument.

pub fn number(value: f64) -> Self

Return a new Number object whose [[NumberData]] internal slot is set to argument.

pub fn string<S>(value: S) -> Self where
S: Into<RcString>,

Return a new String object whose [[StringData]] internal slot is set to argument.

pub fn bigint(value: RcBigInt) -> Self

Return a new BigInt object whose [[BigIntData]] internal slot is set to argument.

pub fn from(value: &Value) -> Result<Self, ()>

Converts the Value to an Object type.

More information:

ECMAScript reference

pub fn is_callable(&self) -> bool

It determines if Object is a callable function with a [[Call]] internal method.

More information:

EcmaScript reference

pub fn is_constructable(&self) -> bool

It determines if Object is a function object with a [[Construct]] internal method.

More information:

EcmaScript reference

pub fn is_array(&self) -> bool

Checks if it an Array object.

pub fn as_array(&self) -> Option<()>

pub fn is_string(&self) -> bool

Checks if it a String object.

pub fn as_string(&self) -> Option<RcString>

pub fn is_function(&self) -> bool

Checks if it a Function object.

pub fn as_function(&self) -> Option<&Function>

pub fn is_symbol(&self) -> bool

Checks if it a Symbol object.

pub fn as_symbol(&self) -> Option<RcSymbol>

pub fn is_error(&self) -> bool

Checks if it an Error object.

pub fn as_error(&self) -> Option<()>

pub fn is_boolean(&self) -> bool

Checks if it a Boolean object.

pub fn as_boolean(&self) -> Option<bool>

pub fn is_number(&self) -> bool

Checks if it a Number object.

pub fn as_number(&self) -> Option<f64>

pub fn is_bigint(&self) -> bool

Checks if it a BigInt object.

pub fn as_bigint(&self) -> Option<&BigInt>

pub fn is_ordinary(&self) -> bool

Checks if it an ordinary object.

pub fn internal_slots(&self) -> &FxHashMap<String, Value>

pub fn internal_slots_mut(&mut self) -> &mut FxHashMap<String, Value>

pub fn properties(&self) -> &FxHashMap<RcString, Property>

pub fn properties_mut(&mut self) -> &mut FxHashMap<RcString, Property>

pub fn symbol_properties(&self) -> &FxHashMap<u32, Property>

pub fn symbol_properties_mut(&mut self) -> &mut FxHashMap<u32, Property>

pub fn state(&self) -> &Option<InternalStateCell>

pub fn state_mut(&mut self) -> &mut Option<InternalStateCell>