Struct boa::builtins::string::String

pub(crate) struct String;

JavaScript String implementation.

Implementations

impl String

pub(crate) const NAME: &'static str

The name of the object.

pub(crate) const LENGTH: usize

The amount of arguments this function object takes.

pub(crate) const MAX_STRING_LENGTH: f64

JavaScript strings must be between 0 and less than positive Infinity and cannot be a negative number. The range of allowed values can be described like this: [0, +∞).

The resulting string can also not be larger than the maximum string size, which can differ in JavaScript engines. In Boa it is 2^32 - 1

fn this_string_value(this: &Value, ctx: &mut Context) -> Result<RcString>

pub(crate) fn make_string(
this: &Value,
args: &[Value],
ctx: &mut Context
) -> Result<Value>

[[Construct]] - Creates a new instance this

[[Call]] - Returns a new native string https://tc39.es/ecma262/#sec-string-constructor-string-value

pub(crate) fn to_string(
this: &Value,
_: &[Value],
ctx: &mut Context
) -> Result<Value>

Get the string value to a primitive string

pub(crate) fn char_at(
this: &Value,
args: &[Value],
ctx: &mut Context
) -> Result<Value>

String.prototype.charAt( index )

The String object's charAt() method returns a new string consisting of the single UTF-16 code unit located at the specified offset into the string.

Characters in a string are indexed from left to right. The index of the first character is 0, and the index of the last character—in a string called stringName—is stringName.length - 1. If the index you supply is out of this range, JavaScript returns an empty string.

If no index is provided to charAt(), the default is 0.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn char_code_at(
this: &Value,
args: &[Value],
ctx: &mut Context
) -> Result<Value>

String.prototype.charCodeAt( index )

The charCodeAt() method returns an integer between 0 and 65535 representing the UTF-16 code unit at the given index.

Unicode code points range from 0 to 1114111 (0x10FFFF). The first 128 Unicode code points are a direct match of the ASCII character encoding.

charCodeAt() returns NaN if the given index is less than 0, or if it is equal to or greater than the length of the string.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn concat(
this: &Value,
args: &[Value],
ctx: &mut Context
) -> Result<Value>

String.prototype.concat( str1[, ...strN] )

The concat() method concatenates the string arguments to the calling string and returns a new string.

Changes to the original string or the returned string don't affect the other.

If the arguments are not of the type string, they are converted to string values before concatenating.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn repeat(
this: &Value,
args: &[Value],
ctx: &mut Context
) -> Result<Value>

String.prototype.repeat( count )

The repeat() method constructs and returns a new string which contains the specified number of copies of the string on which it was called, concatenated together.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn slice(
this: &Value,
args: &[Value],
ctx: &mut Context
) -> Result<Value>

String.prototype.slice( beginIndex [, endIndex] )

The slice() method extracts a section of a string and returns it as a new string, without modifying the original string.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn starts_with(
this: &Value,
args: &[Value],
ctx: &mut Context
) -> Result<Value>

String.prototype.startWith( searchString[, position] )

The startsWith() method determines whether a string begins with the characters of a specified string, returning true or false as appropriate.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn ends_with(
this: &Value,
args: &[Value],
ctx: &mut Context
) -> Result<Value>

String.prototype.endsWith( searchString[, length] )

The endsWith() method determines whether a string ends with the characters of a specified string, returning true or false as appropriate.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn includes(
this: &Value,
args: &[Value],
ctx: &mut Context
) -> Result<Value>

String.prototype.includes( searchString[, position] )

The includes() method determines whether one string may be found within another string, returning true or false as appropriate.

More information:

ECMAScript reference

MDN documentation

fn get_regex_string(value: &Value) -> StdString

Return either the string itself or the string of the regex equivalent

pub(crate) fn replace(
this: &Value,
args: &[Value],
ctx: &mut Context
) -> Result<Value>

String.prototype.replace( regexp|substr, newSubstr|function )

The replace() method returns a new string with some or all matches of a pattern replaced by a replacement.

The pattern can be a string or a RegExp, and the replacement can be a string or a function to be called for each match. If pattern is a string, only the first occurrence will be replaced.

The original string is left unchanged.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn index_of(
this: &Value,
args: &[Value],
ctx: &mut Context
) -> Result<Value>

String.prototype.indexOf( searchValue[, fromIndex] )

The indexOf() method returns the index within the calling String object of the first occurrence of the specified value, starting the search at fromIndex.

Returns -1 if the value is not found.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn last_index_of(
this: &Value,
args: &[Value],
ctx: &mut Context
) -> Result<Value>

String.prototype.lastIndexOf( searchValue[, fromIndex] )

The lastIndexOf() method returns the index within the calling String object of the last occurrence of the specified value, searching backwards from fromIndex.

Returns -1 if the value is not found.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn match(
this: &Value,
args: &[Value],
ctx: &mut Context
) -> Result<Value>

String.prototype.match( regexp )

The match() method retrieves the result of matching a string against a regular expression.

More information:

ECMAScript reference

MDN documentation

fn string_pad(
primitive: RcString,
max_length: i32,
fill_string: Option<RcString>,
at_start: bool
) -> Result<Value>

Abstract method StringPad.

Performs the actual string padding for padStart/End. https://tc39.es/ecma262/#sec-stringpad/

pub(crate) fn pad_end(
this: &Value,
args: &[Value],
ctx: &mut Context
) -> Result<Value>

String.prototype.padEnd( targetLength[, padString] )

The padEnd() method pads the current string with a given string (repeated, if needed) so that the resulting string reaches a given length.

The padding is applied from the end of the current string.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn pad_start(
this: &Value,
args: &[Value],
ctx: &mut Context
) -> Result<Value>

String.prototype.padStart( targetLength [, padString] )

The padStart() method pads the current string with another string (multiple times, if needed) until the resulting string reaches the given length.

The padding is applied from the start of the current string.

More information:

ECMAScript reference

MDN documentation

fn is_trimmable_whitespace(c: char) -> bool

Helper function to check if a char is trimmable.

pub(crate) fn trim(
this: &Value,
_: &[Value],
ctx: &mut Context
) -> Result<Value>

String.prototype.trim()

The trim() method removes whitespace from both ends of a string.

Whitespace in this context is all the whitespace characters (space, tab, no-break space, etc.) and all the line terminator characters (LF, CR, etc.).

More information:

ECMAScript reference

MDN documentation

pub(crate) fn trim_start(
this: &Value,
_: &[Value],
ctx: &mut Context
) -> Result<Value>

String.prototype.trimStart()

The trimStart() method removes whitespace from the beginning of a string.

Whitespace in this context is all the whitespace characters (space, tab, no-break space, etc.) and all the line terminator characters (LF, CR, etc.).

More information:

ECMAScript reference

MDN documentation

pub(crate) fn trim_end(
this: &Value,
_: &[Value],
ctx: &mut Context
) -> Result<Value>

String.prototype.trimEnd()

The trimEnd() method removes whitespace from the end of a string.

Whitespace in this context is all the whitespace characters (space, tab, no-break space, etc.) and all the line terminator characters (LF, CR, etc.).

More information:

ECMAScript reference

MDN documentation

pub(crate) fn to_lowercase(
this: &Value,
_: &[Value],
ctx: &mut Context
) -> Result<Value>

String.prototype.toLowerCase()

The toLowerCase() method returns the calling string value converted to lower case.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn to_uppercase(
this: &Value,
_: &[Value],
ctx: &mut Context
) -> Result<Value>

String.prototype.toUpperCase()

The toUpperCase() method returns the calling string value converted to uppercase.

The value will be converted to a string if it isn't one

More information:

ECMAScript reference

MDN documentation

pub(crate) fn substring(
this: &Value,
args: &[Value],
ctx: &mut Context
) -> Result<Value>

String.prototype.substring( indexStart[, indexEnd] )

The substring() method returns the part of the string between the start and end indexes, or to the end of the string.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn substr(
this: &Value,
args: &[Value],
ctx: &mut Context
) -> Result<Value>

String.prototype.substr( start[, length] )

The substr() method returns a portion of the string, starting at the specified index and extending for a given number of characters afterward.

More information:

ECMAScript reference

MDN documentation

https://tc39.es/ecma262/#sec-string.prototype.substr

pub(crate) fn value_of(
this: &Value,
args: &[Value],
ctx: &mut Context
) -> Result<Value>

String.prototype.valueOf()

The valueOf() method returns the primitive value of a String object.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn match_all(
this: &Value,
args: &[Value],
ctx: &mut Context
) -> Result<Value>

String.prototype.matchAll( regexp )

The matchAll() method returns an iterator of all results matching a string against a regular expression, including capturing groups.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn init(interpreter: &mut Context) -> (&'static str, Value)

Initialise the String object on the global object.