Module boa::builtins::value

This module implements the JavaScript Value.

Javascript values, utility methods and conversion between Javascript values and Rust values.

Enums

ValueData
A Javascript value

Traits

FromValue
Conversion to Rust values from Javascript values

ToValue
Conversion to Javascript values from Rust values

Functions

display_obj
A helper function for specifically printing object values

from_value
A utility function that just calls FromValue::from_value

log_string_from

same_value
The internal comparison abstract operation SameValue(x, y), where x and y are ECMAScript language values, produces true or false. Such a comparison is performed as follows:

same_value_non_number

to_value
A utility function that just calls ToValue::to_value

undefined

Type Definitions

ResultValue
The result of a Javascript expression is represented like this so it can succeed (Ok) or fail (Err)

Value
A Garbage-collected Javascript value as represented in the interpreter.