Enum boa::builtins::value::ValueData

pub enum ValueData {
    Null,
    Undefined,
    Boolean(bool),
    String(String),
    Number(f64),
    Integer(i32),
    Object(GcCell<Object>),
    Function(Box<GcCell<Function>>),
    Symbol(GcCell<Object>),
}

A Javascript value

Variants

Null

null - A null value, for when a value doesn't exist

Undefined

undefined - An undefined value, for when a field or index doesn't exist

Boolean(bool)

boolean - A true / false value, for if a certain criteria is met

String(String)

String - A UTF-8 string, such as "Hello, world"

Number(f64)

Number - A 64-bit floating point number, such as 3.1415

Integer(i32)

Number - A 32-bit integer, such as 42

Object(GcCell<Object>)

Object - An object, such as Math, represented by a binary tree of string keys to Javascript values

Function(Box<GcCell<Function>>)

Function - A runnable block of code, such as Math.sqrt, which can take some variables and return a useful value or act upon an object

Symbol(GcCell<Object>)

Symbol - A Symbol Type - Internally Symbols are similar to objects, except there are no properties, only internal slots

Methods

impl ValueData

pub fn new_obj(global: Option<&Value>) -> Value

Returns a new empty object

pub fn new_obj_from_prototype(proto: Value, kind: ObjectKind) -> Value

Similar to new_obj, but you can pass a prototype to create from, plus a kind

pub fn is_extensible(&self) -> bool

This will tell us if we can exten an object or not, not properly implemented yet

For now always returns true.

For scalar types it should be false, for objects check the private field for extensibilaty. By default true.

<https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Object/seal would turn extensible to false/> <https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Object/freeze would also turn extensible to false/>

pub fn is_object(&self) -> bool

Returns true if the value is an object

pub fn is_symbol(&self) -> bool

Returns true if the value is a symbol

pub fn is_function(&self) -> bool

Returns true if the value is a function

pub fn is_undefined(&self) -> bool

Returns true if the value is undefined

pub fn is_null(&self) -> bool

Returns true if the value is null

pub fn is_null_or_undefined(&self) -> bool

Returns true if the value is null or undefined

pub fn is_double(&self) -> bool

Returns true if the value is a 64-bit floating-point number

pub fn is_num(&self) -> bool

Returns true if the value is a number

pub fn is_string(&self) -> bool

Returns true if the value is a string

pub fn is_boolean(&self) -> bool

Returns true if the value is a boolean

pub fn is_true(&self) -> bool

Returns true if the value is true

toBoolean

pub fn to_num(&self) -> f64

Converts the value into a 64-bit floating point number

pub fn to_int(&self) -> i32

Converts the value into a 32-bit integer

pub fn remove_prop(&self, field: &str)

remove_prop removes a property from a Value object.

It will return a boolean based on if the value was removed, if there was no value to remove false is returned

pub fn get_prop(&self, field: &str) -> Option<Property>

Resolve the property in the object.

A copy of the Property is returned.

pub fn update_prop(
&self,
field: &str,
value: Option<Value>,
enumerable: Option<bool>,
writable: Option<bool>,
configurable: Option<bool>
)

update_prop will overwrite individual [Property] fields, unlike Set_prop, which will overwrite prop with a new Property Mostly used internally for now

pub fn get_internal_slot(&self, field: &str) -> Value

Resolve the property in the object.

Returns a copy of the Property.

pub fn get_field(&self, field: Value) -> Value

Resolve the property in the object and get its value, or undefined if this is not an object or the field doesn't exist get_field recieves a Property from get_prop(). It should then return the [[Get]] result value if that's set, otherwise fall back to [[Value]] TODO: this function should use the get Value if its set

pub fn has_internal_state(&self) -> bool

Check whether an object has an internal state set.

pub fn get_internal_state(&self) -> Option<InternalStateCell>

Get the internal state of an object.

pub fn with_internal_state_ref<S: Any + InternalState, R, F: FnOnce(&S) -> R>(
&self,
f: F
) -> R

Run a function with a reference to the internal state.

Panics

This will panic if this value doesn't have an internal state or if the internal state doesn't have the concrete type S.

pub fn with_internal_state_mut<S: Any + InternalState, R, F: FnOnce(&mut S) -> R>(
&self,
f: F
) -> R

Run a function with a mutable reference to the internal state.

Panics

This will panic if this value doesn't have an internal state or if the internal state doesn't have the concrete type S.

pub fn has_field(&self, field: &str) -> bool

Check to see if the Value has the field, mainly used by environment records

pub fn get_field_slice(&self, field: &str) -> Value

Resolve the property in the object and get its value, or undefined if this is not an object or the field doesn't exist

pub fn set_field(&self, field: Value, val: Value) -> Value

Set the field in the value Field could be a Symbol, so we need to accept a Value (not a string)

pub fn set_field_slice<'a>(&self, field: &'a str, val: Value) -> Value

Set the field in the value

pub fn set_internal_slot(&self, field: &str, val: Value) -> Value

Set the private field in the value

pub fn set_kind(&self, kind: ObjectKind) -> ObjectKind

Set the kind of an object

pub fn set_prop(&self, field: String, prop: Property) -> Property

Set the property in the value

pub fn set_prop_slice<'t>(&self, field: &'t str, prop: Property) -> Property

Set the property in the value

pub fn set_internal_state<T: Any + InternalState>(&self, state: T)

Set internal state of an Object. Discards the previous state if it was set.

pub fn from_json(json: JSONValue) -> Self

Convert from a JSON value to a JS value

pub fn to_json(&self) -> JSONValue

Conversts the Value to JSON.

pub fn get_type(&self) -> &'static str

Get the type of the value

https://tc39.es/ecma262/#sec-typeof-operator

pub fn as_num_to_power(&self, other: ValueData) -> ValueData