Struct boa::builtins::value::Value

pub struct Value(pub(crate) Gc<ValueData>);

A Garbage-collected Javascript value as represented in the interpreter.

Methods

impl Value

pub fn strict_equals(&self, other: &Self) -> bool

Strict equality comparison.

This method is executed when doing strict equality comparisons with the === operator. For more information, check https://tc39.es/ecma262/#sec-strict-equality-comparison.

pub fn equals(
&mut self,
other: &mut Self,
interpreter: &mut Interpreter
) -> bool

Abstract equality comparison.

This method is executed when doing abstract equality comparisons with the == operator. For more information, check https://tc39.es/ecma262/#sec-abstract-equality-comparison

impl Value

pub fn undefined() -> Self

Creates a new undefined value.

pub fn null() -> Self

Creates a new null value.

pub fn string<S>(value: S) -> Self where
S: Into<String>,

Creates a new string value.

pub fn rational<N>(value: N) -> Self where
N: Into<f64>,

Creates a new number value.

pub fn integer<I>(value: I) -> Self where
I: Into<i32>,

Creates a new number value.

pub fn number<N>(value: N) -> Self where
N: Into<f64>,

Creates a new number value.

pub fn bigint(value: BigInt) -> Self

Creates a new bigint value.

pub fn boolean(value: bool) -> Self

Creates a new boolean value.

pub fn object(object: Object) -> Self

Creates a new object value.

pub fn data(&self) -> &ValueData

Gets the underlying ValueData structure.

pub fn as_num_to_power(&self, other: Self) -> Self

Helper function to convert the Value to a number and compute its power.

pub fn new_object(global: Option<&Value>) -> Self

Returns a new empty object

pub fn new_object_from_prototype(proto: Value, kind: ObjectKind) -> Self

Similar to new_object, but you can pass a prototype to create from, plus a kind

pub fn from_json(json: JSONValue, interpreter: &mut Interpreter) -> Self

Convert from a JSON value to a JS value

pub fn to_json(&self, interpreter: &mut Interpreter) -> Result<JSONValue, Value>

Conversts the Value to JSON.