Struct boa::context::Context

pub struct Context {
    pub(crate) realm: Realm,
    executor: Interpreter,
    console: Console,
    iterator_prototypes: IteratorPrototypes,
    typed_array_constructor: StandardConstructor,
    standard_objects: StandardObjects,
    intrinsic_objects: IntrinsicObjects,
    strict: StrictType,
}

Javascript context. It is the primary way to interact with the runtime.

Contexts constructed in a thread share the same runtime, therefore it is possible to share objects from one context to another context, but they have to be in the same thread.

Examples

Execute Function of Script File

use boa::{Context, object::ObjectInitializer, property::{Attribute, PropertyDescriptor}};

let script = r#"
function test(arg1) {
    if(arg1 != null) {
        return arg1.x;
    }
    return 112233;
}
"#;

let mut context = Context::new();

// Populate the script definition to the context.
context.eval(script).unwrap();

// Create an object that can be used in eval calls.
let arg = ObjectInitializer::new(&mut context)
    .property("x", 12, Attribute::READONLY)
    .build();
context.register_global_property(
    "arg",
    arg,
    Attribute::all()
);

let value = context.eval("test(arg)").unwrap();

assert_eq!(value.as_number(), Some(12.0))

Fields
realm: Realm

realm holds both the global object and the environment

executor: Interpreter

The current executor.

console: Console

console object state.

iterator_prototypes: IteratorPrototypes

Cached iterator prototypes.

typed_array_constructor: StandardConstructor

Cached TypedArray constructor.

standard_objects: StandardObjects

Cached standard objects and their prototypes.

intrinsic_objects: IntrinsicObjects

Cached intrinsic objects

strict: StrictType

Whether or not strict mode is active.

Implementations

impl Context

pub fn new() -> Self

Create a new Context.

pub fn executor(&mut self) -> &mut Interpreter

pub(crate) fn console(&self) -> &Console

A helper function for getting an immutable reference to the console object.

pub(crate) fn console_mut(&mut self) -> &mut Console

A helper function for getting a mutable reference to the console object.

pub fn strict(&self) -> bool

Returns if strict mode is currently active.

pub(crate) fn strict_type(&self) -> StrictType

Returns the strict mode type.

pub(crate) fn set_strict(&mut self, strict: StrictType)

Set strict type.

pub fn set_strict_mode_off(&mut self)

Disable the strict mode.

pub fn set_strict_mode_global(&mut self)

Enable the global strict mode.

fn create_intrinsics(&mut self)

Sets up the default global objects within Global

pub fn construct_object(&self) -> JsObject

Constructs an object with the %Object.prototype% prototype.

pub(crate) fn call(
&mut self,
f: &JsValue,
this: &JsValue,
args: &[JsValue]
) -> JsResult<JsValue>

https://tc39.es/ecma262/#sec-call

pub fn global_object(&self) -> JsObject

Return the global object.

pub fn construct_error<M>(&mut self, message: M) -> JsValue where
M: Into<Box<str>>,

Constructs a Error with the specified message.

pub fn throw_error<M>(&mut self, message: M) -> JsResult<JsValue> where
M: Into<Box<str>>,

Throws a Error with the specified message.

pub fn construct_range_error<M>(&mut self, message: M) -> JsValue where
M: Into<Box<str>>,

Constructs a RangeError with the specified message.

pub fn throw_range_error<M>(&mut self, message: M) -> JsResult<JsValue> where
M: Into<Box<str>>,

Throws a RangeError with the specified message.

pub fn construct_type_error<M>(&mut self, message: M) -> JsValue where
M: Into<Box<str>>,

Constructs a TypeError with the specified message.

pub fn throw_type_error<M>(&mut self, message: M) -> JsResult<JsValue> where
M: Into<Box<str>>,

Throws a TypeError with the specified message.

pub fn construct_reference_error<M>(&mut self, message: M) -> JsValue where
M: Into<Box<str>>,

Constructs a ReferenceError with the specified message.

pub fn throw_reference_error<M>(&mut self, message: M) -> JsResult<JsValue> where
M: Into<Box<str>>,

Throws a ReferenceError with the specified message.

pub fn construct_syntax_error<M>(&mut self, message: M) -> JsValue where
M: Into<Box<str>>,

Constructs a SyntaxError with the specified message.

pub fn throw_syntax_error<M>(&mut self, message: M) -> JsResult<JsValue> where
M: Into<Box<str>>,

Throws a SyntaxError with the specified message.

pub fn construct_eval_error<M>(&mut self, message: M) -> JsValue where
M: Into<Box<str>>,

Constructs a EvalError with the specified message.

pub fn construct_uri_error<M>(&mut self, message: M) -> JsValue where
M: Into<Box<str>>,

Constructs a URIError with the specified message.

pub fn throw_eval_error<M>(&mut self, message: M) -> JsResult<JsValue> where
M: Into<Box<str>>,

Throws a EvalError with the specified message.

pub fn throw_uri_error<M>(&mut self, message: M) -> JsResult<JsValue> where
M: Into<Box<str>>,

Throws a URIError with the specified message.

pub(crate) fn create_function<N, P>(
&mut self,
name: N,
params: P,
body: StatementList,
constructor: bool,
this_mode: ThisMode
) -> JsResult<JsValue> where
N: Into<JsString>,
P: Into<Box<[FormalParameter]>>,

Utility to create a function Value for Function Declarations, Arrow Functions or Function Expressions

pub fn register_global_function(
&mut self,
name: &str,
length: usize,
body: NativeFunctionSignature
) -> JsResult<()>

Register a global native function.

This is more efficient that creating a closure function, since this does not allocate, it is just a function pointer.

The function will be both constructable (call with new).

The function will be bound to the global object with writable, non-enumerable and configurable attributes. The same as when you create a function in JavaScript.

Note

If you want to make a function only constructable, or wish to bind it differently to the global object, you can create the function object with FunctionBuilder. And bind it to the global object with Context::register_global_property method.

pub fn register_global_closure<F>(
&mut self,
name: &str,
length: usize,
body: F
) -> JsResult<()> where
F: Fn(&JsValue, &[JsValue], &mut Context) -> JsResult<JsValue> + Copy + 'static,

Register a global closure function.

The function will be both constructable (call with new).

The function will be bound to the global object with writable, non-enumerable and configurable attributes. The same as when you create a function in JavaScript.

Note #1

If you want to make a function only constructable, or wish to bind it differently to the global object, you can create the function object with FunctionBuilder. And bind it to the global object with Context::register_global_property method.

Note #2

This function will only accept Copy closures, meaning you cannot move Clone types, just Copy types. If you need to move Clone types as captures, see FunctionBuilder::closure_with_captures.

See https://github.com/boa-dev/boa/issues/1515 for an explanation on why we need to restrict the set of accepted closures.

pub(crate) fn has_property(
&mut self,
obj: &JsValue,
key: &PropertyKey
) -> JsResult<bool>

https://tc39.es/ecma262/#sec-hasproperty

pub(crate) fn set_value(
&mut self,
node: &Node,
value: JsValue
) -> JsResult<JsValue>

pub fn register_global_class<T>(&mut self) -> JsResult<()> where
T: Class,

Register a global class of type T, where T implements Class.

Example

#[derive(Debug, Trace, Finalize)]
struct MyClass;

impl Class for MyClass {
   // ...
}

context.register_global_class::<MyClass>();

pub fn register_global_property<K, V>(
&mut self,
key: K,
value: V,
attribute: Attribute
) where
K: Into<PropertyKey>,
V: Into<JsValue>,

Register a global property.

Example

use boa::{Context, property::{Attribute, PropertyDescriptor}, object::ObjectInitializer};

let mut context = Context::new();

context.register_global_property(
    "myPrimitiveProperty",
    10,
    Attribute::all()
);

let object = ObjectInitializer::new(&mut context)
   .property(
        "x",
        0,
        Attribute::all()
    )
    .property(
        "y",
        1,
        Attribute::all()
    )
   .build();
context.register_global_property(
    "myObjectProperty",
    object,
    Attribute::all()
);

pub fn eval<T: AsRef<[u8]>>(&mut self, src: T) -> JsResult<JsValue>

Evaluates the given code.

Examples

let mut context = Context::new();

let value = context.eval("1 + 3").unwrap();

assert!(value.is_number());
assert_eq!(value.as_number().unwrap(), 4.0);

pub fn iterator_prototypes(&self) -> &IteratorPrototypes

Return the cached iterator prototypes.

pub(crate) fn typed_array_constructor(&self) -> &StandardConstructor

Return the cached TypedArray constructor.

pub fn standard_objects(&self) -> &StandardObjects

Return the core standard objects.

pub fn intrinsics(&self) -> &IntrinsicObjects

Return the intrinsic objects.

impl Context

pub(crate) fn push_environment<T: Into<Environment>>(&mut self, env: T)

pub(crate) fn pop_environment(&mut self) -> Option<Environment>

pub(crate) fn get_this_binding(&mut self) -> JsResult<JsValue>

pub(crate) fn create_mutable_binding(
&mut self,
name: &str,
deletion: bool,
scope: VariableScope
) -> JsResult<()>

pub(crate) fn create_immutable_binding(
&mut self,
name: &str,
deletion: bool,
scope: VariableScope
) -> JsResult<()>

pub(crate) fn set_mutable_binding(
&mut self,
name: &str,
value: JsValue,
strict: bool
) -> JsResult<()>

pub(crate) fn initialize_binding(
&mut self,
name: &str,
value: JsValue
) -> JsResult<()>

pub(crate) fn get_current_environment(&mut self) -> Environment

When neededing to clone an environment (linking it with another environnment) cloning is more suited. The GC will remove the env once nothing is linking to it anymore

pub(crate) fn has_binding(&mut self, name: &str) -> JsResult<bool>

pub(crate) fn get_binding_value(&mut self, name: &str) -> JsResult<JsValue>