Module boa::environment::environment_record_trait

Environment Records

https://tc39.es/ecma262/#sec-environment-records https://tc39.es/ecma262/#sec-lexical-environments

Some environments are stored as JSObjects. This is for GC, i.e we want to keep an environment if a variable is closed-over (a closure is returned). All of the logic to handle scope/environment records are stored in here.

There are 5 Environment record kinds. They all have methods in common, these are implemented as a the EnvironmentRecordTrait

Traits

EnvironmentRecordTrait
https://tc39.es/ecma262/#sec-environment-records