Module boa::environment::function_environment_record

Function Environment Records

A function Environment Record is a declarative Environment Record that is used to represent the top-level scope of a function and, if the function is not an ArrowFunction, provides a this binding. If a function is not an ArrowFunction function and references super, its function Environment Record also contains the state that is used to perform super method invocations from within the function. More info: https://tc39.es/ecma262/#sec-function-environment-records

Structs

FunctionEnvironmentRecord
https://tc39.es/ecma262/#table-16

Enums

BindingStatus
Different binding status for this. Usually set on a function environment record