Struct boa::environment::function_environment_record::FunctionEnvironmentRecord

pub struct FunctionEnvironmentRecord {
    pub declarative_record: DeclarativeEnvironmentRecord,
    pub this_value: JsValue,
    pub this_binding_status: BindingStatus,
    pub function: JsObject,
    pub home_object: JsValue,
    pub new_target: JsValue,
}

https://tc39.es/ecma262/#table-16

Fields
declarative_record: DeclarativeEnvironmentRecordthis_value: JsValue

This is the this value used for this invocation of the function.

this_binding_status: BindingStatus

If the value is “lexical”, this is an ArrowFunction and does not have a local this value.

function: JsObject

The function object whose invocation caused this Environment Record to be created.

home_object: JsValue

If the associated function has super property accesses and is not an ArrowFunction, [[HomeObject]] is the object that the function is bound to as a method. The default value for [[HomeObject]] is undefined.

new_target: JsValue

If this Environment Record was created by the [[Construct]] internal method, [[NewTarget]] is the value of the [[Construct]] newTarget parameter. Otherwise, its value is undefined.

Implementations

impl FunctionEnvironmentRecord

pub fn new(
f: JsObject,
this: Option<JsValue>,
outer: Option<Environment>,
binding_status: BindingStatus,
new_target: JsValue,
context: &mut Context
) -> JsResult<FunctionEnvironmentRecord>

pub fn bind_this_value(
&mut self,
value: JsValue,
context: &mut Context
) -> JsResult<JsValue>

9.1.1.3.1 BindThisValue ( V )

More information:

ECMAScript reference

pub fn get_super_base(&self) -> JsValue

9.1.1.3.5 GetSuperBase ( )

More information:

ECMAScript reference