Module boa::environment::lexical_environment

Lexical Environment

https://tc39.es/ecma262/#sec-lexical-environment-operations

The following operations are used to operate upon lexical environments This is the entrypoint to lexical environments.

Structs

EnvironmentError
An error that occurred during lexing or compiling of the source input.

LexicalEnvironment

Enums

EnvironmentType
Give each environment an easy way to declare its own type This helps with comparisons

VariableScope
The scope of a given variable

Type Definitions

Environment
Environments are wrapped in a Box and then in a GC wrapper