Struct boa::exec::Interpreter

pub struct Interpreter {
    state: InterpreterState,
    pub realm: Realm,
    symbol_count: u32,
}

A Javascript intepreter

Fields
state: InterpreterState

the current state of the interpreter.

realm: Realm

realm holds both the global object and the environment

symbol_count: u32

This is for generating an unique internal Symbol hash.

Methods

impl Interpreter

pub fn construct_range_error<M>(&mut self, message: M) -> Value where
M: Into<String>,

Constructs a RangeError with the specified message.

pub fn throw_range_error<M>(&mut self, message: M) -> ResultValue where
M: Into<String>,

Throws a RangeError with the specified message.

pub fn construct_type_error<M>(&mut self, message: M) -> Value where
M: Into<String>,

Constructs a TypeError with the specified message.

pub fn throw_type_error<M>(&mut self, message: M) -> ResultValue where
M: Into<String>,

Throws a TypeError with the specified message.

pub fn construct_reference_error<M>(&mut self, message: M) -> Value where
M: Into<String>,

Constructs a ReferenceError with the specified message.

pub fn throw_reference_error<M>(&mut self, message: M) -> ResultValue where
M: Into<String>,

Throws a ReferenceError with the specified message.

impl Interpreter

pub fn new(realm: Realm) -> Self

Creates a new interpreter.

pub(crate) fn realm(&self) -> &Realm

Retrieves the Realm of this executor.

pub(crate) fn realm_mut(&mut self) -> &mut Realm

Retrieves the Realm of this executor as a mutable reference.

pub(crate) fn generate_hash(&mut self) -> u32

Generates a new Symbol internal hash.

This currently is an incremented value.

pub(crate) fn create_function<P, B>(
&mut self,
params: P,
body: B,
this_mode: ThisMode,
constructable: bool,
callable: bool
) -> Value where
P: Into<Box<[FormalParameter]>>,
B: Into<StatementList>,

Utility to create a function Value for Function Declarations, Arrow Functions or Function Expressions

pub(crate) fn call(
&mut self,
f: &Value,
this: &Value,
arguments_list: &[Value]
) -> ResultValue

https://tc39.es/ecma262/#sec-call

pub fn to_string(&mut self, value: &Value) -> Result<RcString, Value>

Converts a value into a rust heap allocated string.

pub fn to_bigint(&mut self, value: &Value) -> Result<RcBigInt, Value>

Helper function.

pub fn to_index(&mut self, value: &Value) -> Result<usize, Value>

Converts a value to a non-negative integer if it is a valid integer index value.

See: https://tc39.es/ecma262/#sec-toindex

pub fn to_integer(&mut self, value: &Value) -> Result<f64, Value>

Converts a value to an integral Number value.

See: https://tc39.es/ecma262/#sec-tointeger

pub fn to_int32(&mut self, value: &Value) -> Result<i32, Value>

Converts a value to an integral 32 bit signed integer.

See: https://tc39.es/ecma262/#sec-toint32

pub fn to_uint32(&mut self, value: &Value) -> Result<u32, Value>

Converts a value to an integral 32 bit unsigned integer.

See: https://tc39.es/ecma262/#sec-toint32

pub fn to_length(&mut self, value: &Value) -> Result<usize, Value>

Converts argument to an integer suitable for use as the length of an array-like object.

See: https://tc39.es/ecma262/#sec-tolength

pub fn to_number(&mut self, value: &Value) -> Result<f64, Value>

Converts a value to a double precision floating point.

See: https://tc39.es/ecma262/#sec-tonumber

pub fn to_numeric(&mut self, value: &Value) -> ResultValue

It returns value converted to a numeric value of type Number or BigInt.

See: https://tc39.es/ecma262/#sec-tonumeric

pub(crate) fn to_numeric_number(&mut self, value: &Value) -> Result<f64, Value>

This is a more specialized version of to_numeric.

It returns value converted to a numeric value of type Number.

See: https://tc39.es/ecma262/#sec-tonumeric

pub(crate) fn extract_array_properties(
&mut self,
value: &Value
) -> Result<Vec<Value>, ()>

Converts an array object into a rust vector of values.

This is useful for the spread operator, for any other object an Err is returned

pub(crate) fn ordinary_to_primitive(
&mut self,
o: &Value,
hint: PreferredType
) -> ResultValue

Converts an object to a primitive.

More information:

ECMAScript

pub(crate) fn to_primitive(
&mut self,
input: &Value,
preferred_type: PreferredType
) -> ResultValue

The abstract operation ToPrimitive takes an input argument and an optional argument PreferredType.

https://tc39.es/ecma262/#sec-toprimitive

pub(crate) fn to_property_key(&mut self, value: &Value) -> ResultValue

The abstract operation ToPropertyKey takes argument argument. It converts argument to a value that can be used as a property key.

https://tc39.es/ecma262/#sec-topropertykey

pub(crate) fn has_property(&self, obj: &Value, key: &Value) -> bool

https://tc39.es/ecma262/#sec-hasproperty

pub(crate) fn to_object(&mut self, value: &Value) -> ResultValue

The abstract operation ToObject converts argument to a value of type Object https://tc39.es/ecma262/#sec-toobject

fn set_value(&mut self, node: &Node, value: Value) -> ResultValue

pub(crate) fn set_current_state(&mut self, new_state: InterpreterState)

pub(crate) fn get_current_state(&self) -> &InterpreterState

pub fn require_object_coercible<'a>(
&mut self,
value: &'a Value
) -> Result<&'a Value, Value>

Check if the Value can be converted to an Object

The abstract operation RequireObjectCoercible takes argument argument. It throws an error if argument is a value that cannot be converted to an Object using ToObject. It is defined by Table 15

More information:

ECMAScript reference