Module boa::object::internal_methods::function

Statics

CONSTRUCTOR_INTERNAL_METHODS

FUNCTION_INTERNAL_METHODS

Definitions of the internal object methods for function objects.

Functions

call_construct

Internal implementation of call and construct.

function_call

Call this object.

function_construct

Construct an instance of this object with the specified arguments.