Struct boa::object::ConstructorBuilder

pub struct ConstructorBuilder<'context> {
    context: &'context mut Context,
    constructor_function: NativeFunctionSignature,
    constructor_object: JsObject,
    constructor_has_prototype: bool,
    prototype: JsObject,
    name: JsString,
    length: usize,
    callable: bool,
    constructor: bool,
    inherit: Option<JsPrototype>,
    custom_prototype: Option<JsPrototype>,
}

Builder for creating constructors objects, like Array.

Fields
context: &'context mut Contextconstructor_function: NativeFunctionSignatureconstructor_object: JsObjectconstructor_has_prototype: boolprototype: JsObjectname: JsStringlength: usizecallable: boolconstructor: boolinherit: Option<JsPrototype>custom_prototype: Option<JsPrototype>
Implementations

impl<'context> ConstructorBuilder<'context>

pub fn new(
context: &'context mut Context,
constructor: NativeFunctionSignature
) -> Self

Create a new ConstructorBuilder.

pub(crate) fn with_standard_object(
context: &'context mut Context,
constructor: NativeFunctionSignature,
object: StandardConstructor
) -> Self

pub fn method<B>(
&mut self,
function: NativeFunctionSignature,
binding: B,
length: usize
) -> &mut Self where
B: Into<FunctionBinding>,

Add new method to the constructors prototype.

pub fn static_method<B>(
&mut self,
function: NativeFunctionSignature,
binding: B,
length: usize
) -> &mut Self where
B: Into<FunctionBinding>,

Add new static method to the constructors object itself.

pub fn property<K, V>(
&mut self,
key: K,
value: V,
attribute: Attribute
) -> &mut Self where
K: Into<PropertyKey>,
V: Into<JsValue>,

Add new data property to the constructor’s prototype.

pub fn static_property<K, V>(
&mut self,
key: K,
value: V,
attribute: Attribute
) -> &mut Self where
K: Into<PropertyKey>,
V: Into<JsValue>,

Add new static data property to the constructor object itself.

pub fn accessor<K>(
&mut self,
key: K,
get: Option<JsObject>,
set: Option<JsObject>,
attribute: Attribute
) -> &mut Self where
K: Into<PropertyKey>,

Add new accessor property to the constructor’s prototype.

pub fn static_accessor<K>(
&mut self,
key: K,
get: Option<JsObject>,
set: Option<JsObject>,
attribute: Attribute
) -> &mut Self where
K: Into<PropertyKey>,

Add new static accessor property to the constructor object itself.

pub fn property_descriptor<K, P>(&mut self, key: K, property: P) -> &mut Self where
K: Into<PropertyKey>,
P: Into<PropertyDescriptor>,

Add new property to the constructor’s prototype.

pub fn static_property_descriptor<K, P>(
&mut self,
key: K,
property: P
) -> &mut Self where
K: Into<PropertyKey>,
P: Into<PropertyDescriptor>,

Add new static property to the constructor object itself.

pub fn length(&mut self, length: usize) -> &mut Self

Specify how many arguments the constructor function takes.

Default is 0.

pub fn name<N>(&mut self, name: N) -> &mut Self where
N: AsRef<str>,

Specify the name of the constructor function.

Default is "[object]"

pub fn callable(&mut self, callable: bool) -> &mut Self

Specify whether the constructor function can be called.

Default is true

pub fn constructor(&mut self, constructor: bool) -> &mut Self

Specify whether the constructor function can be called with new keyword.

Default is true

pub fn inherit<O: Into<JsPrototype>>(&mut self, prototype: O) -> &mut Self

Specify the prototype this constructor object inherits from.

Default is Object.prototype

pub fn custom_prototype<O: Into<JsPrototype>>(
&mut self,
prototype: O
) -> &mut Self

Specify the proto for this constructor.

Default is Function.prototype

pub fn constructor_has_prototype(
&mut self,
constructor_has_prototype: bool
) -> &mut Self

Specify whether the constructor function has a ‘prototype’ property.

Default is true

pub fn context(&mut self) -> &mut Context

Return the current context.

pub fn build(&mut self) -> JsObject

Build the constructor function object.