Struct boa::object::GcObject

pub struct GcObject(Gc<GcCell<Object>>);

Garbage collected Object.

Implementations

impl GcObject

pub fn new(object: Object) -> Self

Create a new GcObject from a Object.

pub fn borrow(&self) -> Ref<'_, Object>

Immutably borrows the Object.

The borrow lasts until the returned Ref exits scope. Multiple immutable borrows can be taken out at the same time.

Panics

Panics if the object is currently mutably borrowed.

pub fn borrow_mut(&self) -> RefMut<'_, Object, Object>

Mutably borrows the Object.

The borrow lasts until the returned RefMut exits scope. The object cannot be borrowed while this borrow is active.

Panics

Panics if the object is currently borrowed.

pub fn try_borrow(&self) -> StdResult<Ref<'_, Object>, BorrowError>

Immutably borrows the Object, returning an error if the value is currently mutably borrowed.

The borrow lasts until the returned GcCellRef exits scope. Multiple immutable borrows can be taken out at the same time.

This is the non-panicking variant of borrow.

pub fn try_borrow_mut(
&self
) -> StdResult<RefMut<'_, Object, Object>, BorrowMutError>

Mutably borrows the object, returning an error if the value is currently borrowed.

The borrow lasts until the returned GcCellRefMut exits scope. The object be borrowed while this borrow is active.

This is the non-panicking variant of borrow_mut.

pub fn equals(lhs: &Self, rhs: &Self) -> bool

Checks if the garbage collected memory is the same.

fn call_construct(
&self,
this_target: &JsValue,
args: &[JsValue],
context: &mut Context,
construct: bool
) -> Result<JsValue>

Internal implementation of call and construct.

Panics

Panics if the object is currently mutably borrowed.

https://tc39.es/ecma262/#sec-prepareforordinarycall https://tc39.es/ecma262/#sec-ordinarycallbindthis https://tc39.es/ecma262/#sec-runtime-semantics-evaluatebody https://tc39.es/ecma262/#sec-ordinarycallevaluatebody

pub fn call(
&self,
this: &JsValue,
args: &[JsValue],
context: &mut Context
) -> Result<JsValue>

Call this object.

Panics

Panics if the object is currently mutably borrowed.

pub fn construct(
&self,
args: &[JsValue],
new_target: &JsValue,
context: &mut Context
) -> Result<JsValue>

Construct an instance of this object with the specified arguments.

Panics

Panics if the object is currently mutably borrowed.

pub(crate) fn ordinary_to_primitive(
&self,
context: &mut Context,
hint: PreferredType
) -> Result<JsValue>

Converts an object to a primitive.

Diverges from the spec to prevent a stack overflow when the object is recursive. For example,

let a = [1];
a[1] = a;
console.log(a.toString()); // We print "1,"

The spec doesn’t mention what to do in this situation, but a naive implementation would overflow the stack recursively calling toString(). We follow v8 and SpiderMonkey instead by returning a default value for the given hint – either 0. or "". Example in v8: https://repl.it/repls/IvoryCircularCertification#index.js

More information:

ECMAScript

pub(crate) fn to_json(&self, context: &mut Context) -> Result<Option<JSONValue>>

Converts an object to JSON, checking for reference cycles and throwing a TypeError if one is found

pub fn is<T>(&self) -> bool where
T: NativeObject,

Return true if it is a native object and the native type is T.

Panics

Panics if the object is currently mutably borrowed.

pub fn downcast_ref<T>(&self) -> Option<Ref<'_, T>> where
T: NativeObject,

Downcast a reference to the object, if the object is type native object type T.

Panics

Panics if the object is currently mutably borrowed.

pub fn downcast_mut<T>(&mut self) -> Option<RefMut<'_, Object, T>> where
T: NativeObject,

Downcast a mutable reference to the object, if the object is type native object type T.

Panics

Panics if the object is currently borrowed.

pub fn prototype_instance(&self) -> JsValue

Get the prototype of the object.

Panics

Panics if the object is currently mutably borrowed.

pub fn set_prototype_instance(&self, prototype: JsValue) -> bool

Set the prototype of the object.

Panics

Panics if the object is currently mutably borrowed or if th prototype is not an object or undefined.

pub fn is_array(&self) -> bool

Checks if it an Array object.

Panics

Panics if the object is currently mutably borrowed.

pub fn is_array_iterator(&self) -> bool

Checks if it is an ArrayIterator object.

Panics

Panics if the object is currently mutably borrowed.

pub fn is_map(&self) -> bool

Checks if it is a Map object.pub

Panics

Panics if the object is currently mutably borrowed.

pub fn is_string(&self) -> bool

Checks if it a String object.

Panics

Panics if the object is currently mutably borrowed.

pub fn is_function(&self) -> bool

Checks if it a Function object.

Panics

Panics if the object is currently mutably borrowed.

pub fn is_symbol(&self) -> bool

Checks if it a Symbol object.

Panics

Panics if the object is currently mutably borrowed.

pub fn is_error(&self) -> bool

Checks if it an Error object.

Panics

Panics if the object is currently mutably borrowed.

pub fn is_boolean(&self) -> bool

Checks if it a Boolean object.

Panics

Panics if the object is currently mutably borrowed.

pub fn is_number(&self) -> bool

Checks if it a Number object.

Panics

Panics if the object is currently mutably borrowed.

pub fn is_bigint(&self) -> bool

Checks if it a BigInt object.

Panics

Panics if the object is currently mutably borrowed.

pub fn is_regexp(&self) -> bool

Checks if it a RegExp object.

Panics

Panics if the object is currently mutably borrowed.

pub fn is_ordinary(&self) -> bool

Checks if it an ordinary object.

Panics

Panics if the object is currently mutably borrowed.

pub fn is_native_object(&self) -> bool

Returns true if it holds an Rust type that implements NativeObject.

Panics

Panics if the object is currently mutably borrowed.

pub fn get_method<K>(
&self,
context: &mut Context,
key: K
) -> Result<Option<GcObject>> where
K: Into<PropertyKey>,

Retrieves value of specific property, when the value of the property is expected to be a function.

More information:

EcmaScript reference

pub(crate) fn ordinary_has_instance(
&self,
context: &mut Context,
value: &JsValue
) -> Result<bool>

Determines if value inherits from the instance object inheritance path.

More information:

EcmaScript reference

pub(crate) fn species_constructor(
&self,
default_constructor: JsValue,
context: &mut Context
) -> Result<JsValue>

7.3.22 SpeciesConstructor ( O, defaultConstructor )

The abstract operation SpeciesConstructor takes arguments O (an Object) and defaultConstructor (a constructor). It is used to retrieve the constructor that should be used to create new objects that are derived from O. defaultConstructor is the constructor to use if a constructor @@species property cannot be found starting from O.

More information:

ECMAScript reference

pub fn to_property_descriptor(
&self,
context: &mut Context
) -> Result<PropertyDescriptor>

pub fn copy_data_properties<K>(
&mut self,
source: &JsValue,
excluded_keys: Vec<K>,
context: &mut Context
) -> Result<()> where
K: Into<PropertyKey>,

7.3.25 CopyDataProperties ( target, source, excludedItems )

More information:

ECMAScript

impl GcObject

pub fn has_property<K>(&self, key: K, _context: &mut Context) -> Result<bool> where
K: Into<PropertyKey>,

Check if object has property.

More information:

ECMAScript reference

pub fn is_extensible(&self, _context: &mut Context) -> Result<bool>

Check if it is extensible.

More information:

ECMAScript reference

pub fn delete<K>(&self, key: K) -> bool where
K: Into<PropertyKey>,

Delete property, if deleted return true.

pub fn delete_property_or_throw<K>(
&self,
key: K,
context: &mut Context
) -> Result<bool> where
K: Into<PropertyKey>,

Defines the property or throws a TypeError if the operation fails.

More information:

EcmaScript reference

pub fn has_own_property<K>(
&self,
key: K,
_context: &mut Context
) -> Result<bool> where
K: Into<PropertyKey>,

Check if object has an own property.

More information:

ECMAScript reference

pub fn get<K>(&self, key: K, context: &mut Context) -> Result<JsValue> where
K: Into<PropertyKey>,

Get property from object or throw.

More information:

ECMAScript reference

pub fn set<K, V>(
&self,
key: K,
value: V,
throw: bool,
context: &mut Context
) -> Result<bool> where
K: Into<PropertyKey>,
V: Into<JsValue>,

set property of object or throw if bool flag is passed.

More information:

ECMAScript reference

pub fn define_property_or_throw<K, P>(
&self,
key: K,
desc: P,
context: &mut Context
) -> Result<bool> where
K: Into<PropertyKey>,
P: Into<PropertyDescriptor>,

Define property or throw.

More information:

ECMAScript reference

pub fn create_data_property<K, V>(
&self,
key: K,
value: V,
context: &mut Context
) -> Result<bool> where
K: Into<PropertyKey>,
V: Into<JsValue>,

Create data property

More information:

ECMAScript reference

pub fn create_data_property_or_throw<K, V>(
&self,
key: K,
value: V,
context: &mut Context
) -> Result<bool> where
K: Into<PropertyKey>,
V: Into<JsValue>,

Create data property or throw

More information:

ECMAScript reference

pub(crate) fn __has_property__(&self, key: &PropertyKey) -> bool

[[hasProperty]]

pub(crate) fn __is_extensible__(&self) -> bool

Check if it is extensible.

More information:

ECMAScript reference

pub fn __prevent_extensions__(&mut self) -> bool

Disable extensibility.

More information:

ECMAScript reference

pub(crate) fn __delete__(&self, key: &PropertyKey) -> bool

Delete property.

pub fn __get__(
&self,
key: &PropertyKey,
receiver: JsValue,
context: &mut Context
) -> Result<JsValue>

[[Get]]

pub fn __set__(
&self,
key: PropertyKey,
value: JsValue,
receiver: JsValue,
context: &mut Context
) -> Result<bool>

[[Set]]

pub fn __define_own_property__(
&self,
key: PropertyKey,
desc: PropertyDescriptor,
context: &mut Context
) -> Result<bool>

[[defineOwnProperty]]

pub fn ordinary_define_own_property(
&self,
key: PropertyKey,
desc: PropertyDescriptor
) -> bool

Define an own property for an ordinary object.

More information:

ECMAScript reference

fn array_define_own_property(
&self,
key: PropertyKey,
desc: PropertyDescriptor,
context: &mut Context
) -> Result<bool>

Define an own property for an array.

More information:

ECMAScript reference

pub fn __get_own_property__(
&self,
key: &PropertyKey
) -> Option<PropertyDescriptor>

Gets own property of ‘Object’

pub fn string_get_own_property(
&self,
key: &PropertyKey
) -> Option<PropertyDescriptor>

StringGetOwnProperty abstract operation

More information:

ECMAScript reference

pub fn string_exotic_get_own_property(
&self,
key: &PropertyKey
) -> Option<PropertyDescriptor>

Gets own property of ‘String’ exotic object

More information:

ECMAScript reference

pub fn ordinary_get_own_property(
&self,
key: &PropertyKey
) -> Option<PropertyDescriptor>

The specification returns a Property Descriptor or Undefined.

These are 2 separate types and we can’t do that here.

More information:

ECMAScript reference

pub fn own_property_keys(&self) -> Vec<PropertyKey>

Essential internal method OwnPropertyKeys

More information:

ECMAScript reference

pub fn define_properties(
&mut self,
props: JsValue,
context: &mut Context
) -> Result<()>

The abstract operation ObjectDefineProperties

More information:

ECMAScript reference

pub fn __set_prototype_of__(&mut self, val: JsValue) -> bool

Object.setPropertyOf(obj, prototype)

This method sets the prototype (i.e., the internal [[Prototype]] property) of a specified object to another object or null.

More information:

ECMAScript reference

MDN documentation

pub fn __get_prototype_of__(&self) -> JsValue

Returns either the prototype or null

More information:

ECMAScript reference

MDN documentation

pub(crate) fn insert<K, P>(
&self,
key: K,
property: P
) -> Option<PropertyDescriptor> where
K: Into<PropertyKey>,
P: Into<PropertyDescriptor>,

Helper function for property insertion.

pub(crate) fn remove(&self, key: &PropertyKey) -> Option<PropertyDescriptor>

Helper function for property removal.

pub fn insert_property<K, P>(
&self,
key: K,
property: P
) -> Option<PropertyDescriptor> where
K: Into<PropertyKey>,
P: Into<PropertyDescriptor>,

Inserts a field in the object properties without checking if it’s writable.

If a field was already in the object with the same name that a Some is returned with that field, otherwise None is returned.

pub fn is_callable(&self) -> bool

It determines if Object is a callable function with a [[Call]] internal method.

More information:

EcmaScript reference

pub fn is_constructable(&self) -> bool

It determines if Object is a function object with a [[Construct]] internal method.

More information:

EcmaScript reference

pub fn is_global(&self) -> bool

Returns true if the GcObject is the global for a Realm

pub(crate) fn create_list_from_array_like(
&self,
element_types: &[Type],
context: &mut Context
) -> Result<Vec<JsValue>>

It is used to create List value whose elements are provided by the indexed properties of self.

More information:

EcmaScript reference

pub(crate) fn enumerable_own_property_names(
&self,
kind: PropertyNameKind,
context: &mut Context
) -> Result<Vec<JsValue>>

It is used to iterate over names of object’s keys.

More information:

EcmaScript reference

pub(crate) fn length_of_array_like(
&self,
context: &mut Context
) -> Result<usize>