Struct boa::object::Object

pub struct Object {
    pub data: ObjectData,
    properties: PropertyMap,
    prototype: JsValue,
    extensible: bool,
}

The internal representation of an JavaScript object.

Fields
data: ObjectData

The type of the object.

properties: PropertyMapprototype: JsValue

Instance prototype __proto__.

extensible: bool

Whether it can have new properties added to it.

Implementations

impl Object

pub(crate) fn insert<K, P>(
&mut self,
key: K,
property: P
) -> Option<PropertyDescriptor> where
K: Into<PropertyKey>,
P: Into<PropertyDescriptor>,

Helper function for property insertion.

pub(crate) fn remove(&mut self, key: &PropertyKey) -> Option<PropertyDescriptor>

Helper function for property removal.

pub fn insert_property<K, P>(
&mut self,
key: K,
property: P
) -> Option<PropertyDescriptor> where
K: Into<PropertyKey>,
P: Into<PropertyDescriptor>,

Inserts a field in the object properties without checking if it’s writable.

If a field was already in the object with the same name that a Some is returned with that field, otherwise None is retuned.

impl Object

pub fn new() -> Self

pub fn function(function: Function, prototype: JsValue) -> Self

Return a new ObjectData struct, with kind set to Ordinary

pub fn create(proto: JsValue) -> Self

ObjectCreate is used to specify the runtime creation of new ordinary objects.

More information:

ECMAScript reference

pub fn boolean(value: bool) -> Self

Return a new Boolean object whose [[BooleanData]] internal slot is set to argument.

pub fn number(value: f64) -> Self

Return a new Number object whose [[NumberData]] internal slot is set to argument.

pub fn string<S>(value: S) -> Self where
S: Into<JsString>,

Return a new String object whose [[StringData]] internal slot is set to argument.

pub fn bigint(value: JsBigInt) -> Self

Return a new BigInt object whose [[BigIntData]] internal slot is set to argument.

pub fn native_object<T>(value: T) -> Self where
T: NativeObject,

Create a new native object of type T.

pub fn is_callable(&self) -> bool

It determines if Object is a callable function with a [[Call]] internal method.

More information:

EcmaScript reference

pub fn is_constructable(&self) -> bool

It determines if Object is a function object with a [[Construct]] internal method.

More information:

EcmaScript reference

pub fn is_array(&self) -> bool

Checks if it an Array object.

pub fn as_array(&self) -> Option<()>

pub fn is_array_iterator(&self) -> bool

Checks if it is an ArrayIterator object.

pub fn as_array_iterator(&self) -> Option<&ArrayIterator>

pub fn as_array_iterator_mut(&mut self) -> Option<&mut ArrayIterator>

pub fn as_string_iterator_mut(&mut self) -> Option<&mut StringIterator>

pub fn as_regexp_string_iterator_mut(
&mut self
) -> Option<&mut RegExpStringIterator>

pub fn as_for_in_iterator(&self) -> Option<&ForInIterator>

pub fn as_for_in_iterator_mut(&mut self) -> Option<&mut ForInIterator>

pub fn is_map(&self) -> bool

Checks if it is a Map object.pub

pub fn as_map_ref(&self) -> Option<&OrderedMap<JsValue>>

pub fn as_map_mut(&mut self) -> Option<&mut OrderedMap<JsValue>>

pub fn as_map_iterator_mut(&mut self) -> Option<&mut MapIterator>

pub fn is_set(&self) -> bool

pub fn as_set_ref(&self) -> Option<&OrderedSet<JsValue>>

pub fn as_set_mut(&mut self) -> Option<&mut OrderedSet<JsValue>>

pub fn as_set_iterator_mut(&mut self) -> Option<&mut SetIterator>

pub fn is_string(&self) -> bool

Checks if it a String object.

pub fn as_string(&self) -> Option<JsString>

pub fn is_function(&self) -> bool

Checks if it a Function object.

pub fn as_function(&self) -> Option<&Function>

pub fn is_symbol(&self) -> bool

Checks if it a Symbol object.

pub fn as_symbol(&self) -> Option<JsSymbol>

pub fn is_error(&self) -> bool

Checks if it an Error object.

pub fn as_error(&self) -> Option<()>

pub fn is_boolean(&self) -> bool

Checks if it a Boolean object.

pub fn as_boolean(&self) -> Option<bool>

pub fn is_number(&self) -> bool

Checks if it a Number object.

pub fn as_number(&self) -> Option<f64>

pub fn is_bigint(&self) -> bool

Checks if it a BigInt object.

pub fn as_bigint(&self) -> Option<&JsBigInt>

pub fn is_regexp(&self) -> bool

Checks if it a RegExp object.

pub fn as_regexp(&self) -> Option<&RegExp>

pub fn is_ordinary(&self) -> bool

Checks if it an ordinary object.

pub fn prototype_instance(&self) -> &JsValue

pub fn set_prototype_instance(&mut self, prototype: JsValue) -> bool

Sets the prototype instance of the object.

More information

pub fn with_prototype(proto: JsValue, data: ObjectData) -> Object

Similar to Value::new_object, but you can pass a prototype to create from, plus a kind

pub fn is_native_object(&self) -> bool

Returns true if it holds an Rust type that implements NativeObject.

pub fn as_native_object(&self) -> Option<&dyn NativeObject>

pub fn is<T>(&self) -> bool where
T: NativeObject,

Reeturn true if it is a native object and the native type is T.

pub fn downcast_ref<T>(&self) -> Option<&T> where
T: NativeObject,

Downcast a reference to the object, if the object is type native object type T.

pub fn downcast_mut<T>(&mut self) -> Option<&mut T> where
T: NativeObject,

Downcast a mutable reference to the object, if the object is type native object type T.

pub fn properties(&self) -> &PropertyMap