Struct boa::object::ObjectData

pub struct ObjectData {
    kind: ObjectKind,
    internal_methods: &'static InternalObjectMethods,
}

Defines the kind of an object and its internal methods

Fields
kind: ObjectKindinternal_methods: &'static InternalObjectMethods
Implementations

impl ObjectData

pub fn array() -> Self

Create the Array object data and reference its exclusive internal methods

pub fn array_iterator(array_iterator: ArrayIterator) -> Self

Create the ArrayIterator object data

pub fn array_buffer(array_buffer: ArrayBuffer) -> Self

Create the ArrayBuffer object data

pub fn map(map: OrderedMap<JsValue>) -> Self

Create the Map object data

pub fn map_iterator(map_iterator: MapIterator) -> Self

Create the MapIterator object data

pub fn reg_exp(reg_exp: Box<RegExp>) -> Self

Create the RegExp object data

pub fn reg_exp_string_iterator(
reg_exp_string_iterator: RegExpStringIterator
) -> Self

Create the RegExpStringIterator object data

pub fn big_int(big_int: JsBigInt) -> Self

Create the BigInt object data

pub fn boolean(boolean: bool) -> Self

Create the Boolean object data

pub fn data_view(data_view: DataView) -> Self

Create the DataView object data

pub fn for_in_iterator(for_in_iterator: ForInIterator) -> Self

Create the ForInIterator object data

pub fn function(function: Function) -> Self

Create the Function object data

pub fn bound_function(bound_function: BoundFunction, constructor: bool) -> Self

Create the BoundFunction object data

pub fn set(set: OrderedSet<JsValue>) -> Self

Create the Set object data

pub fn set_iterator(set_iterator: SetIterator) -> Self

Create the SetIterator object data

pub fn string(string: JsString) -> Self

Create the String object data and reference its exclusive internal methods

pub fn string_iterator(string_iterator: StringIterator) -> Self

Create the StringIterator object data

pub fn number(number: f64) -> Self

Create the Number object data

pub fn symbol(symbol: JsSymbol) -> Self

Create the Symbol object data

pub fn error() -> Self

Create the Error object data

pub fn ordinary() -> Self

Create the Ordinary object data

pub fn proxy(proxy: Proxy, call: bool, construct: bool) -> Self

Create the Proxy object data

pub fn date(date: Date) -> Self

Create the Date object data

pub fn global() -> Self

Create the Global object data

pub fn arguments(arguments: Arguments) -> Self

Create the Arguments object data

pub fn native_object(native_object: Box<dyn NativeObject>) -> Self

Create the NativeObject object data

pub fn integer_indexed(integer_indexed: IntegerIndexed) -> Self

Creates the IntegerIndexed object data