Struct boa::property::attribute::Attribute

pub struct Attribute {
    bits: u8,
}

This struct constains the property flags as describen in the ECMAScript specification.

It contains the following flags:

[[Writable]] (WRITABLE) - If false, attempts by ECMAScript code to change the property’s [[Value]] attribute using [[Set]] will not succeed.

[[Enumerable]] (ENUMERABLE) - If the property will be enumerated by a for-in enumeration.

[[Configurable]] (CONFIGURABLE) - If false, attempts to delete the property, change the property to be an accessor property, or change its attributes (other than [[Value]], or changing [[Writable]] to false) will fail.

Fields
bits: u8
Implementations

impl Attribute

pub const WRITABLE: Attribute

The Writable attribute decides whether the value associated with the property can be changed or not, from its initial value.

pub const ENUMERABLE: Attribute

If the property can be enumerated by a for-in loop.

pub const CONFIGURABLE: Attribute

If the property descriptor can be changed later.

pub const READONLY: Attribute

The property is not writable.

pub const NON_ENUMERABLE: Attribute

The property can not be enumerated in a for-in loop.

pub const PERMANENT: Attribute

The property descriptor cannot be changed.

pub const fn empty() -> Attribute

Returns an empty set of flags

pub const fn all() -> Attribute

Returns the set containing all flags.

pub const fn bits(&self) -> u8

Returns the raw value of the flags currently stored.

pub fn from_bits(bits: u8) -> Option<Attribute>

Convert from underlying bit representation, unless that representation contains bits that do not correspond to a flag.

pub const fn from_bits_truncate(bits: u8) -> Attribute

Convert from underlying bit representation, dropping any bits that do not correspond to flags.

pub const unsafe fn from_bits_unchecked(bits: u8) -> Attribute

Convert from underlying bit representation, preserving all bits (even those not corresponding to a defined flag).

pub const fn is_empty(&self) -> bool

Returns true if no flags are currently stored.

pub const fn is_all(&self) -> bool

Returns true if all flags are currently set.

pub const fn intersects(&self, other: Attribute) -> bool

Returns true if there are flags common to both self and other.

pub const fn contains(&self, other: Attribute) -> bool

Returns true all of the flags in other are contained within self.

pub fn insert(&mut self, other: Attribute)

Inserts the specified flags in-place.

pub fn remove(&mut self, other: Attribute)

Removes the specified flags in-place.

pub fn toggle(&mut self, other: Attribute)

Toggles the specified flags in-place.

pub fn set(&mut self, other: Attribute, value: bool)

Inserts or removes the specified flags depending on the passed value.

impl Attribute

pub fn clear(&mut self)

Clear all flags.

pub fn set_writable(&mut self, value: bool)

Sets the writable flag.

pub fn writable(self) -> bool

Gets the writable flag.

pub fn set_enumerable(&mut self, value: bool)

Sets the enumerable flag.

pub fn enumerable(self) -> bool

Gets the enumerable flag.

pub fn set_configurable(&mut self, value: bool)

Sets the configurable flag.

pub fn configurable(self) -> bool

Gets the configurable flag.