Module boa::realm

Conceptually, a realm consists of a set of intrinsic objects, an ECMAScript global environment, all of the ECMAScript code that is loaded within the scope of that global environment, and other associated state and resources.

A realm is represented in this implementation as a Realm struct with the fields specified from the spec.

Structs

Realm
Representation of a Realm.