Struct boa::Context

pub struct Context {
    realm: Realm,
    executor: Interpreter,
    symbol_count: u64,
    console: Console,
    well_known_symbols: WellKnownSymbols,
    iterator_prototypes: IteratorPrototypes,
    standard_objects: StandardObjects,
}

Javascript context. It is the primary way to interact with the runtime.

Contexts constructed in a thread share the same runtime, therefore it is possible to share objects from one context to another context, but they have to be in the same thread.

Fields
realm: Realm

realm holds both the global object and the environment

executor: Interpreter

The current executor.

symbol_count: u64

Symbol hash.

For now this is an incremented u64 number.

console: Console

console object state.

well_known_symbols: WellKnownSymbols

Cached well known symbols

iterator_prototypes: IteratorPrototypes

Cached iterator prototypes.

standard_objects: StandardObjects

Cached standard objects and their prototypes

Implementations

impl Context

pub fn new() -> Self

Create a new Context.

pub fn realm(&self) -> &Realm

pub fn realm_mut(&mut self) -> &mut Realm

pub fn executor(&mut self) -> &mut Interpreter

pub(crate) fn console(&self) -> &Console

A helper function for getting an immutable reference to the console object.

pub(crate) fn console_mut(&mut self) -> &mut Console

A helper function for getting a mutable reference to the console object.

fn create_intrinsics(&mut self)

Sets up the default global objects within Global

fn generate_hash(&mut self) -> u64

Generates a new Symbol internal hash.

This currently is an incremented value.

pub fn construct_symbol(&mut self, description: Option<RcString>) -> RcSymbol

Construct a new Symbol with an optional description.

pub fn construct_object(&self) -> GcObject

Construct an empty object.

pub(crate) fn call(
&mut self,
f: &Value,
this: &Value,
args: &[Value]
) -> Result<Value>

https://tc39.es/ecma262/#sec-call

pub fn global_object(&self) -> &Value

Return the global object.

pub fn construct_range_error<M>(&mut self, message: M) -> Value where
M: Into<String>,

Constructs a RangeError with the specified message.

pub fn throw_range_error<M>(&mut self, message: M) -> Result<Value> where
M: Into<String>,

Throws a RangeError with the specified message.

pub fn construct_type_error<M>(&mut self, message: M) -> Value where
M: Into<String>,

Constructs a TypeError with the specified message.

pub fn throw_type_error<M>(&mut self, message: M) -> Result<Value> where
M: Into<String>,

Throws a TypeError with the specified message.

pub fn construct_reference_error<M>(&mut self, message: M) -> Value where
M: Into<String>,

Constructs a ReferenceError with the specified message.

pub fn throw_reference_error<M>(&mut self, message: M) -> Result<Value> where
M: Into<String>,

Throws a ReferenceError with the specified message.

pub fn construct_syntax_error<M>(&mut self, message: M) -> Value where
M: Into<String>,

Constructs a SyntaxError with the specified message.

pub fn throw_syntax_error<M>(&mut self, message: M) -> Result<Value> where
M: Into<String>,

Throws a SyntaxError with the specified message.

pub fn construct_eval_error<M>(&mut self, message: M) -> Value where
M: Into<String>,

Constructs a EvalError with the specified message.

pub fn construct_uri_error<M>(&mut self, message: M) -> Value where
M: Into<String>,

Constructs a URIError with the specified message.

pub fn throw_eval_error<M>(&mut self, message: M) -> Result<Value> where
M: Into<String>,

Throws a EvalError with the specified message.

pub fn throw_uri_error<M>(&mut self, message: M) -> Result<Value> where
M: Into<String>,

Throws a URIError with the specified message.

pub(crate) fn create_function<P, B>(
&mut self,
params: P,
body: B,
flags: FunctionFlags
) -> Value where
P: Into<Box<[FormalParameter]>>,
B: Into<StatementList>,

Utility to create a function Value for Function Declarations, Arrow Functions or Function Expressions

pub fn create_builtin_function(
&mut self,
name: &str,
length: usize,
body: NativeFunction
) -> Result<GcObject>

Create a new builin function.

pub fn register_global_function(
&mut self,
name: &str,
length: usize,
body: NativeFunction
) -> Result<()>

Register a global function.

pub(crate) fn extract_array_properties(
&mut self,
value: &Value
) -> StdResult<Vec<Value>, ()>

Converts an array object into a rust vector of values.

This is useful for the spread operator, for any other object an Err is returned TODO: Not needed for spread of arrays. Check in the future for Map and remove if necessary

pub(crate) fn has_property(&self, obj: &Value, key: &PropertyKey) -> bool

https://tc39.es/ecma262/#sec-hasproperty

pub(crate) fn set_value(&mut self, node: &Node, value: Value) -> Result<Value>

pub fn register_global_class<T>(&mut self) -> Result<()> where
T: Class,

Register a global class of type T, where T implements Class.

Example

#[derive(Debug, Trace, Finalize)]
struct MyClass;

impl Class for MyClass {
   // ...
}

context.register_global_class::<MyClass>();

pub fn register_global_property<K, V>(
&mut self,
key: K,
value: V,
attribute: Attribute
) where
K: Into<PropertyKey>,
V: Into<Value>,

Register a global property.

Example

use boa::{Context, property::Attribute, object::ObjectInitializer};

let mut context = Context::new();

context.register_global_property("myPrimitiveProperty", 10, Attribute::all());

let object = ObjectInitializer::new(&mut context)
   .property("x", 0, Attribute::all())
   .property("y", 1, Attribute::all())
   .build();
context.register_global_property("myObjectProperty", object, Attribute::all());

pub fn eval(&mut self, src: &str) -> Result<Value>

Evaluates the given code.

Examples

let mut context = Context::new();

let value = context.eval("1 + 3").unwrap();

assert!(value.is_number());
assert_eq!(value.as_number().unwrap(), 4.0);

pub fn well_known_symbols(&self) -> &WellKnownSymbols

Returns a structure that contains the JavaScript well known symbols.

Examples

let mut context = Context::new();

let iterator = context.well_known_symbols().iterator_symbol();
assert_eq!(iterator.description(), Some("Symbol.iterator"));

This is equivalent to let iterator = Symbol.iterator in JavaScript.

pub fn iterator_prototypes(&self) -> &IteratorPrototypes

Return the cached iterator prototypes.

pub fn standard_objects(&self) -> &StandardObjects

Return the core standard objects.