Struct boa::JsString

pub struct JsString {
    inner: NonNull<Inner>,
    _marker: PhantomData<Rc<str>>,
}

This represents a JavaScript primitive string.

This is similar to Rc<str>. But unlike Rc<str> which stores the length on the stack and a pointer to the data (this is also known as fat pointers). The JsString length and data is stored on the heap. and just an non-null pointer is kept, so its size is the size of a pointer.

Fields
inner: NonNull<Inner>_marker: PhantomData<Rc<str>>
Implementations

impl JsString

pub fn new<S: AsRef<str>>(s: S) -> Self

Create a new JavaScript string.

pub fn concat<T, U>(x: T, y: U) -> JsString where
T: AsRef<str>,
U: AsRef<str>,

Concatinate two string.

fn inner(&self) -> &Inner

Return the inner representation.

pub fn as_str(&self) -> &str

Return the JavaScript string as a rust &str.

pub fn refcount(this: &Self) -> usize

Gets the number of JsStrings which point to this allocation.

pub fn ptr_eq(x: &Self, y: &Self) -> bool

Returns true if the two JsStrings point to the same allocation (in a vein similar to ptr::eq).

pub(crate) fn index_of(
&self,
search_value: &Self,
from_index: usize
) -> Option<usize>

6.1.4.1 StringIndexOf ( string, searchValue, fromIndex )

Note: Instead of returning an isize with -1 as the “not found” value, We make use of the type system and return Option with None as the “not found” value.

More information:

ECMAScript reference