Module boa::syntax::ast::node

This module implements the Node structure, which composes the AST.

Structs

FormalParameter
"Formal parameter" is a fancy way of saying "function parameter".

Enums

MethodDefinitionKind
Method definition kinds.

Node
A Javascript AST Node.

PropertyDefinition
A JavaScript property is a characteristic of an object, often describing attributes associated with a data structure.

Functions

join_nodes
Utility to join multiple Nodes into a single string.