Module boa::syntax::ast::node::declaration

Declaration nodes

Re-exports

pub use self::arrow_function_decl::ArrowFunctionDecl;

pub use self::async_function_decl::AsyncFunctionDecl;

pub use self::async_function_expr::AsyncFunctionExpr;

pub use self::function_decl::FunctionDecl;

pub use self::function_expr::FunctionExpr;

Modules

arrow_function_decl

async_function_decl

Async Function Declaration.

async_function_expr

Async Function Expression.

function_decl

function_expr

generator_decl

generator_expr

Structs

DeclarationPatternArray

DeclarationPatternArray represents an array binding pattern.

DeclarationPatternObject

DeclarationPatternObject represents an object binding pattern.

Enums

BindingPatternTypeArray

BindingPatternTypeArray represents the different types of bindings that an array binding pattern may contain.

BindingPatternTypeObject

BindingPatternTypeObject represents the different types of bindings that an object binding pattern may contain.

Declaration

Declaration represents either an individual binding or a binding pattern.

DeclarationList

DeclarationPattern

DeclarationPattern represents an object or array binding pattern.