Enum boa::syntax::ast::node::declaration::BindingPatternTypeArray

pub enum BindingPatternTypeArray {
    Empty,
    Elision,
    SingleName {
        ident: Box<str>,
        default_init: Option<Node>,
    },
    BindingPattern {
        pattern: DeclarationPattern,
    },
    SingleNameRest {
        ident: Box<str>,
    },
    BindingPatternRest {
        pattern: DeclarationPattern,
    },
}

BindingPatternTypeArray represents the different types of bindings that an array binding pattern may contain.

More information:

ECMAScript reference: 14.3.3 Destructuring Binding Patterns - ArrayBindingPattern

Variants

Empty

Empty represents an empty array binding pattern e.g. [ ].

This may occur because the Elision and BindingRestElement in the first type of array binding pattern are both optional.

More information:

ECMAScript reference: 14.3.3 Destructuring Binding Patterns - ArrayBindingPattern

Elision

Elision represents the elision of an item in the array binding pattern.

An Elision may occur at multiple points in the pattern and may be multiple elisions. This variant strictly represents one elision. If there are multiple, this should be used multiple times.

More information:

ECMAScript reference: 13.2.4 Array Initializer - Elision

SingleName

SingleName represents a SingleNameBinding with an identifier and an optional default initializer.

More information:

ECMAScript reference: 14.3.3 Destructuring Binding Patterns - SingleNameBinding

Fields of SingleName

ident: Box<str>default_init: Option<Node>

BindingPattern

BindingPattern represents a BindingPattern in a BindingElement of an array binding pattern.

The pattern and the optional default initializer are both stored in the DeclarationPattern.

More information:

ECMAScript reference: 14.3.3 Destructuring Binding Patterns - BindingElement

Fields of BindingPattern

pattern: DeclarationPattern

SingleNameRest

SingleNameRest represents a BindingIdentifier in a BindingRestElement of an array binding pattern.

More information:

ECMAScript reference: 14.3.3 Destructuring Binding Patterns - BindingRestElement

Fields of SingleNameRest

ident: Box<str>

BindingPatternRest

SingleNameRest represents a BindingPattern in a BindingRestElement of an array binding pattern.

More information:

ECMAScript reference: 14.3.3 Destructuring Binding Patterns - BindingRestElement

Fields of BindingPatternRest

pattern: DeclarationPattern