Struct boa::syntax::ast::node::declaration::generator_decl::GeneratorDecl

pub struct GeneratorDecl {
    name: Box<str>,
    parameters: Box<[FormalParameter]>,
    body: StatementList,
}

The function* declaration (function keyword followed by an asterisk) defines a generator function, which returns a Generator object.

More information:

ECMAScript reference

MDN documentation

Fields
name: Box<str>parameters: Box<[FormalParameter]>body: StatementList
Implementations

impl GeneratorDecl

pub(in syntax) fn new<N, P, B>(
name: N,
parameters: P,
body: B
) -> Self where
N: Into<Box<str>>,
P: Into<Box<[FormalParameter]>>,
B: Into<StatementList>,

Creates a new generator declaration.

pub fn name(&self) -> &str

Gets the name of the generator declaration.

pub fn parameters(&self) -> &[FormalParameter]

Gets the list of parameters of the generator declaration.

pub fn body(&self) -> &[Node]

Gets the body of the generator declaration.

pub(in syntax::ast::node) fn display(
&self,
f: &mut Formatter<'_>,
indentation: usize
) -> Result

Implements the display formatting with indentation.