Enum boa::syntax::ast::node::Node

pub enum Node {
    ArrayDecl(ArrayDecl),
    ArrowFunctionDecl(ArrowFunctionDecl),
    Assign(Assign),
    BinOp(BinOp),
    Block(Block),
    Break(Option<Box<str>>),
    Call(Call),
    ConditionalOp(Box<Node>, Box<Node>, Box<Node>),
    Const(Const),
    ConstDeclList(ConstDeclList),
    Continue(Option<Box<str>>),
    DoWhileLoop(Box<Node>, Box<Node>),
    FunctionDecl(FunctionDecl),
    FunctionExpr(FunctionExpr),
    GetConstField(Box<Node>, Box<str>),
    GetField(Box<Node>, Box<Node>),
    ForLoop(ForLoop),
    If(Box<Node>, Box<Node>, Option<Box<Node>>),
    LetDeclList(LetDeclList),
    Identifier(Identifier),
    New(New),
    Object(Box<[PropertyDefinition]>),
    Return(Option<Box<Node>>),
    Switch(Box<Node>, Box<[(Node, Box<[Node]>)]>, Option<Box<Node>>),
    Spread(Box<Node>),
    Throw(Box<Node>),
    Try(Try),
    This,
    UnaryOp(UnaryOp),
    VarDeclList(VarDeclList),
    WhileLoop(Box<Node>, Box<Node>),
}

Variants

ArrayDecl(ArrayDecl)

Array declaration node. More information.

ArrowFunctionDecl(ArrowFunctionDecl)

An arrow function expression node. More information.

Assign(Assign)

An assignment operator node. More information.

BinOp(BinOp)

A binary operator node. More information.

Block(Block)

A Block node. More information.

Break(Option<Box<str>>)

The break statement terminates the current loop, switch, or label statement and transfers program control to the statement following the terminated statement.

The break statement includes an optional label that allows the program to break out of a labeled statement. The break statement needs to be nested within the referenced label. The labeled statement can be any block statement; it does not have to be preceded by a loop statement.

More information:

ECMAScript reference

MDN documentation

Call(Call)

A function call. More information.

ConditionalOp(Box<Node>, Box<Node>, Box<Node>)

The conditional (ternary) operator is the only JavaScript operator that takes three operands.

This operator is the only JavaScript operator that takes three operands: a condition followed by a question mark (?), then an expression to execute if the condition is truthy followed by a colon (:), and finally the expression to execute if the condition is false. This operator is frequently used as a shortcut for the if statement.

More information:

ECMAScript reference

MDN documentation

Const(Const)

Literals represent values in JavaScript.

These are fixed values not variables that you literally provide in your script.

More information:

ECMAScript reference

MDN documentation

ConstDeclList(ConstDeclList)

A constant declaration list. More information.

Continue(Option<Box<str>>)

The continue statement terminates execution of the statements in the current iteration of the current or labeled loop, and continues execution of the loop with the next iteration.

The continue statement can include an optional label that allows the program to jump to the next iteration of a labeled loop statement instead of the current loop. In this case, the continue statement needs to be nested within this labeled statement.

More information:

ECMAScript reference

MDN documentation

DoWhileLoop(Box<Node>, Box<Node>)

The do...while statement creates a loop that executes a specified statement until the test condition evaluates to false.

The condition is evaluated after executing the statement, resulting in the specified statement executing at least once.

More information:

ECMAScript reference

MDN documentation

FunctionDecl(FunctionDecl)

A function declaration node. More information.

FunctionExpr(FunctionExpr)

A function expressino node. More information

GetConstField(Box<Node>, Box<str>)

This property accessor provides access to an object's properties by using the dot notation.

In the object.property syntax, the property must be a valid JavaScript identifier. (In the ECMAScript standard, the names of properties are technically "IdentifierNames", not "Identifiers", so reserved words can be used but are not recommended).

One can think of an object as an associative array (a.k.a. map, dictionary, hash, lookup table). The keys in this array are the names of the object's properties.

It's typical when speaking of an object's properties to make a distinction between properties and methods. However, the property/method distinction is little more than a convention. A method is simply a property that can be called (for example, if it has a reference to a Function instance as its value).

More information:

ECMAScript reference

MDN documentation

GetField(Box<Node>, Box<Node>)

This property accessor provides access to an object's properties by using the bracket notation.

In the object[property_name] syntax, the property_name is just a string or Symbol. So, it can be any string, including '1foo', '!bar!', or even ' ' (a space).

One can think of an object as an associative array (a.k.a. map, dictionary, hash, lookup table). The keys in this array are the names of the object's properties.

It's typical when speaking of an object's properties to make a distinction between properties and methods. However, the property/method distinction is little more than a convention. A method is simply a property that can be called (for example, if it has a reference to a Function instance as its value).

More information:

ECMAScript reference

MDN documentation

ForLoop(ForLoop)

A for statement. More information.

If(Box<Node>, Box<Node>, Option<Box<Node>>)

The if statement executes a statement if a specified condition is truthy. If the condition is falsy, another statement can be executed.

Multiple if...else statements can be nested to create an else if clause.

Note that there is no elseif (in one word) keyword in JavaScript.

More information:

ECMAScript reference

MDN documentation

LetDeclList(LetDeclList)

A let declaration list. More information.

Identifier(Identifier)

A local identifier node. More information.

New(New)

A new expression. More information.

Object(Box<[PropertyDefinition]>)

Objects in JavaScript may be defined as an unordered collection of related data, of primitive or reference types, in the form of “key: value” pairs.

Objects can be initialized using new Object(), Object.create(), or using the literal notation.

An object initializer is an expression that describes the initialization of an Object. Objects consist of properties, which are used to describe an object. Values of object properties can either contain primitive data types or other objects.

More information:

ECMAScript reference

MDN documentation

Return(Option<Box<Node>>)

The return statement ends function execution and specifies a value to be returned to the function caller.

Syntax: return [expression];

expression:

The expression whose value is to be returned. If omitted, undefined is returned nstead.

When a return statement is used in a function body, the execution of the function is stopped. If specified, a given value is returned to the function caller.

More information:

ECMAScript reference

MDN documentation

Switch(Box<Node>, Box<[(Node, Box<[Node]>)]>, Option<Box<Node>>)

The switch statement evaluates an expression, matching the expression's value to a case clause, and executes statements associated with that case, as well as statements in cases that follow the matching case.

A switch statement first evaluates its expression. It then looks for the first case clause whose expression evaluates to the same value as the result of the input expression (using the strict comparison, ===) and transfers control to that clause, executing the associated statements. (If multiple cases match the provided value, the first case that matches is selected, even if the cases are not equal to each other.)

More information:

ECMAScript reference

MDN documentation

Spread(Box<Node>)

The spread operator allows an iterable such as an array expression or string to be expanded.

Syntax: ...x

It expands array expressions or strings in places where zero or more arguments (for function calls) or elements (for array literals) are expected, or an object expression to be expanded in places where zero or more key-value pairs (for object literals) are expected.

More information:

ECMAScript reference

MDN documentation

Throw(Box<Node>)

The throw statement throws a user-defined exception.

Syntax: throw expression;

Execution of the current function will stop (the statements after throw won't be executed), and control will be passed to the first catch block in the call stack. If no catch block exists among caller functions, the program will terminate.

More information:

ECMAScript reference

MDN documentation

Try(Try)

A try...catch node. More information.

This

The JavaScript this keyword refers to the object it belongs to.

A property of an execution context (global, function or eval) that, in non–strict mode, is always a reference to an object and in strict mode can be any value.

More information:

ECMAScript reference

MDN documentation

UnaryOp(UnaryOp)

Unary operation node. More information

VarDeclList(VarDeclList)

Array declaration node. More information.

WhileLoop(Box<Node>, Box<Node>)

The while statement creates a loop that executes a specified statement as long as the test condition evaluates to true.

The condition is evaluated before executing the statement.

More information:

ECMAScript reference

MDN documentation

Methods

impl Node

pub(crate) fn hoistable_order(a: &Node, b: &Node) -> Ordering

Returns a node ordering based on the hoistability of each node.

pub fn break_node<OL, L>(label: OL) -> Self where
L: Into<Box<str>>,
OL: Into<Option<L>>,

Creates a Break AST node.

pub fn conditional_op<C, T, F>(condition: C, if_true: T, if_false: F) -> Self where
C: Into<Self>,
T: Into<Self>,
F: Into<Self>,

Creates a ConditionalOp AST node.

pub fn continue_node<OL, L>(label: OL) -> Self where
L: Into<Box<str>>,
OL: Into<Option<L>>,

Creates a Continue AST node.

pub fn do_while_loop<B, C>(body: B, condition: C) -> Self where
B: Into<Self>,
C: Into<Self>,

Creates a DoWhileLoop AST node.

pub fn get_const_field<V, L>(value: V, label: L) -> Self where
V: Into<Self>,
L: Into<Box<str>>,

Creates a GetConstField AST node.

pub fn get_field<V, F>(value: V, field: F) -> Self where
V: Into<Self>,
F: Into<Self>,

Creates a GetField AST node.

pub fn if_node<C, B, E, OE>(condition: C, body: B, else_node: OE) -> Self where
C: Into<Self>,
B: Into<Self>,
E: Into<Self>,
OE: Into<Option<E>>,

Creates an If AST node.

pub fn object<D>(def: D) -> Self where
D: Into<Box<[PropertyDefinition]>>,

Creates an Object AST node.

pub fn return_node<E, OE>(expr: OE) -> Self where
E: Into<Self>,
OE: Into<Option<E>>,

Creates a Return AST node.

pub fn switch<V, C, OD, D>(val: V, cases: C, default: OD) -> Self where
V: Into<Self>,
C: Into<Box<[(Self, Box<[Self]>)]>>,
OD: Into<Option<D>>,
D: Into<Self>,

Creates a Switch AST node.

pub fn spread<V>(val: V) -> Self where
V: Into<Self>,

Creates a Spread AST node.

pub fn throw<V>(val: V) -> Self where
V: Into<Self>,

Creates a Throw AST node.

pub fn this() -> Self

Creates a This AST node.

pub fn while_loop<C, B>(condition: C, body: B) -> Self where
C: Into<Self>,
B: Into<Self>,

Creates a WhileLoop AST node.

fn display(&self, f: &mut Formatter, indentation: usize) -> Result

Implements the display formatting with indentation.