Trait boa::syntax::ast::op::Operator

pub trait Operator {
    fn get_assoc(&self) -> bool;

    fn get_precedence(&self) -> u64;

    fn get_precedence_and_assoc(&self) -> (u64, bool) { ... }
}

Represents an operator

Required methods

fn get_assoc(&self) -> bool

Get the associativity as a boolean that is true if it goes rightwards

fn get_precedence(&self) -> u64

Get the precedence as an unsigned integer, where the lower it is, the more precedence/priority it has

Loading content...
Provided methods

fn get_precedence_and_assoc(&self) -> (u64, bool)

Get the precedence and associativity of this operator

Loading content...