Module boa::syntax::lexer

A lexical analyzer for JavaScript source code.

This module contains the Boa lexer or tokenizer implementation.

The Lexer splits its input source code into a sequence of input elements called tokens, represented by the Token structure. It also removes whitespace and comments and attaches them to the next token.

This is tightly coupled with the parser due to the javascript goal-symbol requirements as documented by the spec.

More information:

ECMAScript reference

Re-exports

pub use crate::profiler::BoaProfiler;

pub use crate::syntax::ast::Position;

pub use error::Error;

pub use token::Token;

pub use token::TokenKind;

Modules

comment
This module implements lexing for comments used in the JavaScript programing language.

cursor
Module implementing the lexer cursor. This is used for managing the input byte stream.

error
This module contains the errors used by the lexer.

identifier
This module implements lexing for identifiers (foo, myvar, etc.) used in the JavaScript programing language.

number
This module implements lexing for number literals (123, 787) used in the JavaScript programing language.

operator
This module implements lexing for operators (+, - etc.) used in the JavaScript programing language.

regex
This module implements lexing for regex literals used in the JavaScript programing language.

spread
This module implements lexing for spread (…) literals used in the JavaScript programing language.

string
This module implements lexing for string literals used in the JavaScript programing language.

template
This module implements lexing for template literals used in the JavaScript programing language.

token
This module implements all of the Tokens used in the JavaScript programing language.

Structs

Lexer
Lexer or tokenizer for the Boa JavaScript Engine.

Enums

InputElement
ECMAScript goal symbols.

Traits

Tokenizer