Struct boa::syntax::lexer::cursor::Cursor

pub(super) struct Cursor<R> {
    iter: InnerIter<R>,
    pos: Position,
    strict_mode: bool,
}

Cursor over the source code.

Fields
iter: InnerIter<R>pos: Positionstrict_mode: bool
Implementations

impl<R> Cursor<R>

pub(super) fn pos(
&self
) -> Position

Gets the current position of the cursor in the source code.

pub(super) fn next_column(
&mut self
)

Advances the position to the next column.

fn next_line(&mut self)

Advances the position to the next line.

pub(super) fn strict_mode(
&self
) -> bool

pub(super) fn set_strict_mode(
&mut self,
strict_mode: bool
)

impl<R> Cursor<R> where
R: Read,

pub(super) fn new(
inner: R
) -> Self

Creates a new Lexer cursor.

pub(super) fn peek(
&mut self
) -> Result<Option<char>, Error>

Peeks the next character.

pub(super) fn next_is(
&mut self,
peek: char
) -> Result<bool>

Compares the character passed in to the next character, if they match true is returned and the buffer is incremented

pub(super) fn next_is_pred<F>(
&mut self,
pred: &F
) -> Result<bool> where
F: Fn(char) -> bool,

Applies the predicate to the next character and returns the result. Returns false if there is no next character.

The buffer is not incremented.

pub(super) fn take_until(
&mut self,
stop: char,
buf: &mut String
) -> Result<()>

Fills the buffer with all characters until the stop character is found.

Note: It will not add the stop character to the buffer.

pub(super) fn take_while_pred<F>(
&mut self,
buf: &mut String,
pred: &F
) -> Result<()> where
F: Fn(char) -> bool,

Fills the buffer with characters until the first character (x) for which the predicate (pred) is false (or the next character is none).

Note that all characters up until x are added to the buffer including the character right before.

pub(super) fn fill_bytes(
&mut self,
buf: &mut [u8]
) -> Result<()>

It will fill the buffer with checked ASCII bytes.

This expects for the buffer to be fully filled. If it's not, it will fail with an UnexpectedEof I/O error.

pub(crate) fn next_char(&mut self) -> Result<Option<char>, Error>

Retrieves the next UTF-8 character.