Struct boa::syntax::lexer::Lexer

pub struct Lexer<'a> {
    pub tokens: Vec<Token>,
    line_number: u64,
    column_number: u64,
    buffer: Peekable<Chars<'a>>,
}

A lexical analyzer for JavaScript source code

Fields
tokens: Vec<Token>line_number: u64column_number: u64buffer: Peekable<Chars<'a>>
Methods

impl<'a> Lexer<'a>

pub fn new(buffer: &'a str) -> Lexer<'a>

Returns a Lexer with a buffer inside

Arguments

buffer - A string slice that holds the source code. The buffer needs to have a lifetime as long as the Lexer instance itself

Example

let buffer = std::fs::read_to_string("yourSourceCode.js").unwrap();
let lexer = boa::syntax::lexer::Lexer::new(&buffer);

fn push_token(&mut self, tk: TokenKind)

Push tokens onto the token queue

fn push_punc(&mut self, punc: Punctuator)

Push a punctuation token

fn next(&mut self) -> char

next fetches the next token and return it, or a LexerError if there are no more.

fn preview_next(&mut self) -> Option<char>

Preview the next character but don't actually increment

fn preview_multiple_next(&mut self, nb_next: usize) -> Option<char>

Preview a char x indexes further in buf, without incrementing

fn take_char_while<F>(&mut self, f: F) -> Result<String, LexerError> where
F: FnMut(char) -> bool,

Utility Function, while f(char) is true, read chars and move curser. All chars are returned as a string

fn next_is(&mut self, peek: char) -> bool

Compares the character passed in to the next character, if they match true is returned and the buffer is incremented

fn read_integer_in_base(
&mut self,
base: u32,
buf: String
) -> Result<u64, LexerError>

fn check_after_numeric_literal(&mut self) -> Result<(), LexerError>

pub fn lex(&mut self) -> Result<(), LexerError>