Enum boa::syntax::lexer::token::TokenKind

pub enum TokenKind {

    BooleanLiteral(bool),
    EOF,
    Identifier(Box<str>),
    Keyword(Keyword),
    NullLiteral,
    NumericLiteral(Numeric),
    Punctuator(Punctuator),
    StringLiteral(Box<str>),
    TemplateNoSubstitution(TemplateString),
    TemplateMiddle(TemplateString),
    RegularExpressionLiteral(Box<str>, RegExpFlags),
    LineTerminator,
    Comment,

}

Represents the type of Token and the data it has inside.

Variants

BooleanLiteral(bool)

A boolean literal, which is either true or false.

EOF

The end of the file.

Identifier(Box<str>)

An identifier.

Keyword(Keyword)

A keyword.

see: Keyword

NullLiteral

A null literal.

NumericLiteral(Numeric)

A numeric literal.

Punctuator(Punctuator)

A piece of punctuation

see: Punctuator

StringLiteral(Box<str>)

A string literal.

TemplateNoSubstitution(TemplateString)

A part of a template literal without substitution.

TemplateMiddle(TemplateString)

The part of a template literal between substitutions

RegularExpressionLiteral(Box<str>, RegExpFlags)

A regular expression, consisting of body and flags.

LineTerminator

Indicates the end of a line (\n).

Comment

Indicates a comment, the content isn’t stored.

Implementations

impl TokenKind

pub fn boolean_literal(lit: bool) -> Self

Creates a BooleanLiteral token kind.

pub fn eof() -> Self

Creates an EOF token kind.

pub fn identifier<I>(ident: I) -> Self where
I: Into<Box<str>>,

Creates an Identifier token type.

pub fn keyword(keyword: Keyword) -> Self

Creates a Keyword token kind.

pub fn numeric_literal<L>(lit: L) -> Self where
L: Into<Numeric>,

Creates a NumericLiteral token kind.

pub fn punctuator(punc: Punctuator) -> Self

Creates a Punctuator token type.

pub fn string_literal<S>(lit: S) -> Self where
S: Into<Box<str>>,

Creates a StringLiteral token type.

pub fn template_middle(template_string: TemplateString) -> Self

pub fn template_no_substitution(template_string: TemplateString) -> Self

pub fn regular_expression_literal<B, R>(body: B, flags: R) -> Self where
B: Into<Box<str>>,
R: Into<RegExpFlags>,

Creates a RegularExpressionLiteral token kind.

pub fn line_terminator() -> Self

Creates a LineTerminator token kind.

pub fn comment() -> Self

Creates a ‘Comment’ token kind.