Module boa::syntax::parser

Boa parser implementation.

Re-exports

pub use self::error::ParseError;

pub use self::error::ParseResult;

Modules

cursor
Cursor implementation for the parser.

error
Error and result implementation for the parser.

expression
Expression parsing.

function
Function definition parsing.

statement
Statement and declaration parsing.

Structs

AllowAwait
Boolean representing if the parser should allow a await keyword.

AllowDefault
Boolean representing if the parser should allow a default keyword.

AllowIn
Boolean representing if the parser should allow a in keyword.

AllowReturn
Boolean representing if the parser should allow a return keyword.

AllowYield
Boolean representing if the parser should allow a yield keyword.

Parser

Script
Parses a full script.

ScriptBody
Parses a script body.

Traits

TokenParser
Trait implemented by parsers.