Module boa::syntax::parser::cursor::buffered_lexer

Structs

BufferedLexer

Constants

MAX_PEEK_SKIP
The maximum number of tokens which can be peeked ahead.

PEEK_BUF_SIZE
The fixed size of the buffer used for storing values that are peeked ahead.