Struct boa::syntax::parser::cursor::Cursor

pub(super) struct Cursor<R> {
    buffered_lexer: BufferedLexer<R>,
}

Token cursor.

This internal structure gives basic testable operations to the parser.

Fields
buffered_lexer: BufferedLexer<R>
Implementations

impl<R> Cursor<R> where
R: Read,

pub(super) fn new(reader: R) -> Self

Creates a new cursor with the given reader.

pub(super) fn set_goal(&mut self, elm: InputElement)

pub(super) fn lex_regex(&mut self, start: Position) -> Result<Token, ParseError>

pub(super) fn lex_template(
&mut self,
start: Position
) -> Result<Token, ParseError>

pub(super) fn next(&mut self) -> Result<Option<Token>, ParseError>

pub(super) fn peek(
&mut self,
skip_n: usize
) -> Result<Option<&Token>, ParseError>

pub(super) fn strict_mode(&self) -> bool

pub(super) fn set_strict_mode(&mut self, strict_mode: bool)

pub(super) fn expect<K>(
&mut self,
kind: K,
context: &'static str
) -> Result<Token, ParseError> where
K: Into<TokenKind>,

Returns an error if the next token is not of kind kind.

pub(super) fn peek_semicolon(
&mut self
) -> Result<SemicolonResult<'_>, ParseError>

It will peek for the next token, to see if it’s a semicolon.

It will automatically insert a semicolon if needed, as specified in the spec.

pub(super) fn expect_semicolon(
&mut self,
context: &'static str
) -> Result<(), ParseError>

Consumes the next token if it is a semicolon, or returns a ParseError if it’s not.

It will automatically insert a semicolon if needed, as specified in the spec.

pub(super) fn peek_expect_no_lineterminator(
&mut self,
skip_n: usize,
context: &'static str
) -> Result<&Token, ParseError>

It will make sure that the peeked token (skipping n tokens) is not a line terminator.

It expects that the token stream does not end here.

This is just syntatic sugar for a .peek(skip_n) call followed by a check that the result is not a line terminator or None.

pub(super) fn next_if<K>(
&mut self,
kind: K
) -> Result<Option<Token>, ParseError> where
K: Into<TokenKind>,

Advance the cursor to the next token and retrieve it, only if it’s of kind type.

When the next token is a kind token, get the token, otherwise return None.

No next token also returns None.