Enum boa::syntax::parser::error::ParseError

pub enum ParseError {
    Expected {
        expected: Box<[TokenKind]>,
        found: Token,
        context: &'static str,
    },
    Unexpected {
        found: Token,
        message: Option<&'static str>,
    },
    AbruptEnd,
    Lex {
        err: LexError,
    },
    General {
        message: &'static str,
        position: Position,
    },
    Unimplemented {
        message: &'static str,
        position: Position,
    },
}

ParseError is an enum which represents errors encounted during parsing an expression

Variants

Expected

When it expected a certain kind of token, but got another as part of something

Fields of Expected

expected: Box<[TokenKind]>found: Tokencontext: &'static str

Unexpected

When a token is unexpected

Fields of Unexpected

found: Tokenmessage: Option<&'static str>

AbruptEnd

When there is an abrupt end to the parsing

Lex

A lexing error.

Fields of Lex

err: LexError

General

Catch all General Error

Fields of General

message: &'static strposition: Position

Unimplemented

Unimplemented syntax error

Fields of Unimplemented

message: &'static strposition: Position

Implementations

impl ParseError

fn context(self, new_context: &'static str) -> Self

Changes the context of the error, if any.

pub(super) fn expected<E>(
expected: E,
found: Token,
context: &'static str
) -> Self where
E: Into<Box<[TokenKind]>>,

Creates an Expected parsing error.

pub(super) fn unexpected<C>(found: Token, message: C) -> Self where
C: Into<Option<&'static str>>,

Creates an Expected parsing error.

pub(super) fn general(message: &'static str, position: Position) -> Self

Creates a “general” parsing error.

pub(super) fn lex(e: LexError) -> Self

Creates a parsing error from a lexing error.

pub(super) fn unimplemented(message: &'static str, position: Position) -> Self

Creates a new Unimplemented parsing error.