Struct boa::syntax::parser::Parser

pub struct Parser<'a> {
    cursor: Cursor<'a>,
}

Fields
cursor: Cursor<'a>

Cursor in the parser, the internal structure used to read tokens.

Methods

impl<'a> Parser<'a>

pub fn new(tokens: &'a [Token]) -> Self

Create a new parser, using tokens as input

pub fn parse_all(&mut self) -> ParseResult

Parse all expressions in the token array

fn get_next_token(&mut self) -> Result<&Token, ParseError>

consume the next token and increment position

pub fn peek_skip_lineterminator(&mut self) -> Option<&'a Token>

Peek the next token skipping line terminators.

pub fn next_skip_lineterminator(&mut self) -> Option<&'a Token>

Consume the next token skipping line terminators.

fn next_if(&mut self, kind: TokenKind) -> Option<&'a Token>

Advance the cursor to the next token and retrieve it, only if it's of kind type.

When the next token is a kind token, get the token, otherwise return None.

fn next_if_skip_lineterminator(&mut self, kind: TokenKind) -> Option<&'a Token>

Advance the cursor to the next token and retrieve it, only if it's of kind type.

When the next token is a kind token, get the token, otherwise return None. This function skips line terminators.

fn expect(
&mut self,
kind: TokenKind,
routine: Option<&'static str>
) -> Result<(), ParseError>

Returns an error if the next Punctuator is not tk

fn expect_punc(
&mut self,
p: Punctuator,
routine: Option<&'static str>
) -> Result<(), ParseError>

Returns an error if the next symbol is not the punctuator p Consumes the next symbol otherwise

fn read_statement_list(&mut self) -> ParseResult

Reads a list of statements as a Node::StatementList.

It will end at the end of file.

fn read_block_statement(&mut self) -> ParseResult

Reads a code block as a Node::Block.

Note: it will start after the { character and stop after reading }.

fn read_block(&mut self) -> ParseResult

Read a list of statements and stop after }

Note: it will start after the { character and stop after reading }.

fn read_statements(
&mut self,
break_when_closingbrase: bool
) -> Result<Vec<Node>, ParseError>

Reads a list of statements.

If break_when_closingbrase is true, it will stop as soon as it finds a } character.

fn read_statement_list_item(&mut self) -> ParseResult

Reads an individual statement list item.

A statement list item can either be an statement or a declaration.

More information:

ECMAScript reference: https://tc39.es/ecma262/#prod-StatementListItem.

MDN information page about statements and declarations: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Statements.

fn read_declaration(&mut self) -> ParseResult

Parses a declaration.

More information:: https://tc39.es/ecma262/#prod-Declaration

fn read_binding_list(&mut self, is_const: bool) -> ParseResult

Reads a binding list.

It will return an error if a const declaration is being parsed and there is no initializer.

More information: https://tc39.es/ecma262/#prod-BindingList.

fn read_function_declaration(&mut self) -> ParseResult

Parses a function declaration.

More information:

ECMAScript specification: https://tc39.es/ecma262/#prod-FunctionDeclaration.

MDN documentation: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Statements/function

fn read_statement(&mut self) -> ParseResult

https://tc39.es/ecma262/#prod-Statement

fn read_switch_statement(&mut self) -> ParseResult

https://tc39.es/ecma262/#prod-SwitchStatement

fn read_expression_statement(&mut self) -> ParseResult

https://tc39.es/ecma262/#sec-expression-statement

fn read_break_statement(&mut self) -> ParseResult

https://tc39.es/ecma262/#sec-break-statement

fn read_continue_statement(&mut self) -> ParseResult

https://tc39.es/ecma262/#sec-continue-statement

fn read_throw_statement(&mut self) -> ParseResult

https://tc39.es/ecma262/#prod-ThrowStatement

fn read_return_statement(&mut self) -> ParseResult

https://tc39.es/ecma262/#prod-ReturnStatement

fn read_if_statement(&mut self) -> ParseResult

https://tc39.es/ecma262/#sec-if-statement

fn read_while_statement(&mut self) -> ParseResult

https://tc39.es/ecma262/#sec-while-statement

fn read_do_while_statement(&mut self) -> ParseResult

https://tc39.es/ecma262/#sec-do-while-statement

fn read_try_statement(&mut self) -> ParseResult

https://tc39.es/ecma262/#sec-try-statement

fn read_for_statement(&mut self) -> ParseResult

https://tc39.es/ecma262/#sec-for-statement

fn read_variable_statement(&mut self) -> ParseResult

https://tc39.es/ecma262/#prod-VariableStatement

fn read_variable_declaration_list(&mut self) -> ParseResult

https://tc39.es/ecma262/#prod-VariableDeclarationList

fn lexical_declaration_continuation(&mut self) -> Result<bool, ParseError>

Checks if the lexical declaration continues with more bindings.

If it does, it will advance the internal cursor to the next identifier token. A Lexical Declaration continues its binding list if we find a , character. A New line indicates the same as a ;.

More information: https://tc39.es/ecma262/#prod-LexicalDeclaration.

fn read_variable_declaration(
&mut self
) -> Result<(String, Option<Node>), ParseError>

https://tc39.es/ecma262/#prod-VariableDeclaration

fn read_initializer(&mut self) -> ParseResult

Reads an initializer of variables.

Note: it will expect the = token to have been already read.

More information: https://tc39.es/ecma262/#prod-Initializer

fn read_expression(&mut self) -> ParseResult

fn read_assignment_expression(&mut self) -> ParseResult

https://tc39.es/ecma262/#prod-AssignmentExpression

fn read_conditional_expression(&mut self) -> ParseResult

https://tc39.es/ecma262/#prod-ConditionalExpression

fn read_arrow_function(&mut self) -> ParseResult

https://tc39.es/ecma262/#sec-arrow-function-definitions

fn read_formal_parameters(&mut self) -> Result<FormalParameters, ParseError>

Collect parameters from functions or arrow functions

fn read_function_rest_parameter(
&mut self
) -> Result<FormalParameter, ParseError>

https://tc39.es/ecma262/#prod-FunctionRestParameter

fn read_formal_parameter(&mut self) -> Result<FormalParameter, ParseError>

https://tc39.es/ecma262/#prod-FormalParameter

fn read_logical_or_expression(&mut self) -> ParseResult

fn read_logical_and_expression(&mut self) -> ParseResult

fn read_bitwise_or_expression(&mut self) -> ParseResult

fn read_bitwise_xor_expression(&mut self) -> ParseResult

fn read_bitwise_and_expression(&mut self) -> ParseResult

fn read_equality_expression(&mut self) -> ParseResult

fn read_relational_expression(&mut self) -> ParseResult

fn read_shift_expression(&mut self) -> ParseResult

fn read_additive_expression(&mut self) -> ParseResult

fn read_multiplicate_expression(&mut self) -> ParseResult

fn read_exponentiation_expression(&mut self) -> ParseResult

https://tc39.es/ecma262/#prod-MultiplicativeExpression

fn is_unary_expression(&mut self) -> bool

fn read_unary_expression(&mut self) -> ParseResult

https://tc39.es/ecma262/#prod-UnaryExpression

fn read_update_expression(&mut self) -> ParseResult

https://tc39.es/ecma262/#prod-UpdateExpression

fn read_left_hand_side_expression(&mut self) -> ParseResult

https://tc39.es/ecma262/#prod-LeftHandSideExpression

fn read_new_expression(&mut self, first_member_expr: Node) -> ParseResult

https://tc39.es/ecma262/#prod-NewExpression

fn read_member_expression(&mut self) -> ParseResult

https://tc39.es/ecma262/#prod-MemberExpression

fn read_call_expression(&mut self, first_member_expr: Node) -> ParseResult

https://tc39.es/ecma262/#prod-CallExpression

fn read_arguments(&mut self) -> Result<Vec<Node>, ParseError>

https://tc39.es/ecma262/#prod-Arguments

fn read_primary_expression(&mut self) -> ParseResult

https://tc39.es/ecma262/#prod-PrimaryExpression

fn read_function_expression(&mut self) -> ParseResult

https://tc39.es/ecma262/#prod-FunctionExpression

fn read_array_literal(&mut self) -> ParseResult

https://tc39.es/ecma262/#prod-ArrayLiteral

fn read_object_literal(&mut self) -> ParseResult

https://tc39.es/ecma262/#prod-ObjectLiteral

fn get_object_property_name(&mut self) -> Result<String, ParseError>

fn read_property_definition(&mut self) -> Result<PropertyDefinition, ParseError>

https://tc39.es/ecma262/#prod-PropertyDefinition

fn read_property_method_definition(
&mut self,
identifier: &str
) -> Result<PropertyDefinition, ParseError>

https://tc39.es/ecma262/#prod-MethodDefinition