Trait boa::syntax::parser::TokenParser

trait TokenParser: Sized {
    type Output;
    fn parse(self, cursor: &mut Cursor) -> Result<Self::Output, ParseError>;

    fn try_parse(self, cursor: &mut Cursor) -> Option<Self::Output> { ... }
}

Trait implemented by parsers.

This makes it possible to abstract over the underlying implementation of a parser.

Associated Types

type Output

Output type for the parser.

Loading content...
Required methods

fn parse(self, cursor: &mut Cursor) -> Result<Self::Output, ParseError>

Parses the token stream using the current parser.

This method needs to be provided by the implementor type.

Loading content...
Provided methods

fn try_parse(self, cursor: &mut Cursor) -> Option<Self::Output>

Tries to parse the following tokens with this parser.

It will return the cursor to the initial position if an error occurs during parsing.

Loading content...