Trait boa::vm::code_block::Readable

pub(crate) unsafe trait Readable { }

This represents whether a value can be read from CodeBlock code.

Safety

This trait is safe to implement as long as the type doesn’t implement Drop. At some point, if negative impls are stabilized, we might be able to remove the unsafe bound.