Struct boa::vm::CodeBlock

pub struct CodeBlock {
    pub(crate) name: JsString,
    pub(crate) length: u32,
    pub(crate) strict: bool,
    pub(crate) constructor: bool,
    pub(crate) this_mode: ThisMode,
    pub(crate) params: Box<[FormalParameter]>,
    pub(crate) code: Vec<u8>,
    pub(crate) literals: Vec<JsValue>,
    pub(crate) variables: Vec<JsString>,
    pub(crate) functions: Vec<Gc<CodeBlock>>,
    pub(crate) lexical_name_argument: bool,
}

The internal representation of a JavaScript function.

A CodeBlock is generated for each function compiled by the ByteCompiler. It stores the bytecode and the other attributes of the function.

Fields
name: JsString

Name of this function

length: u32

The number of arguments expected.

strict: bool

Is this function in strict mode.

constructor: bool

Is this function a constructor.

this_mode: ThisMode

[ThisMode]

params: Box<[FormalParameter]>

Parameters passed to this function.

code: Vec<u8>

Bytecode

literals: Vec<JsValue>

Literals

variables: Vec<JsString>

Variables names

functions: Vec<Gc<CodeBlock>>

Functions inside this function

lexical_name_argument: bool

Indicates if the codeblock contains a lexical name arguments

Implementations

impl CodeBlock

pub fn new(name: JsString, length: u32, strict: bool, constructor: bool) -> Self

Constructs a new CodeBlock.

pub(crate) unsafe fn read_unchecked<T>(&self, offset: usize) -> T where
T: Readable,

Read type T from code.

Safety

Does not check if read happens out-of-bounds.

pub(crate) fn read<T>(&self, offset: usize) -> T where
T: Readable,

Read type T from code.

pub(crate) fn instruction_operands(&self, pc: &mut usize) -> String

Get the operands after the Opcode pointed to by pc as a String. Modifies the pc to point to the next instruction.

Returns an empty String if no operands are present.