Module boa::builtins

Builtins live here, such as Object, String, Math, etc.

Modules

array
This module implements the global Array object.

bigint
This module implements the global BigInt object.

boolean
This module implements the global Boolean object.

console
This module implements the global console object.

date

error
This module implements the global Error object.

function
This module implements the global Function object as well as creates Native Functions.

global_this
This module implements the global globalThis property.

infinity
This module implements the global Infinity property.

iterable

json
This module implements the global JSON object.

map
This module implements the global Map objest.

math
This module implements the global Math object.

nan
This module implements the global NaN property.

number
This module implements the global Number object.

object
This module implements the global Object object.

reflect
This module implements the global Reflect object.

regexp
This module implements the global RegExp object.

set
This module implements the global Set objest.

string
This module implements the global String object.

symbol
This module implements the global Symbol object.

undefined
This module implements the global undefined property.

Traits

BuiltIn

Functions

init
Initializes builtin objects and functions