Struct boa::builtins::console::Console

pub(crate) struct Console {
    count_map: FxHashMap<RcString, u32>,
    timer_map: FxHashMap<RcString, u128>,
    groups: Vec<String>,
}

This is the internal console object state.

Fields
count_map: FxHashMap<RcString, u32>timer_map: FxHashMap<RcString, u128>groups: Vec<String>
Implementations

impl Console

pub(crate) const NAME: &'static str

The name of the object.

pub(crate) fn assert(
_: &Value,
args: &[Value],
ctx: &mut Context
) -> Result<Value>

console.assert(condition, ...data)

Prints a JavaScript value to the standard error if first argument evaluates to false or there were no arguments.

More information:

MDN documentation

WHATWG console specification

pub(crate) fn clear(_: &Value, _: &[Value], ctx: &mut Context) -> Result<Value>

console.clear()

Removes all groups and clears console if possible.

More information:

MDN documentation

WHATWG console specification

pub(crate) fn debug(
_: &Value,
args: &[Value],
ctx: &mut Context
) -> Result<Value>

console.debug(...data)

Prints a JavaScript values with "debug" logLevel.

More information:

MDN documentation

WHATWG console specification

pub(crate) fn error(
_: &Value,
args: &[Value],
ctx: &mut Context
) -> Result<Value>

console.error(...data)

Prints a JavaScript values with "error" logLevel.

More information:

MDN documentation

WHATWG console specification

pub(crate) fn info(
_: &Value,
args: &[Value],
ctx: &mut Context
) -> Result<Value>

console.info(...data)

Prints a JavaScript values with "info" logLevel.

More information:

MDN documentation

WHATWG console specification

pub(crate) fn log(_: &Value, args: &[Value], ctx: &mut Context) -> Result<Value>

console.log(...data)

Prints a JavaScript values with "log" logLevel.

More information:

MDN documentation

WHATWG console specification

pub(crate) fn trace(
_: &Value,
args: &[Value],
ctx: &mut Context
) -> Result<Value>

console.trace(...data)

Prints a stack trace with "trace" logLevel, optionally labelled by data.

More information:

MDN documentation

WHATWG console specification

pub(crate) fn warn(
_: &Value,
args: &[Value],
ctx: &mut Context
) -> Result<Value>

console.warn(...data)

Prints a JavaScript values with "warn" logLevel.

More information:

MDN documentation

WHATWG console specification

pub(crate) fn count(
_: &Value,
args: &[Value],
ctx: &mut Context
) -> Result<Value>

console.count(label)

Prints number of times the function was called with that particular label.

More information:

MDN documentation

WHATWG console specification

pub(crate) fn count_reset(
_: &Value,
args: &[Value],
ctx: &mut Context
) -> Result<Value>

console.countReset(label)

Resets the counter for label.

More information:

MDN documentation

WHATWG console specification

fn system_time_in_ms() -> u128

Returns current system time in ms.

pub(crate) fn time(
_: &Value,
args: &[Value],
ctx: &mut Context
) -> Result<Value>

console.time(label)

Starts the timer for given label.

More information:

MDN documentation

WHATWG console specification

pub(crate) fn time_log(
_: &Value,
args: &[Value],
ctx: &mut Context
) -> Result<Value>

console.timeLog(label, ...data)

Prints elapsed time for timer with given label.

More information:

MDN documentation

WHATWG console specification

pub(crate) fn time_end(
_: &Value,
args: &[Value],
ctx: &mut Context
) -> Result<Value>

console.timeEnd(label)

Removes the timer with given label.

More information:

MDN documentation

WHATWG console specification

pub(crate) fn group(
_: &Value,
args: &[Value],
ctx: &mut Context
) -> Result<Value>

console.group(...data)

Adds new group with name from formatted data to stack.

More information:

MDN documentation

WHATWG console specification

pub(crate) fn group_end(
_: &Value,
_: &[Value],
ctx: &mut Context
) -> Result<Value>

console.groupEnd(label)

Removes the last group from the stack.

More information:

MDN documentation

WHATWG console specification

pub(crate) fn dir(_: &Value, args: &[Value], ctx: &mut Context) -> Result<Value>

console.dir(item, options)

Prints info about item

More information:

MDN documentation

WHATWG console specification