Struct boa::builtins::date::Date

pub struct Date(Option<NaiveDateTime>);

Implementations

impl Date

pub(crate) const LENGTH: usize

The amount of arguments this function object takes.

pub fn time_clip(time: f64) -> Option<f64>

Check if the time (number of miliseconds) is in the expected range. Returns None if the time is not in the range, otherwise returns the time itself in option.

More information:

ECMAScript reference

pub fn to_local(&self) -> Option<DateTime<Local>>

Converts the Date to a local DateTime.

If the Date is invalid (i.e. NAN), this function will return None.

pub fn to_utc(&self) -> Option<DateTime<Utc>>

Converts the Date to a UTC DateTime.

If the Date is invalid (i.e. NAN), this function will return None.

pub fn set_components(
&mut self,
utc: bool,
year: Option<f64>,
month: Option<f64>,
day: Option<f64>,
hour: Option<f64>,
minute: Option<f64>,
second: Option<f64>,
millisecond: Option<f64>
)

Optionally sets the individual components of the Date.

Each component does not have to be within the range of valid values. For example, if month is too large then year will be incremented by the required amount.

pub(crate) fn constructor(
new_target: &Value,
args: &[Value],
context: &mut Context
) -> Result<Value>

Date()

Creates a JavaScript Date instance that represents a single moment in time in a platform-independent format.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn make_date_string() -> Value

Date()

The Date() function is used to create a string that represent the current date and time.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn make_date_now(this: &Value) -> Value

Date()

The newly-created Date object represents the current date and time as of the time of instantiation.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn make_date_single(
this: &Value,
args: &[Value],
context: &mut Context
) -> Result<Value>

Date(value)

The newly-created Date object represents the value provided to the constructor.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn make_date_multiple(
this: &Value,
args: &[Value],
context: &mut Context
) -> Result<Value>

Date(year, month [ , date [ , hours [ , minutes [ , seconds [ , ms ] ] ] ] ])

The newly-created Date object represents the date components provided to the constructor.

More information:

ECMAScript reference

MDN documentation

pub fn get_date(&self) -> f64

Date.prototype.getDate()

The getDate() method returns the day of the month for the specified date according to local time.

More information:

ECMAScript reference

MDN documentation

pub fn get_day(&self) -> f64

Date.prototype.getDay()

The getDay() method returns the day of the week for the specified date according to local time, where 0 represents Sunday.

More information:

ECMAScript reference

MDN documentation

pub fn get_full_year(&self) -> f64

Date.prototype.getFullYear()

The getFullYear() method returns the year of the specified date according to local time.

More information:

ECMAScript reference

MDN documentation

pub fn get_hours(&self) -> f64

Date.prototype.getHours()

The getHours() method returns the hour for the specified date, according to local time.

More information:

ECMAScript reference

MDN documentation

pub fn get_milliseconds(&self) -> f64

Date.prototype.getMilliseconds()

The getMilliseconds() method returns the milliseconds in the specified date according to local time.

More information:

ECMAScript reference

MDN documentation

pub fn get_minutes(&self) -> f64

Date.prototype.getMinutes()

The getMinutes() method returns the minutes in the specified date according to local time.

More information:

ECMAScript reference

MDN documentation

pub fn get_month(&self) -> f64

Date.prototype.getMonth()

The getMonth() method returns the month in the specified date according to local time, as a zero-based value (where zero indicates the first month of the year).

More information:

ECMAScript reference

MDN documentation

pub fn get_seconds(&self) -> f64

Date.prototype.getSeconds()

The getSeconds() method returns the seconds in the specified date according to local time.

More information:

ECMAScript reference

MDN documentation

pub fn get_year(&self) -> f64

Date.prototype.getYear()

The getYear() method returns the year in the specified date according to local time. Because getYear() does not return full years (“year 2000 problem”), it is no longer used and has been replaced by the getFullYear() method.

More information:

ECMAScript reference

MDN documentation

pub fn get_time(&self) -> f64

Date.prototype.getTime()

The getTime() method returns the number of milliseconds since the Unix Epoch.

More information:

ECMAScript reference

MDN documentation

pub fn get_timezone_offset() -> f64

Date.prototype.getTimeZoneOffset()

The getTimezoneOffset() method returns the time zone difference, in minutes, from current locale (host system settings) to UTC.

More information:

ECMAScript reference

MDN documentation

pub fn get_utc_date(&self) -> f64

Date.prototype.getUTCDate()

The getUTCDate() method returns the day (date) of the month in the specified date according to universal time.

More information:

ECMAScript reference

MDN documentation

pub fn get_utc_day(&self) -> f64

Date.prototype.getUTCDay()

The getUTCDay() method returns the day of the week in the specified date according to universal time, where 0 represents Sunday.

More information:

ECMAScript reference

MDN documentation

pub fn get_utc_full_year(&self) -> f64

Date.prototype.getUTCFullYear()

The getUTCFullYear() method returns the year in the specified date according to universal time.

More information:

ECMAScript reference

MDN documentation

pub fn get_utc_hours(&self) -> f64

Date.prototype.getUTCHours()

The getUTCHours() method returns the hours in the specified date according to universal time.

More information:

ECMAScript reference

MDN documentation

pub fn get_utc_milliseconds(&self) -> f64

Date.prototype.getUTCMilliseconds()

The getUTCMilliseconds() method returns the milliseconds portion of the time object’s value.

More information:

ECMAScript reference

MDN documentation

pub fn get_utc_minutes(&self) -> f64

Date.prototype.getUTCMinutes()

The getUTCMinutes() method returns the minutes in the specified date according to universal time.

More information:

ECMAScript reference

MDN documentation

pub fn get_utc_month(&self) -> f64

Date.prototype.getUTCMonth()

The getUTCMonth() returns the month of the specified date according to universal time, as a zero-based value (where zero indicates the first month of the year).

More information:

ECMAScript reference

MDN documentation

pub fn get_utc_seconds(&self) -> f64

Date.prototype.getUTCSeconds()

The getUTCSeconds() method returns the seconds in the specified date according to universal time.

More information:

ECMAScript reference

MDN documentation

pub fn set_date(&mut self, day: Option<f64>)

Date.prototype.setDate()

The setDate() method sets the day of the Date object relative to the beginning of the currently set month.

More information:

ECMAScript reference

MDN documentation

pub fn set_full_year(
&mut self,
year: Option<f64>,
month: Option<f64>,
day: Option<f64>
)

Date.prototype.setFullYear()

The setFullYear() method sets the full year for a specified date according to local time. Returns new timestamp.

More information:

ECMAScript reference

MDN documentation

pub fn set_hours(
&mut self,
hour: Option<f64>,
minute: Option<f64>,
second: Option<f64>,
millisecond: Option<f64>
)

Date.prototype.setHours()

The setHours() method sets the hours for a specified date according to local time, and returns the number of milliseconds since January 1, 1970 00:00:00 UTC until the time represented by the updated Date instance.

More information:

ECMAScript reference

MDN documentation

pub fn set_milliseconds(&mut self, millisecond: Option<f64>)

Date.prototype.setMilliseconds()

The setMilliseconds() method sets the milliseconds for a specified date according to local time.

More information:

ECMAScript reference

MDN documentation

pub fn set_minutes(
&mut self,
minute: Option<f64>,
second: Option<f64>,
millisecond: Option<f64>
)

Date.prototype.setMinutes()

The setMinutes() method sets the minutes for a specified date according to local time.

More information:

ECMAScript reference

MDN documentation

pub fn set_month(&mut self, month: Option<f64>, day: Option<f64>)

Date.prototype.setMonth()

The setMonth() method sets the month for a specified date according to the currently set year.

More information:

ECMAScript reference

MDN documentation

pub fn set_seconds(&mut self, second: Option<f64>, millisecond: Option<f64>)

Date.prototype.setSeconds()

The setSeconds() method sets the seconds for a specified date according to local time.

More information:

ECMAScript reference

MDN documentation

pub fn set_year(
&mut self,
year: Option<f64>,
month: Option<f64>,
day: Option<f64>
)

Date.prototype.setYear()

The setYear() method sets the year for a specified date according to local time.

More information:

ECMAScript reference

MDN documentation

pub fn set_time(&mut self, time: Option<f64>)

Date.prototype.setTime()

The setTime() method sets the Date object to the time represented by a number of milliseconds since January 1, 1970, 00:00:00 UTC.

More information:

ECMAScript reference

MDN documentation

pub fn set_utc_date(&mut self, day: Option<f64>)

Date.prototype.setUTCDate()

The setUTCDate() method sets the day of the month for a specified date according to universal time.

More information:

ECMAScript reference

MDN documentation

pub fn set_utc_full_year(
&mut self,
year: Option<f64>,
month: Option<f64>,
day: Option<f64>
)

Date.prototype.setFullYear()

The setFullYear() method sets the full year for a specified date according to local time. Returns new timestamp.

More information:

ECMAScript reference

MDN documentation

pub fn set_utc_hours(
&mut self,
hour: Option<f64>,
minute: Option<f64>,
second: Option<f64>,
millisecond: Option<f64>
)

Date.prototype.setUTCHours()

The setUTCHours() method sets the hour for a specified date according to universal time, and returns the number of milliseconds since January 1, 1970 00:00:00 UTC until the time represented by the updated Date instance.

More information:

ECMAScript reference

MDN documentation

pub fn set_utc_milliseconds(&mut self, millisecond: Option<f64>)

Date.prototype.setUTCMilliseconds()

The setUTCMilliseconds() method sets the milliseconds for a specified date according to universal time.

More information:

ECMAScript reference

MDN documentation

pub fn set_utc_minutes(
&mut self,
minute: Option<f64>,
second: Option<f64>,
millisecond: Option<f64>
)

Date.prototype.setUTCMinutes()

The setUTCMinutes() method sets the minutes for a specified date according to universal time.

More information:

ECMAScript reference

MDN documentation

pub fn set_utc_month(&mut self, month: Option<f64>, day: Option<f64>)

Date.prototype.setUTCMonth()

The setUTCMonth() method sets the month for a specified date according to universal time.

More information:

ECMAScript reference

MDN documentation

pub fn set_utc_seconds(&mut self, second: Option<f64>, millisecond: Option<f64>)

Date.prototype.setUTCSeconds()

The setUTCSeconds() method sets the seconds for a specified date according to universal time.

More information:

ECMAScript reference

MDN documentation

pub fn to_date_string(&self) -> String

Date.prototype.toDateString()

The toDateString() method returns the date portion of a Date object in English.

More information:

ECMAScript reference

MDN documentation

pub fn to_gmt_string(&self) -> String

Date.prototype.toGMTString()

The toGMTString() method converts a date to a string, using Internet Greenwich Mean Time (GMT) conventions.

More information:

ECMAScript reference

MDN documentation

pub fn to_iso_string(&self) -> String

Date.prototype.toISOString()

The toISOString() method returns a string in simplified extended ISO format (ISO 8601).

More information:

ISO 8601

ECMAScript reference

MDN documentation

pub fn to_json(&self) -> String

Date.prototype.toJSON()

The toJSON() method returns a string representation of the Date object.

More information:

ECMAScript reference

MDN documentation

pub fn to_time_string(&self) -> String

Date.prototype.toTimeString()

The toTimeString() method returns the time portion of a Date object in human readable form in American English.

More information:

ECMAScript reference

MDN documentation

pub fn to_utc_string(&self) -> String

Date.prototype.toUTCString()

The toUTCString() method returns a string representing the specified Date object.

More information:

ECMAScript reference

MDN documentation

pub fn value_of(&self) -> f64

Date.prototype.valueOf()

The valueOf() method returns the primitive value of a Date object.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn now(_: &Value, _: &[Value], _: &mut Context) -> Result<Value>

Date.now()

The static Date.now() method returns the number of milliseconds elapsed since January 1, 1970 00:00:00 UTC.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn parse(
_: &Value,
args: &[Value],
context: &mut Context
) -> Result<Value>

Date.parse()

The Date.parse() method parses a string representation of a date, and returns the number of milliseconds since January 1, 1970, 00:00:00 UTC or NaN if the string is unrecognized or, in some cases, contains illegal date values.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn utc(
_: &Value,
args: &[Value],
context: &mut Context
) -> Result<Value>

Date.UTC()

The Date.UTC() method accepts parameters similar to the Date constructor, but treats them as UTC.

More information:

ECMAScript reference

MDN documentation