Module boa::builtins::function

This module implements the global Function object as well as creates Native Functions.

Objects wrap Functions and expose them via call/construct slots.

The Function` object is used for matching text with a pattern.

More information:

ECMAScript reference

MDN documentation

Structs

BuiltInFunction

BuiltInFunctionObject

FunctionFlags

Enums

Function
Boa representation of a Function Object.

Functions

create_unmapped_arguments_object
Arguments.

make_builtin_fn
Creates a new member function of a Object or prototype.

Type Definitions

NativeFunction
fn(this, arguments, context) -> ResultValue - The signature of a built-in function