Struct boa::builtins::map::ordered_map::OrderedMap

pub struct OrderedMap<K, V, S = RandomState>(IndexMap<K, V, S>)
 where
    K: Hash + Eq;

A newtype wrapping indexmap::IndexMap

Implementations

impl<K, V> OrderedMap<K, V> where
K: Hash + Eq,

pub fn new() -> Self

pub fn with_capacity(capacity: usize) -> Self

pub fn len(&self) -> usize

Return the number of key-value pairs in the map.

Computes in O(1) time.

pub fn is_empty(&self) -> bool

Returns true if the map contains no elements.

Computes in O(1) time.

pub fn insert(&mut self, key: K, value: V) -> Option<V>

Insert a key-value pair in the map.

If an equivalent key already exists in the map: the key remains and retains in its place in the order, its corresponding value is updated with value and the older value is returned inside Some(_).

If no equivalent key existed in the map: the new key-value pair is inserted, last in order, and None is returned.

Computes in O(1) time (amortized average).

pub fn remove(&mut self, key: &K) -> Option<V>

Remove the key-value pair equivalent to key and return its value.

Like Vec::remove, the pair is removed by shifting all of the elements that follow it, preserving their relative order. This perturbs the index of all of those elements!

Return None if key is not in map.

Computes in O(n) time (average).

pub fn get(&self, key: &K) -> Option<&V>

Return a reference to the value stored for key, if it is present, else None.

Computes in O(1) time (average).

pub fn get_index(&self, index: usize) -> Option<(&K, &V)>

Get a key-value pair by index Valid indices are 0 <= index < self.len() Computes in O(1) time.

pub fn iter(&self) -> Iter<'_, K, V>

Return an iterator over the key-value pairs of the map, in their order

pub fn contains_key(&self, key: &K) -> bool

Return true if an equivalent to key exists in the map.

Computes in O(1) time (average).