Module boa::builtins::number

This module implements the global Number object.

The Number JavaScript object is a wrapper object allowing you to work with numerical values. A Number object is created using the Number() constructor. A primitive type object number is created using the Number() function.

The JavaScript Number type is double-precision 64-bit binary format IEEE 754 value. In more recent implementations, JavaScript also supports integers with arbitrary precision using the BigInt type.

More information:

ECMAScript reference

MDN documentation

Modules

conversions

Structs

Number
Number implementation.

Constants

BUF_SIZE

PARSE_FLOAT_MAX_ARG_COUNT
Maximum number of arguments expected to the builtin parseFloat() function.

PARSE_INT_MAX_ARG_COUNT
Maximum number of arguments expected to the builtin parseInt() function.