Struct boa::builtins::number::Number

pub(crate) struct Number;

Number implementation.

Implementations

impl Number

pub(crate) const LENGTH: usize

The amount of arguments this function object takes.

pub(crate) const MAX_SAFE_INTEGER: f64

The Number.MAX_SAFE_INTEGER constant represents the maximum safe integer in JavaScript (2^53 - 1).

/// More information:

ECMAScript reference

MDN documentation

pub(crate) const MIN_SAFE_INTEGER: f64

The Number.MIN_SAFE_INTEGER constant represents the minimum safe integer in JavaScript (-(253 - 1)).

More information:

ECMAScript reference

MDN documentation

pub(crate) const MAX_VALUE: f64

The Number.MAX_VALUE property represents the maximum numeric value representable in JavaScript.

The MAX_VALUE property has a value of approximately 1.79E+308, or 2^1024. Values larger than MAX_VALUE are represented as Infinity.

More information:

ECMAScript reference

MDN documentation

pub(crate) const MIN_VALUE: f64

The Number.MIN_VALUE property represents the smallest positive numeric value representable in JavaScript.

The MIN_VALUE property is the number closest to 0, not the most negative number, that JavaScript can represent. It has a value of approximately 5e-324. Values smaller than MIN_VALUE (“underflow values”) are converted to 0.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn constructor(
new_target: &Value,
args: &[Value],
context: &mut Context
) -> Result<Value>

Number( value )

fn this_number_value(value: &Value, context: &mut Context) -> Result<f64>

This function returns a Result of the number Value.

If the Value is a Number primitive of Number object the number is returned. Otherwise an TypeError is thrown.

More information:

ECMAScript reference

fn num_to_exponential(n: f64) -> String

Helper function that formats a float as a ES6-style exponential number string.

pub(crate) fn to_exponential(
this: &Value,
_: &[Value],
context: &mut Context
) -> Result<Value>

Number.prototype.toExponential( [fractionDigits] )

The toExponential() method returns a string representing the Number object in exponential notation.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn to_fixed(
this: &Value,
args: &[Value],
context: &mut Context
) -> Result<Value>

Number.prototype.toFixed( [digits] )

The toFixed() method formats a number using fixed-point notation

More information:

ECMAScript reference

MDN documentation

pub(crate) fn to_locale_string(
this: &Value,
_: &[Value],
context: &mut Context
) -> Result<Value>

Number.prototype.toLocaleString( [locales [, options]] )

The toLocaleString() method returns a string with a language-sensitive representation of this number.

Note that while this technically conforms to the Ecma standard, it does no actual internationalization logic.

More information:

ECMAScript reference

MDN documentation

fn flt_str_to_exp(flt: &str) -> i32

flt_str_to_exp - used in to_precision

This function traverses a string representing a number, returning the floored log10 of this number.

fn round_to_precision(digits: &mut String, precision: usize) -> bool

round_to_precision - used in to_precision

This procedure has two roles:

If there are enough or more than enough digits in the string to show the required precision, the number represented by these digits is rounded using string manipulation.

Else, zeroes are appended to the string.

Additionnally, sometimes the exponent was wrongly computed and while up-rounding we find that we need an extra digit. When this happens, we return true so that the calling context can adjust the exponent. The string is kept at an exact length of precision.

When this procedure returns, digits is exactly precision long.

pub(crate) fn to_precision(
this: &Value,
args: &[Value],
context: &mut Context
) -> Result<Value>

Number.prototype.toPrecision( [precision] )

The toPrecision() method returns a string representing the Number object to the specified precision.

More information:

ECMAScript reference

MDN documentation

fn next_after(x: f64, y: f64) -> f64

pub(crate) fn to_native_string_radix(value: f64, radix: u8) -> String

pub(crate) fn to_native_string(x: f64) -> String

pub(crate) fn to_string(
this: &Value,
args: &[Value],
context: &mut Context
) -> Result<Value>

Number.prototype.toString( [radix] )

The toString() method returns a string representing the specified Number object.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn value_of(
this: &Value,
_: &[Value],
context: &mut Context
) -> Result<Value>

Number.prototype.toString()

The valueOf() method returns the wrapped primitive value of a Number object.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn parse_int(
_: &Value,
args: &[Value],
context: &mut Context
) -> Result<Value>

Builtin javascript ‘parseInt(str, radix)’ function.

Parses the given string as an integer using the given radix as a base.

An argument of type Number (i.e. Integer or Rational) is also accepted in place of string.

The radix must be an integer in the range [2, 36] inclusive.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn parse_float(
_: &Value,
args: &[Value],
context: &mut Context
) -> Result<Value>

Builtin javascript ‘parseFloat(str)’ function.

Parses the given string as a floating point value.

An argument of type Number (i.e. Integer or Rational) is also accepted in place of string.

To improve performance an Integer type Number is returned in place of a Rational if the given string can be parsed and stored as an Integer.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn global_is_finite(
_: &Value,
args: &[Value],
context: &mut Context
) -> Result<Value>

Builtin javascript ‘isFinite(number)’ function.

Converts the argument to a number, throwing a type error if the conversion is invalid.

If the number is NaN, +∞, or -∞ false is returned.

Otherwise true is returned.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn global_is_nan(
_: &Value,
args: &[Value],
context: &mut Context
) -> Result<Value>

Builtin javascript ‘isNaN(number)’ function.

Converts the argument to a number, throwing a type error if the conversion is invalid.

If the number is NaN true is returned.

Otherwise false is returned.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn number_is_finite(
_: &Value,
args: &[Value],
_ctx: &mut Context
) -> Result<Value>

Number.isFinite( number )

Checks if the argument is a number, returning false if it isn’t.

If the number is NaN, +∞, or -∞ false is returned.

Otherwise true is returned.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn number_is_integer(
_: &Value,
args: &[Value],
_ctx: &mut Context
) -> Result<Value>

Number.isInteger( number )

Checks if the argument is an integer.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn number_is_nan(
_: &Value,
args: &[Value],
_ctx: &mut Context
) -> Result<Value>

Number.isNaN( number )

Checks if the argument is a number, returning false if it isn’t.

If the number is NaN true is returned.

Otherwise false is returned.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn is_safe_integer(
_: &Value,
args: &[Value],
_ctx: &mut Context
) -> Result<Value>

Number.isSafeInteger( number )

Checks if the argument is an integer, returning false if it isn’t.

If abs(number) ≤ MAX_SAFE_INTEGER true is returned.

Otherwise false is returned.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn is_integer(val: &Value) -> bool

Checks if the argument is a finite integer Number value.

More information:

ECMAScript reference

pub(crate) fn is_float_integer(number: f64) -> bool

Checks if the float argument is an integer.

pub(crate) fn equal(x: f64, y: f64) -> bool

The abstract operation Number::equal takes arguments x (a Number) and y (a Number). It performs the following steps when called:

https://tc39.es/ecma262/#sec-numeric-types-number-equal

pub(crate) fn same_value(a: f64, b: f64) -> bool

The abstract operation Number::sameValue takes arguments x (a Number) and y (a Number). It performs the following steps when called:

https://tc39.es/ecma262/#sec-numeric-types-number-sameValue

pub(crate) fn same_value_zero(x: f64, y: f64) -> bool

The abstract operation Number::sameValueZero takes arguments x (a Number) and y (a Number). It performs the following steps when called:

https://tc39.es/ecma262/#sec-numeric-types-number-sameValueZero

pub(crate) fn less_than(x: f64, y: f64) -> AbstractRelation