Module boa::builtins::object

This module implements the global Object object.

The Object class represents one of JavaScript's data types.

It is used to store various keyed collections and more complex entities. Objects can be created using the Object() constructor or the object initializer / literal syntax.

More information:

ECMAScript reference

MDN documentation

Modules

internal_methods
This module defines the object internal methods.

internal_state
Implementations for storing normal rust structs inside any object as internal state.

Structs

InternalStateCell
Wrapper around Rc to implement Trace and Finalize.

Object
The internal representation of an JavaScript object.

Enums

ObjectData
Defines the different types of objects.

Statics

INSTANCE_PROTOTYPE
Static __proto__, usually set on Object instances as a key to point to their respective prototype object.

PROTOTYPE
Static prototype, usually set on constructors as a key to point to their respective prototype object.

Traits

InternalState
This trait must be implemented by all structs used for internal state.

Functions

create
Create a new Object object.

define_property
Define a property in an object

get_prototype_of
Get the prototype of an object.

has_own_property
Object.prototype.hasOwnPrototype( property )

init
Initialise the Object object on the global object.

is
Uses the SameValue algorithm to check equality of objects

make_object
Create a new object.

property_is_enumerable

set_prototype_of
Set the prototype of an object.

to_string
Object.prototype.toString()