Struct boa::builtins::object::Object

pub struct Object;

The global JavaScript object.

Implementations

impl Object

const LENGTH: usize

fn constructor(
new_target: &JsValue,
args: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

pub fn create(
_: &JsValue,
args: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

Object.create( proto, [propertiesObject] )

Creates a new object from the provided prototype.

More information:

ECMAScript reference

MDN documentation

pub fn get_own_property_descriptor(
_: &JsValue,
args: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

Object.getOwnPropertyDescriptor( object, property )

Returns an object describing the configuration of a specific property on a given object.

More information:

ECMAScript reference

MDN documentation

pub fn get_own_property_descriptors(
_: &JsValue,
args: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

Object.getOwnPropertyDescriptors( object )

Returns all own property descriptors of a given object.

More information:

ECMAScript reference

MDN documentation

fn from_property_descriptor(
desc: PropertyDescriptor,
context: &mut Context
) -> JsValue

The abstract operation FromPropertyDescriptor.

ECMAScript reference

pub fn is(_: &JsValue, args: &[JsValue], _: &mut Context) -> JsResult<JsValue>

Uses the SameValue algorithm to check equality of objects

pub fn get_prototype_of(
_: &JsValue,
args: &[JsValue],
ctx: &mut Context
) -> JsResult<JsValue>

Get the prototype of an object.

pub fn set_prototype_of(
_: &JsValue,
args: &[JsValue],
ctx: &mut Context
) -> JsResult<JsValue>

Set the prototype of an object.

More information

pub fn is_prototype_of(
this: &JsValue,
args: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

Object.prototype.isPrototypeOf( proto )

Check whether or not an object exists within another object’s prototype chain.

More information:

ECMAScript reference

MDN documentation

pub fn define_property(
_: &JsValue,
args: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

Define a property in an object

pub fn define_properties(
_: &JsValue,
args: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

Object.defineProperties( proto, [propertiesObject] )

Creates or update own properties to the object

More information:

ECMAScript reference

MDN documentation

pub fn value_of(
this: &JsValue,
_: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

Object.prototype.valueOf()

More information:

ECMAScript reference

MDN documentation

pub fn to_string(
this: &JsValue,
_: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

Object.prototype.toString()

This method returns a string representing the object.

More information:

ECMAScript reference

MDN documentation

pub fn has_own_property(
this: &JsValue,
args: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

Object.prototype.hasOwnPrototype( property )

The method returns a boolean indicating whether the object has the specified property as its own property (as opposed to inheriting it).

More information:

ECMAScript reference

MDN documentation

pub fn property_is_enumerable(
this: &JsValue,
args: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

Object.prototype.propertyIsEnumerable( property )

This method returns a Boolean indicating whether the specified property is enumerable and is the object’s own property.

More information:

ECMAScript reference

MDN documentation

pub fn assign(
_: &JsValue,
args: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

Object.assign( target, ...sources )

This method copies all enumerable own properties from one or more source objects to a target object. It returns the target object.

More information:

ECMAScript reference

MDN documentation

pub fn keys(
_: &JsValue,
args: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

Object.keys( target )

This method returns an array of a given object’s own enumerable property names, iterated in the same order that a normal loop would.

More information:

ECMAScript reference

MDN documentation

pub fn values(
_: &JsValue,
args: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

Object.values( target )

More information:

ECMAScript reference

MDN documentation

pub fn entries(
_: &JsValue,
args: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

Object.entries( target )

This method returns an array of a given object’s own enumerable string-keyed property [key, value] pairs. This is the same as iterating with a for…in loop, except that a for…in loop enumerates properties in the prototype chain as well).

More information:

ECMAScript reference

MDN documentation

pub fn seal(
_: &JsValue,
args: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

Object.seal( target )

More information:

ECMAScript reference

MDN documentation

pub fn is_sealed(
_: &JsValue,
args: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

Object.isSealed( target )

More information:

ECMAScript reference

MDN documentation

pub fn freeze(
_: &JsValue,
args: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

Object.freeze( target )

More information:

ECMAScript reference

MDN documentation

pub fn is_frozen(
_: &JsValue,
args: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

Object.isFrozen( target )

More information:

ECMAScript reference

MDN documentation

pub fn prevent_extensions(
_: &JsValue,
args: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

Object.preventExtensions( target )

More information:

ECMAScript reference

MDN documentation

pub fn is_extensible(
_: &JsValue,
args: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

Object.isExtensible( target )

More information:

ECMAScript reference

MDN documentation