Struct boa::builtins::regexp::RegExp

pub struct RegExp {
    matcher: Regex,
    use_last_index: bool,
    flags: Box<str>,
    dot_all: bool,
    global: bool,
    ignore_case: bool,
    multiline: bool,
    sticky: bool,
    unicode: bool,
    pub(crate) original_source: Box<str>,
    original_flags: Box<str>,
}

The internal representation on a RegExp object.

Fields
matcher: Regex

Regex matcher.

use_last_index: bool

Update last_index, set if global or sticky flags are set.

flags: Box<str>

String of parsed flags.

dot_all: bool

Flag 's' - dot matches newline characters.

global: bool

Flag 'g'

ignore_case: bool

Flag 'i' - ignore case.

multiline: bool

Flag 'm' - '^' and '$' match beginning/end of line.

sticky: bool

Flag 'y'

unicode: bool

Flag 'u' - Unicode.

original_source: Box<str>original_flags: Box<str>
Implementations

impl RegExp

pub(crate) const NAME: &'static str

The name of the object.

pub(crate) const LENGTH: usize

The amount of arguments this function object takes.

pub(crate) fn constructor(
this: &Value,
args: &[Value],
_: &mut Context
) -> Result<Value>

Create a new RegExp

pub(crate) fn test(
this: &Value,
args: &[Value],
ctx: &mut Context
) -> Result<Value>

RegExp.prototype.test( string )

The test() method executes a search for a match between a regular expression and a specified string.

Returns true or false.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn exec(
this: &Value,
args: &[Value],
ctx: &mut Context
) -> Result<Value>

RegExp.prototype.exec( string )

The exec() method executes a search for a match in a specified string.

Returns a result array, or null.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn match(
this: &Value,
arg: RcString,
ctx: &mut Context
) -> Result<Value>

RegExp.prototype[ @@match ]( string )

This method retrieves the matches when matching a string against a regular expression.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn to_string(
this: &Value,
_: &[Value],
context: &mut Context
) -> Result<Value>

RegExp.prototype.toString()

Return a string representing the regular expression.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn match_all(this: &Value, arg_str: String) -> Result<Value>

RegExp.prototype[ @@matchAll ]( string )

The [@@matchAll] method returns all matches of the regular expression against a string.

More information:

ECMAScript reference

MDN documentation