Struct boa::builtins::regexp::RegExp

pub struct RegExp {
    matcher: Regex,
    use_last_index: bool,
    flags: Box<str>,
    dot_all: bool,
    global: bool,
    ignore_case: bool,
    multiline: bool,
    sticky: bool,
    unicode: bool,
    pub(crate) original_source: Box<str>,
    original_flags: Box<str>,
}

The internal representation on a RegExp object.

Fields
matcher: Regex

Regex matcher.

use_last_index: bool

Update last_index, set if global or sticky flags are set.

flags: Box<str>

String of parsed flags.

dot_all: bool

Flag ‘s’ - dot matches newline characters.

global: bool

Flag ‘g’

ignore_case: bool

Flag ‘i’ - ignore case.

multiline: bool

Flag ‘m’ - ‘^’ and ‘$’ match beginning/end of line.

sticky: bool

Flag ‘y’

unicode: bool

Flag ‘u’ - Unicode.

original_source: Box<str>original_flags: Box<str>
Implementations

impl RegExp

pub(crate) const NAME: &'static str

The name of the object.

pub(crate) const LENGTH: usize

The amount of arguments this function object takes.

pub(crate) fn constructor(
new_target: &Value,
args: &[Value],
ctx: &mut Context
) -> Result<Value>

Create a new RegExp

fn get_species(this: &Value, _: &[Value], _: &mut Context) -> Result<Value>

get RegExp[@@species]

The RegExp[@@species] accessor property returns the RegExp constructor.

More information:

ECMAScript reference

MDN documentation

fn regexp_has_flag(
this: &Value,
flag: char,
context: &mut Context
) -> Result<Value>

pub(crate) fn get_global(
this: &Value,
_: &[Value],
context: &mut Context
) -> Result<Value>

get RegExp.prototype.global

The global property indicates whether or not the “g” flag is used with the regular expression.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn get_ignore_case(
this: &Value,
_: &[Value],
context: &mut Context
) -> Result<Value>

get RegExp.prototype.ignoreCase

The ignoreCase property indicates whether or not the “i” flag is used with the regular expression.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn get_multiline(
this: &Value,
_: &[Value],
context: &mut Context
) -> Result<Value>

get RegExp.prototype.multiline

The multiline property indicates whether or not the “m” flag is used with the regular expression.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn get_dot_all(
this: &Value,
_: &[Value],
context: &mut Context
) -> Result<Value>

get RegExp.prototype.dotAll

The dotAll property indicates whether or not the “s” flag is used with the regular expression.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn get_unicode(
this: &Value,
_: &[Value],
context: &mut Context
) -> Result<Value>

get RegExp.prototype.unicode

The unicode property indicates whether or not the “u” flag is used with a regular expression. unicode is a read-only property of an individual regular expression instance.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn get_sticky(
this: &Value,
_: &[Value],
context: &mut Context
) -> Result<Value>

get RegExp.prototype.sticky

This flag indicates that it matches only from the index indicated by the lastIndex property of this regular expression in the target string (and does not attempt to match from any later indexes).

More information:

ECMAScript reference

MDN documentation

pub(crate) fn get_flags(
this: &Value,
_: &[Value],
context: &mut Context
) -> Result<Value>

get RegExp.prototype.flags

The flags property returns a string consisting of the flags of the current regular expression object.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn test(
this: &Value,
args: &[Value],
context: &mut Context
) -> Result<Value>

RegExp.prototype.test( string )

The test() method executes a search for a match between a regular expression and a specified string.

Returns true or false.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn exec(
this: &Value,
args: &[Value],
context: &mut Context
) -> Result<Value>

RegExp.prototype.exec( string )

The exec() method executes a search for a match in a specified string.

Returns a result array, or null.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn match(
this: &Value,
arg: RcString,
context: &mut Context
) -> Result<Value>

RegExp.prototype[ @@match ]( string )

This method retrieves the matches when matching a string against a regular expression.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn to_string(
this: &Value,
_: &[Value],
context: &mut Context
) -> Result<Value>

RegExp.prototype.toString()

Return a string representing the regular expression.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn match_all(
this: &Value,
arg_str: String,
context: &mut Context
) -> Result<Value>

RegExp.prototype[ @@matchAll ]( string )

The [@@matchAll] method returns all matches of the regular expression against a string.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn search(
this: &Value,
args: &[Value],
context: &mut Context
) -> Result<Value>

RegExp.prototype[ @@search ]( string )

This method executes a search for a match between a this regular expression and a string.

More information:

ECMAScript reference

MDN documentation