Module boa::builtins::string

This module implements the global String object.

The String global object is a constructor for strings or a sequence of characters.

More information:

ECMAScript reference

MDN documentation

Modules

string_iterator

Structs

String

JavaScript String implementation.

Functions

code_point_at

get_substitution

22.1.3.17.1 GetSubstitution ( matched, str, position, captures, namedCaptures, replacement )

is_leading_surrogate

is_trailing_surrogate

is_trimmable_whitespace

Helper function to check if a char is trimmable.

split_match

22.1.3.21.1 SplitMatch ( S, q, R )