Module boa::builtins::value

This module implements the JavaScript Value.

Javascript values, utility methods and conversion between Javascript values and Rust values.

Re-exports

pub use crate::builtins::value::val_type::Type;

pub use conversions::*;

pub use equality::*;

pub use operations::*;

Modules

conversions

display

equality

operations

val_type

Structs

Value
A Garbage-collected Javascript value as represented in the interpreter.

Enums

ValueData
A Javascript value

Type Definitions

ResultValue
The result of a Javascript expression is represented like this so it can succeed (Ok) or fail (Err)