Enum boa::builtins::value::ValueData

pub enum ValueData {
    Null,
    Undefined,
    Boolean(bool),
    String(String),
    Rational(f64),
    Integer(i32),
    BigInt(BigInt),
    Object(Box<GcCell<Object>>),
    Symbol(Symbol),
}

A Javascript value

Variants

Null

null - A null value, for when a value doesn't exist.

Undefined

undefined - An undefined value, for when a field or index doesn't exist.

Boolean(bool)

boolean - A true / false value, for if a certain criteria is met.

String(String)

String - A UTF-8 string, such as "Hello, world".

Rational(f64)

Number - A 64-bit floating point number, such as 3.1415

Integer(i32)

Number - A 32-bit integer, such as 42.

BigInt(BigInt)

BigInt - holds any arbitrary large signed integer.

Object(Box<GcCell<Object>>)

Object - An object, such as Math, represented by a binary tree of string keys to Javascript values.

Symbol(Symbol)

Symbol - A Symbol Primitive type.

Methods

impl ValueData

pub fn get_type(&self) -> Type

Get the type of the value.

This is similar to typeof as described at https://tc39.es/ecma262/#sec-typeof-operator but instead of returning a string it returns a Type enum which implements fmt::Display to allow getting the string if required using to_string().

impl ValueData

pub fn is_extensible(&self) -> bool

This will tell us if we can exten an object or not, not properly implemented yet

For now always returns true.

For scalar types it should be false, for objects check the private field for extensibilaty. By default true.

<https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Object/seal would turn extensible to false/> <https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Object/freeze would also turn extensible to false/>

pub fn is_object(&self) -> bool

Returns true if the value is an object

pub fn as_object(&self) -> Option<GcCellRef<Object>>

pub fn as_object_mut(&self) -> Option<GcCellRefMut<Object>>

pub fn is_symbol(&self) -> bool

Returns true if the value is a symbol.

pub fn is_function(&self) -> bool

Returns true if the value is a function

pub fn is_undefined(&self) -> bool

Returns true if the value is undefined.

pub fn is_null(&self) -> bool

Returns true if the value is null.

pub fn is_null_or_undefined(&self) -> bool

Returns true if the value is null or undefined.

pub fn is_double(&self) -> bool

Returns true if the value is a 64-bit floating-point number.

pub fn is_integer(&self) -> bool

Returns true if the value is integer.

pub fn is_number(&self) -> bool

Returns true if the value is a number.

pub fn is_string(&self) -> bool

Returns true if the value is a string.

pub fn is_boolean(&self) -> bool

Returns true if the value is a boolean.

pub fn is_bigint(&self) -> bool

Returns true if the value is a bigint.

pub fn as_bigint(&self) -> Option<&BigInt>

Returns an optional reference to a BigInt if the value is a BigInt primitive.

pub fn is_true(&self) -> bool

Returns true if the value is true.

toBoolean

pub fn to_number(&self) -> f64

Converts the value into a 64-bit floating point number

pub fn to_integer(&self) -> i32

Converts the value into a 32-bit integer

pub fn to_boolean(&self) -> bool

Creates a new boolean value from the input

pub fn remove_property(&self, field: &str) -> bool

Removes a property from a Value object.

It will return a boolean based on if the value was removed, if there was no value to remove false is returned.

pub fn get_property(&self, field: &str) -> Option<Property>

Resolve the property in the object.

A copy of the Property is returned.

pub fn update_property(
&self,
field: &str,
value: Option<Value>,
enumerable: Option<bool>,
writable: Option<bool>,
configurable: Option<bool>
)

update_prop will overwrite individual [Property] fields, unlike Set_prop, which will overwrite prop with a new Property Mostly used internally for now

pub fn get_internal_slot(&self, field: &str) -> Value

Resolve the property in the object.

Returns a copy of the Property.

pub fn get_field<F>(&self, field: F) -> Value where
F: Into<Value>,

Resolve the property in the object and get its value, or undefined if this is not an object or the field doesn't exist get_field recieves a Property from get_prop(). It should then return the [[Get]] result value if that's set, otherwise fall back to [[Value]] TODO: this function should use the get Value if its set

pub fn has_internal_state(&self) -> bool

Check whether an object has an internal state set.

pub fn get_internal_state(&self) -> Option<InternalStateCell>

Get the internal state of an object.

pub fn with_internal_state_ref<S, R, F>(&self, f: F) -> R where
S: Any + InternalState,
F: FnOnce(&S) -> R,

Run a function with a reference to the internal state.

Panics

This will panic if this value doesn't have an internal state or if the internal state doesn't have the concrete type S.

pub fn with_internal_state_mut<S, R, F>(&self, f: F) -> R where
S: Any + InternalState,
F: FnOnce(&mut S) -> R,

Run a function with a mutable reference to the internal state.

Panics

This will panic if this value doesn't have an internal state or if the internal state doesn't have the concrete type S.

pub fn has_field(&self, field: &str) -> bool

Check to see if the Value has the field, mainly used by environment records.

pub fn set_field<F, V>(&self, field: F, val: V) -> Value where
F: Into<Value>,
V: Into<Value>,

Set the field in the value Field could be a Symbol, so we need to accept a Value (not a string)

pub fn set_internal_slot(&self, field: &str, value: Value) -> Value

Set the private field in the value

pub fn set_data(&self, data: ObjectData)

Set the kind of an object.

pub fn set_property<S>(&self, field: S, property: Property) -> Property where
S: Into<String>,

Set the property in the value.

pub fn set_internal_state<T: Any + InternalState>(&self, state: T)

Set internal state of an Object. Discards the previous state if it was set.

pub fn from_func(function: Function) -> Value

Consume the function and return a Value