Trait boa::class::Class

pub trait Class: NativeObject + Sized {
    const NAME: &'static str;
    const LENGTH: usize;
    const ATTRIBUTES: Attribute;

    fn constructor(
        this: &JsValue, 
        args: &[JsValue], 
        context: &mut Context
    ) -> JsResult<Self>;

    fn init(class: &mut ClassBuilder<'_>) -> JsResult<()>;
}

Native class.

Associated Constants

const NAME: &'static str

The binding name of the object.

const LENGTH: usize

The amount of arguments the class constructor takes, default is 0.

const ATTRIBUTES: Attribute

The attibutes the class will be binded with, default is writable, enumerable, configurable.

Required methods

fn constructor(
this: &JsValue,
args: &[JsValue],
context: &mut Context
) -> JsResult<Self>

The constructor of the class.

fn init(class: &mut ClassBuilder<'_>) -> JsResult<()>

Initializes the internals and the methods of the class.