Trait boa::class::ClassConstructor

pub trait ClassConstructor: Class {
    fn raw_constructor(
        this: &Value, 
        args: &[Value], 
        ctx: &mut Context
    ) -> Result<Value>
    where
        Self: Sized;
}

This is a wrapper around Class::constructor that sets the internal data of a class.

This is automatically implemented, when a type implements Class.

Required methods

fn raw_constructor(
this: &Value,
args: &[Value],
ctx: &mut Context
) -> Result<Value> where
Self: Sized,

The raw constructor that mathces the NativeFunction signature.

Loading content...