Module boa::environment::declarative_environment_record

Declarative Records

Each declarative Environment Record is associated with an ECMAScript program scope containing variable, constant, let, class, module, import, and/or function declarations. A declarative Environment Record binds the set of identifiers defined by the declarations contained within its scope. More info: ECMA-262 sec-declarative-environment-records

Structs

DeclarativeEnvironmentRecord

A declarative Environment Record binds the set of identifiers defined by the declarations contained within its scope.

DeclarativeEnvironmentRecordBinding

Declarative Bindings have a few properties for book keeping purposes, such as mutability (const vs let). Can it be deleted? and strict mode.