Struct boa::environment::function_environment_record::FunctionEnvironmentRecord

pub struct FunctionEnvironmentRecord {
    pub declarative_record: DeclarativeEnvironmentRecord,
    pub this_value: Value,
    pub this_binding_status: BindingStatus,
    pub function: GcObject,
    pub home_object: Value,
    pub new_target: Value,
}

https://tc39.es/ecma262/#table-16

Fields
declarative_record: DeclarativeEnvironmentRecordthis_value: Value

This is the this value used for this invocation of the function.

this_binding_status: BindingStatus

If the value is “lexical”, this is an ArrowFunction and does not have a local this value.

function: GcObject

The function object whose invocation caused this Environment Record to be created.

home_object: Value

If the associated function has super property accesses and is not an ArrowFunction, [[HomeObject]] is the object that the function is bound to as a method. The default value for [[HomeObject]] is undefined.

new_target: Value

If this Environment Record was created by the [[Construct]] internal method, [[NewTarget]] is the value of the [[Construct]] newTarget parameter. Otherwise, its value is undefined.

Implementations

impl FunctionEnvironmentRecord

pub fn new(
f: GcObject,
this: Option<Value>,
outer: Option<Environment>,
binding_status: BindingStatus,
new_target: Value
) -> FunctionEnvironmentRecord

pub fn bind_this_value(&mut self, value: Value) -> Result<Value>

pub fn get_super_base(&self) -> Value