Struct boa::exec::Interpreter

pub struct Interpreter {
    is_return: bool,
    pub realm: Realm,
}

A Javascript intepreter

Fields
is_return: boolrealm: Realm

realm holds both the global object and the environment

Methods

impl Interpreter

pub(crate) fn get_realm(&self) -> &Realm

Get the Interpreter's realm

pub(crate) fn create_function(
&mut self,
args: Box<[FormalParameter]>,
expr: &Node,
this_mode: ThisMode
) -> Value

Utility to create a function Value for Function Declarations, Arrow Functions or Function Expressions

pub(crate) fn call(
&mut self,
f: &Value,
this: &mut Value,
arguments_list: &[Value]
) -> ResultValue

https://tc39.es/ecma262/#sec-call

fn ordinary_to_primitive(&mut self, o: &mut Value, hint: &str) -> Value

https://tc39.es/ecma262/#sec-ordinarytoprimitive

pub fn to_primitive(
&mut self,
input: &mut Value,
preferred_type: Option<&str>
) -> Value

The abstract operation ToPrimitive takes an input argument and an optional argument PreferredType. https://tc39.es/ecma262/#sec-toprimitive

pub fn to_string(&mut self, value: &Value) -> Value

Converts a value into a String.

https://tc39.es/ecma262/#sec-tostring

pub fn to_property_key(&mut self, value: &mut Value) -> Value

The abstract operation ToPropertyKey takes argument argument. It converts argument to a value that can be used as a property key. https://tc39.es/ecma262/#sec-topropertykey

pub fn has_property(&self, obj: &mut Value, key: &Value) -> bool

https://tc39.es/ecma262/#sec-hasproperty

pub fn to_object(&mut self, value: &Value) -> ResultValue

The abstract operation ToObject converts argument to a value of type Object https://tc39.es/ecma262/#sec-toobject

pub fn value_to_rust_string(&mut self, value: &Value) -> String

value_to_rust_string() converts a value into a rust heap allocated string

pub fn value_to_rust_number(&mut self, value: &Value) -> f64

fn extract_array_properties(&mut self, value: &Value) -> Result<Vec<Value>, ()>

extract_array_properties converts an array object into a rust vector of Values. This is useful for the spread operator, for any other object an Err is returned

fn set_value(&mut self, node: &Node, value: Value) -> ResultValue