Struct boa::exec::Interpreter

pub struct Interpreter {
    state: InterpreterState,
    pub realm: Realm,
    symbol_count: u32,
    console: Console,
}

A Javascript intepreter

Fields
state: InterpreterState

the current state of the interpreter.

realm: Realm

realm holds both the global object and the environment

symbol_count: u32

This is for generating an unique internal Symbol hash.

console: Console

console object state.

Implementations

impl Interpreter

pub fn construct_range_error<M>(&mut self, message: M) -> Value where
M: Into<String>,

Constructs a RangeError with the specified message.

pub fn throw_range_error<M>(&mut self, message: M) -> Result<Value> where
M: Into<String>,

Throws a RangeError with the specified message.

pub fn construct_type_error<M>(&mut self, message: M) -> Value where
M: Into<String>,

Constructs a TypeError with the specified message.

pub fn throw_type_error<M>(&mut self, message: M) -> Result<Value> where
M: Into<String>,

Throws a TypeError with the specified message.

pub fn construct_reference_error<M>(&mut self, message: M) -> Value where
M: Into<String>,

Constructs a ReferenceError with the specified message.

pub fn throw_reference_error<M>(&mut self, message: M) -> Result<Value> where
M: Into<String>,

Throws a ReferenceError with the specified message.

pub fn construct_syntax_error<M>(&mut self, message: M) -> Value where
M: Into<String>,

Constructs a SyntaxError with the specified message.

pub fn throw_syntax_error<M>(&mut self, message: M) -> Result<Value> where
M: Into<String>,

Throws a SyntaxError with the specified message.

impl Interpreter

pub fn new(realm: Realm) -> Self

Creates a new interpreter.

fn create_intrinsics(&mut self)

Sets up the default global objects within Global

pub(crate) fn realm(&self) -> &Realm

Retrieves the Realm of this executor.

pub(crate) fn realm_mut(&mut self) -> &mut Realm

Retrieves the Realm of this executor as a mutable reference.

pub(crate) fn global(&self) -> &Value

Retrieves the global object of the Realm of this executor.

pub(crate) fn generate_hash(&mut self) -> u32

Generates a new Symbol internal hash.

This currently is an incremented value.

pub(crate) fn create_function<P, B>(
&mut self,
params: P,
body: B,
this_mode: ThisMode,
constructable: bool,
callable: bool
) -> Value where
P: Into<Box<[FormalParameter]>>,
B: Into<StatementList>,

Utility to create a function Value for Function Declarations, Arrow Functions or Function Expressions

pub(crate) fn call(
&mut self,
f: &Value,
this: &Value,
arguments_list: &[Value]
) -> Result<Value>

https://tc39.es/ecma262/#sec-call

pub(crate) fn extract_array_properties(
&mut self,
value: &Value
) -> StdResult<Vec<Value>, ()>

Converts an array object into a rust vector of values.

This is useful for the spread operator, for any other object an Err is returned

pub(crate) fn ordinary_to_primitive(
&mut self,
o: &Value,
hint: PreferredType
) -> Result<Value>

Converts an object to a primitive.

More information:

ECMAScript

pub(crate) fn has_property(&self, obj: &Value, key: &PropertyKey) -> bool

https://tc39.es/ecma262/#sec-hasproperty

fn set_value(&mut self, node: &Node, value: Value) -> Result<Value>

pub(crate) fn set_current_state(&mut self, new_state: InterpreterState)

pub(crate) fn get_current_state(&self) -> &InterpreterState

pub fn require_object_coercible<'a>(
&mut self,
value: &'a Value
) -> Result<&'a Value>

Check if the Value can be converted to an Object

The abstract operation RequireObjectCoercible takes argument argument. It throws an error if argument is a value that cannot be converted to an Object using ToObject. It is defined by Table 15

More information:

ECMAScript reference

pub(crate) fn console(&self) -> &Console

A helper function for getting a immutable reference to the console object.

pub(crate) fn console_mut(&mut self) -> &mut Console

A helper function for getting a mutable reference to the console object.