Module boa::gc

This module represents the main way to interact with the garbacge collector.

Macros

custom_trace
This rule implements the trace method.

empty_trace
This rule implements the trace methods with empty implementations.

Structs

GcObject
Garbage collected Object.

Ref
A wrapper type for an immutably borrowed value from a GcCell<T>.

RefMut
A wrapper type for a mutably borrowed value from a GcCell<T>.

Traits

Finalize
The Finalize trait. Can be specialized for a specific type to define finalization logic for that type.

Trace
The Trace trait, which needs to be implemented on garbage-collected objects.

Functions

force_collect
Immediately triggers a garbage collection on the current thread.

Derive Macros

Finalize

Trace