Module boa::gc

This module represents the main way to interact with the garbage collector.

Macros

custom_trace
This rule implements the trace method.

empty_trace
This rule implements the trace methods with empty implementations.

Structs

Ref
A wrapper type for an immutably borrowed value from a GcCell<T>.

RefMut
A wrapper type for a mutably borrowed value from a GcCell<T>.

Traits

Finalize
The Finalize trait, which needs to be implemented on garbage-collected objects to define finalization logic.

Trace
The Trace trait, which needs to be implemented on garbage-collected objects.

Functions

force_collect
Immediately triggers a garbage collection on the current thread.

Derive Macros

Finalize

Trace