Struct boa::object::ConstructorBuilder

pub struct ConstructorBuilder<'context> {
    context: &'context mut Context,
    constrcutor_function: NativeFunction,
    constructor_object: GcObject,
    prototype: GcObject,
    name: Option<String>,
    length: usize,
    callable: bool,
    constructable: bool,
    inherit: Option<Value>,
}

Builder for creating constructors objects, like Array.

Fields
context: &'context mut Contextconstrcutor_function: NativeFunctionconstructor_object: GcObjectprototype: GcObjectname: Option<String>length: usizecallable: boolconstructable: boolinherit: Option<Value>
Implementations

impl<'context> ConstructorBuilder<'context>

pub fn new(context: &'context mut Context, constructor: NativeFunction) -> Self

Create a new ConstructorBuilder.

pub(crate) fn with_standard_object(
context: &'context mut Context,
constructor: NativeFunction,
object: StandardConstructor
) -> Self

pub fn method<B>(
&mut self,
function: NativeFunction,
binding: B,
length: usize
) -> &mut Self where
B: Into<FunctionBinding>,

Add new method to the constructors prototype.

pub fn static_method<B>(
&mut self,
function: NativeFunction,
binding: B,
length: usize
) -> &mut Self where
B: Into<FunctionBinding>,

Add new static method to the constructors object itself.

pub fn property<K, V>(
&mut self,
key: K,
value: V,
attribute: Attribute
) -> &mut Self where
K: Into<PropertyKey>,
V: Into<Value>,

Add new property to the constructors prototype.

pub fn static_property<K, V>(
&mut self,
key: K,
value: V,
attribute: Attribute
) -> &mut Self where
K: Into<PropertyKey>,
V: Into<Value>,

Add new static property to the constructors object itself.

pub fn length(&mut self, length: usize) -> &mut Self

Specify how many arguments the constructor function takes.

Default is 0.

pub fn name<N>(&mut self, name: N) -> &mut Self where
N: AsRef<str>,

Specify the name of the constructor function.

Default is "[object]"

pub fn callable(&mut self, callable: bool) -> &mut Self

Specify whether the constructor function can be called.

Default is true

pub fn constructable(&mut self, constructable: bool) -> &mut Self

Specify whether the constructor function can be called with new keyword.

Default is true

pub fn inherit(&mut self, prototype: Value) -> &mut Self

Specify the prototype this constructor object inherits from.

Default is Object.prototype

pub fn context(&mut self) -> &mut Context

Return the current context.

pub fn build(&mut self) -> GcObject

Build the constructor function object.