Struct boa::object::Object

pub struct Object {
    pub data: ObjectData,
    properties: PropertyMap,
    prototype: JsPrototype,
    extensible: bool,
}

The internal representation of a JavaScript object.

Fields
data: ObjectData

The type of the object.

properties: PropertyMap

The collection of properties contained in the object

prototype: JsPrototype

Instance prototype __proto__.

extensible: bool

Whether it can have new properties added to it.

Implementations

impl Object

pub fn kind(&self) -> &ObjectKind

pub fn is_array(&self) -> bool

Checks if it an Array object.

pub fn as_array(&self) -> Option<()>

pub fn is_array_iterator(&self) -> bool

Checks if it is an ArrayIterator object.

pub fn as_array_iterator(&self) -> Option<&ArrayIterator>

pub fn is_array_buffer(&self) -> bool

Checks if it an ArrayBuffer object.

pub fn as_array_buffer(&self) -> Option<&ArrayBuffer>

pub fn as_array_buffer_mut(&mut self) -> Option<&mut ArrayBuffer>

pub fn as_array_iterator_mut(&mut self) -> Option<&mut ArrayIterator>

pub fn as_string_iterator_mut(&mut self) -> Option<&mut StringIterator>

pub fn as_regexp_string_iterator_mut(
&mut self
) -> Option<&mut RegExpStringIterator>

pub fn as_for_in_iterator(&self) -> Option<&ForInIterator>

pub fn as_for_in_iterator_mut(&mut self) -> Option<&mut ForInIterator>

pub fn is_map(&self) -> bool

Checks if it is a Map object.pub

pub fn as_map_ref(&self) -> Option<&OrderedMap<JsValue>>

pub fn as_map_mut(&mut self) -> Option<&mut OrderedMap<JsValue>>

pub fn is_map_iterator(&self) -> bool

pub fn as_map_iterator_ref(&self) -> Option<&MapIterator>

pub fn as_map_iterator_mut(&mut self) -> Option<&mut MapIterator>

pub fn is_set(&self) -> bool

pub fn as_set_ref(&self) -> Option<&OrderedSet<JsValue>>

pub fn as_set_mut(&mut self) -> Option<&mut OrderedSet<JsValue>>

pub fn as_set_iterator_mut(&mut self) -> Option<&mut SetIterator>

pub fn is_string(&self) -> bool

Checks if it a String object.

pub fn as_string(&self) -> Option<JsString>

pub fn is_function(&self) -> bool

Checks if it a Function object.

pub fn as_function(&self) -> Option<&Function>

pub fn as_function_mut(&mut self) -> Option<&mut Function>

pub fn as_bound_function(&self) -> Option<&BoundFunction>

pub fn is_symbol(&self) -> bool

Checks if it a Symbol object.

pub fn as_symbol(&self) -> Option<JsSymbol>

pub fn is_error(&self) -> bool

Checks if it an Error object.

pub fn as_error(&self) -> Option<()>

pub fn is_boolean(&self) -> bool

Checks if it a Boolean object.

pub fn as_boolean(&self) -> Option<bool>

pub fn is_number(&self) -> bool

Checks if it a Number object.

pub fn as_number(&self) -> Option<f64>

pub fn is_bigint(&self) -> bool

Checks if it a BigInt object.

pub fn as_bigint(&self) -> Option<&JsBigInt>

pub fn is_date(&self) -> bool

pub fn as_date(&self) -> Option<&Date>

pub fn is_regexp(&self) -> bool

Checks if it a RegExp object.

pub fn as_regexp(&self) -> Option<&RegExp>

Gets the regexp data if the object is a regexp.

pub fn is_typed_array(&self) -> bool

Checks if it a TypedArray object.

pub fn as_data_view(&self) -> Option<&DataView>

pub fn as_data_view_mut(&mut self) -> Option<&mut DataView>

pub fn is_arguments(&self) -> bool

Checks if it is an Arguments object.

pub fn as_mapped_arguments(&self) -> Option<&MappedArguments>

Gets the mapped arguments data if this is a mapped arguments object.

pub fn as_typed_array(&self) -> Option<&IntegerIndexed>

Gets the typed array data (integer indexed object) if this is a typed array.

pub fn as_typed_array_mut(&mut self) -> Option<&mut IntegerIndexed>

Gets the typed array data (integer indexed object) if this is a typed array.

pub fn is_ordinary(&self) -> bool

Checks if it an ordinary object.

pub fn is_proxy(&self) -> bool

Checks if it’s an proxy object.

pub fn as_proxy(&self) -> Option<&Proxy>

pub fn as_proxy_mut(&mut self) -> Option<&mut Proxy>

pub fn prototype(&self) -> &JsPrototype

Gets the prototype instance of this object.

pub fn set_prototype<O: Into<JsPrototype>>(&mut self, prototype: O) -> bool

Sets the prototype instance of the object.

More information

pub fn is_native_object(&self) -> bool

Returns true if it holds an Rust type that implements NativeObject.

pub fn as_native_object(&self) -> Option<&dyn NativeObject>

pub fn is<T>(&self) -> bool where
T: NativeObject,

Return true if it is a native object and the native type is T.

pub fn downcast_ref<T>(&self) -> Option<&T> where
T: NativeObject,

Downcast a reference to the object, if the object is type native object type T.

pub fn downcast_mut<T>(&mut self) -> Option<&mut T> where
T: NativeObject,

Downcast a mutable reference to the object, if the object is type native object type T.

pub fn properties(&self) -> &PropertyMap

pub(crate) fn insert<K, P>(
&mut self,
key: K,
property: P
) -> Option<PropertyDescriptor> where
K: Into<PropertyKey>,
P: Into<PropertyDescriptor>,

Helper function for property insertion.

pub(crate) fn remove(&mut self, key: &PropertyKey) -> Option<PropertyDescriptor>

Helper function for property removal.

pub fn insert_property<K, P>(
&mut self,
key: K,
property: P
) -> Option<PropertyDescriptor> where
K: Into<PropertyKey>,
P: Into<PropertyDescriptor>,

Inserts a field in the object properties without checking if it’s writable.

If a field was already in the object with the same name that a Some is returned with that field, otherwise None is returned.