Module boa::property

This module implements the Property Descriptor.

The Property Descriptor type is used to explain the manipulation and reification of Object property attributes. Values of the Property Descriptor type are Records. Each field’s name is an attribute name and its value is a corresponding attribute value as specified in 6.1.7.1. In addition, any field may be present or absent. The schema name used within this specification to tag literal descriptions of Property Descriptor records is PropertyDescriptor.

More information:

MDN documentation

ECMAScript reference

Modules

attribute

This module implements the Attribute struct which contains the attibutes for property descriptors.

Structs

Attribute

This struct constains the property flags as described in the ECMAScript specification.

PropertyDescriptor

This represents a JavaScript Property AKA The Property Descriptor.

PropertyDescriptorBuilder

Enums

DescriptorKind

PropertyKey

This abstracts away the need for IsPropertyKey by transforming the PropertyKey values into an enum with both valid types: String and Symbol

PropertyNameKind