Module boa::syntax::ast::node

This module implements the Node structure, which composes the AST.

Re-exports

pub use self::array::ArrayDecl;

pub use self::await_expr::AwaitExpr;

pub use self::block::Block;

pub use self::break_node::Break;

pub use self::call::Call;

pub use self::conditional::ConditionalOp;

pub use self::conditional::If;

pub use self::declaration::ArrowFunctionDecl;

pub use self::declaration::AsyncFunctionDecl;

pub use self::declaration::AsyncFunctionExpr;

pub use self::declaration::Declaration;

pub use self::declaration::DeclarationList;

pub use self::declaration::FunctionDecl;

pub use self::declaration::FunctionExpr;

pub use self::field::GetConstField;

pub use self::field::GetField;

pub use self::identifier::Identifier;

pub use self::iteration::Continue;

pub use self::iteration::DoWhileLoop;

pub use self::iteration::ForInLoop;

pub use self::iteration::ForLoop;

pub use self::iteration::ForOfLoop;

pub use self::iteration::WhileLoop;

pub use self::new::New;

pub use self::object::Object;

pub use self::operator::Assign;

pub use self::operator::BinOp;

pub use self::operator::UnaryOp;

pub use self::return_smt::Return;

pub use self::spread::Spread;

pub use self::statement_list::RcStatementList;

pub use self::statement_list::StatementList;

pub use self::switch::Case;

pub use self::switch::Switch;

pub use self::template::TaggedTemplate;

pub use self::template::TemplateLit;

pub use self::throw::Throw;

pub use self::try_node::Catch;

pub use self::try_node::Finally;

pub use self::try_node::Try;

Modules

array
Array declaration node.

await_expr
Await expression node.

block
Block AST node.

break_node

call

conditional
Conditional nodes

declaration
Declaration nodes

field
Field nodes

identifier
Local identifier node.

iteration
Iteration nodes

new

object
Object node.

operator
Operator nodes

return_smt

spread

statement_list
Statement list node.

switch
Switch node.

template
Template literal node.

throw

try_node

Structs

FormalParameter
“Formal parameter” is a fancy way of saying “function parameter”.

Enums

MethodDefinitionKind
Method definition kinds.

Node

PropertyDefinition
A JavaScript property is a characteristic of an object, often describing attributes associated with a data structure.

Functions

join_nodes
Utility to join multiple Nodes into a single string.