Enum boa::syntax::ast::node::Node

pub enum Node {
    ArrayDecl(Vec<Node>),
    ArrowFunctionDecl(Vec<FormalParameter>, Box<Node>),
    Assign(Box<Node>, Box<Node>),
    BinOp(BinOp, Box<Node>, Box<Node>),
    Block(Vec<Node>),
    Break(Option<String>),
    Call(Box<Node>, Vec<Node>),
    ConditionalOp(Box<Node>, Box<Node>, Box<Node>),
    Const(Const),
    ConstDecl(Vec<(String, Node)>),
    Continue(Option<String>),
    DoWhileLoop(Box<Node>, Box<Node>),
    FunctionDecl(Option<String>, Vec<FormalParameter>, Box<Node>),
    GetConstField(Box<Node>, String),
    GetField(Box<Node>, Box<Node>),
    ForLoop(Option<Box<Node>>, Option<Box<Node>>, Option<Box<Node>>, Box<Node>),
    If(Box<Node>, Box<Node>, Option<Box<Node>>),
    LetDecl(Vec<(String, Option<Node>)>),
    Local(String),
    New(Box<Node>),
    Object(Vec<PropertyDefinition>),
    Return(Option<Box<Node>>),
    Switch(Box<Node>, Vec<(Node, Vec<Node>)>, Option<Box<Node>>),
    Spread(Box<Node>),
    StatementList(Vec<Node>),
    Throw(Box<Node>),
    TypeOf(Box<Node>),
    Try(Box<Node>, Option<Box<Node>>, Option<Box<Node>>, Option<Box<Node>>),
    This,
    UnaryOp(UnaryOp, Box<Node>),
    VarDecl(Vec<(String, Option<Node>)>),
    WhileLoop(Box<Node>, Box<Node>),
}

A Javascript AST Node.

Variants

ArrayDecl(Vec<Node>)

An array is an ordered collection of data (either primitive or object depending upon the language).

Arrays are used to store multiple values in a single variable. This is compared to a variable that can store only one value.

Each item in an array has a number attached to it, called a numeric index, that allows you to access it. In JavaScript, arrays start at index zero and can be manipulated with various methods.

More information:

ECMAScript reference

MDN documentation

ArrowFunctionDecl(Vec<FormalParameter>, Box<Node>)

An arrow function expression is a syntactically compact alternative to a regular function expression.

Arrow function expressions are ill suited as methods, and they cannot be used as constructors. Arrow functions cannot be used as constructors and will throw an error when used with new.

More information:

ECMAScript reference

MDN documentation

Assign(Box<Node>, Box<Node>)

An assignment operator assigns a value to its left operand based on the value of its right operand.

Assignment operator (=), assigns the value of its right operand to its left operand.

More information:

ECMAScript reference

MDN documentation

BinOp(BinOp, Box<Node>, Box<Node>)

Binary operators requires two operands, one before the operator and one after the operator.

More information:

MDN documentation

Block(Vec<Node>)

A block statement (or compound statement in other languages) is used to group zero or more statements.

The block statement is often called compound statement in other languages. It allows you to use multiple statements where JavaScript expects only one statement. Combining statements into blocks is a common practice in JavaScript. The opposite behavior is possible using an empty statement, where you provide no statement, although one is required.

More information:

ECMAScript reference

MDN documentation

Break(Option<String>)

The break statement terminates the current loop, switch, or label statement and transfers program control to the statement following the terminated statement.

The break statement includes an optional label that allows the program to break out of a labeled statement. The break statement needs to be nested within the referenced label. The labeled statement can be any block statement; it does not have to be preceded by a loop statement.

More information:

ECMAScript reference

MDN documentation

Call(Box<Node>, Vec<Node>)

Calling the function actually performs the specified actions with the indicated parameters.

Defining a function does not execute it. Defining it simply names the function and specifies what to do when the function is called. Functions must be in scope when they are called, but the function declaration can be hoisted The scope of a function is the function in which it is declared (or the entire program, if it is declared at the top level).

More information:

ECMAScript reference

MDN documentation

ConditionalOp(Box<Node>, Box<Node>, Box<Node>)

The conditional (ternary) operator is the only JavaScript operator that takes three operands.

This operator is the only JavaScript operator that takes three operands: a condition followed by a question mark (?), then an expression to execute if the condition is truthy followed by a colon (:), and finally the expression to execute if the condition is falsy. This operator is frequently used as a shortcut for the if statement.

More information:

ECMAScript reference

MDN documentation

Const(Const)

Literals represent values in JavaScript.

These are fixed values not variables that you literally provide in your script.

More information:

ECMAScript reference

MDN documentation

ConstDecl(Vec<(String, Node)>)

The const statements are block-scoped, much like variables defined using the let keyword.

This declaration creates a constant whose scope can be either global or local to the block in which it is declared. Global constants do not become properties of the window object, unlike var variables.

An initializer for a constant is required. You must specify its value in the same statement in which it's declared. (This makes sense, given that it can't be changed later.)

More information:

ECMAScript reference

MDN documentation

Continue(Option<String>)

The continue statement terminates execution of the statements in the current iteration of the current or labeled loop, and continues execution of the loop with the next iteration.

The continue statement can include an optional label that allows the program to jump to the next iteration of a labeled loop statement instead of the current loop. In this case, the continue statement needs to be nested within this labeled statement.

More information:

ECMAScript reference

MDN documentation

DoWhileLoop(Box<Node>, Box<Node>)

The do...while statement creates a loop that executes a specified statement until the test condition evaluates to false.

The condition is evaluated after executing the statement, resulting in the specified statement executing at least once.

More information:

ECMAScript reference

MDN documentation

FunctionDecl(Option<String>, Vec<FormalParameter>, Box<Node>)

The function declaration (function statement) defines a function with the specified parameters.

A function created with a function declaration is a Function object and has all the properties, methods and behavior of Function.

A function can also be created using an expression (see function expression).

By default, functions return undefined. To return any other value, the function must have a return statement that specifies the value to return.

More information:

ECMAScript reference

MDN documentation

GetConstField(Box<Node>, String)

This property accessor provides access to an object's properties by using the dot notation.

In the object.property syntax, the property must be a valid JavaScript identifier. (In the ECMAScript standard, the names of properties are technically "IdentifierNames", not "Identifiers", so reserved words can be used but are not recommended).

One can think of an object as an associative array (a.k.a. map, dictionary, hash, lookup table). The keys in this array are the names of the object's properties.

It's typical when speaking of an object's properties to make a distinction between properties and methods. However, the property/method distinction is little more than a convention. A method is simply a property that can be called (for example, if it has a reference to a Function instance as its value).

More information:

ECMAScript reference

MDN documentation

GetField(Box<Node>, Box<Node>)

This property accessor provides access to an object's properties by using the bracket notation.

In the object[property_name] syntax, the property_name is just a string or Symbol. So, it can be any string, including '1foo', '!bar!', or even ' ' (a space).

One can think of an object as an associative array (a.k.a. map, dictionary, hash, lookup table). The keys in this array are the names of the object's properties.

It's typical when speaking of an object's properties to make a distinction between properties and methods. However, the property/method distinction is little more than a convention. A method is simply a property that can be called (for example, if it has a reference to a Function instance as its value).

More information:

ECMAScript reference

MDN documentation

ForLoop(Option<Box<Node>>, Option<Box<Node>>, Option<Box<Node>>, Box<Node>)

The for statement creates a loop that consists of three optional expressions.

A for loop repeats until a specified condition evaluates to false. The JavaScript for loop is similar to the Java and C for loop.

More information:

ECMAScript reference

MDN documentation

If(Box<Node>, Box<Node>, Option<Box<Node>>)

The if statement executes a statement if a specified condition is truthy. If the condition is falsy, another statement can be executed.

Multiple if...else statements can be nested to create an else if clause.

Note that there is no elseif (in one word) keyword in JavaScript.

More information:

ECMAScript reference

MDN documentation

LetDecl(Vec<(String, Option<Node>)>)

The let statement declares a block scope local variable, optionally initializing it to a value.

let allows you to declare variables that are limited to a scope of a block statement, or expression on which it is used, unlike the var keyword, which defines a variable globally, or locally to an entire function regardless of block scope.

Just like const the let does not create properties of the window object when declared globally (in the top-most scope).

More information:

ECMAScript reference

MDN documentation

Local(String)

An identifier is a sequence of characters in the code that identifies a variable, function, or property.

In JavaScript, identifiers are case-sensitive and can contain Unicode letters, $, _, and digits (0-9), but may not start with a digit.

An identifier differs from a string in that a string is data, while an identifier is part of the code. In JavaScript, there is no way to convert identifiers to strings, but sometimes it is possible to parse strings into identifiers.

More information:

ECMAScript reference

MDN documentation

New(Box<Node>)

The new operator lets developers create an instance of a user-defined object type or of one of the built-in object types that has a constructor function.

The new keyword does the following things:

Creates a blank, plain JavaScript object;

Links (sets the constructor of) this object to another object;

Passes the newly created object from Step 1 as the this context;

Returns this if the function doesn't return its own object.

More information:

ECMAScript reference

MDN documentation

Object(Vec<PropertyDefinition>)

Objects in JavaScript may be defined as an unordered collection of related data, of primitive or reference types, in the form of “key: value” pairs.

Objects can be initialized using new Object(), Object.create(), or using the literal notation.

An object initializer is an expression that describes the initialization of an Object. Objects consist of properties, which are used to describe an object. Values of object properties can either contain primitive data types or other objects.

More information:

ECMAScript reference

MDN documentation

Return(Option<Box<Node>>)

The return statement ends function execution and specifies a value to be returned to the function caller.

Syntax: return [expression];

expression:

The expression whose value is to be returned. If omitted, undefined is returned instead.

When a return statement is used in a function body, the execution of the function is stopped. If specified, a given value is returned to the function caller.

More information:

ECMAScript reference

MDN documentation

Switch(Box<Node>, Vec<(Node, Vec<Node>)>, Option<Box<Node>>)

The switch statement evaluates an expression, matching the expression's value to a case clause, and executes statements associated with that case, as well as statements in cases that follow the matching case.

A switch statement first evaluates its expression. It then looks for the first case clause whose expression evaluates to the same value as the result of the input expression (using the strict comparison, ===) and transfers control to that clause, executing the associated statements. (If multiple cases match the provided value, the first case that matches is selected, even if the cases are not equal to each other.)

More information:

ECMAScript reference

MDN documentation

Spread(Box<Node>)

The spread operator allows an iterable such as an array expression or string to be expanded.

Syntax: ...x

It expands array expressions or strings in places where zero or more arguments (for function calls) or elements (for array literals) are expected, or an object expression to be expanded in places where zero or more key-value pairs (for object literals) are expected.

More information:

ECMAScript reference

MDN documentation

StatementList(Vec<Node>)

Similar to Node::Block but without the braces

More information:

ECMAScript reference

Throw(Box<Node>)

The throw statement throws a user-defined exception.

Syntax: throw expression;

Execution of the current function will stop (the statements after throw won't be executed), and control will be passed to the first catch block in the call stack. If no catch block exists among caller functions, the program will terminate.

More information:

ECMAScript reference

MDN documentation

TypeOf(Box<Node>)

The typeof operator returns a string indicating the type of the unevaluated operand.

Syntax: typeof operand

Returns a string indicating the type of the unevaluated operand.

More information:

ECMAScript reference

MDN documentation

Try(Box<Node>, Option<Box<Node>>, Option<Box<Node>>, Option<Box<Node>>)

The try...catch statement marks a block of statements to try and specifies a response should an exception be thrown.

The try statement consists of a try-block, which contains one or more statements. {} must always be used, even for single statements. At least one catch-block, or a finally-block, must be present.

More information:

ECMAScript reference

MDN documentation

This

The JavaScript this keyword refers to the object it belongs to.

A property of an execution context (global, function or eval) that, in non–strict mode, is always a reference to an object and in strict mode can be any value.

More information:

ECMAScript reference

MDN documentation

UnaryOp(UnaryOp, Box<Node>)

A unary operation is an operation with only one operand.

More information:

ECMAScript reference

MDN documentation

VarDecl(Vec<(String, Option<Node>)>)

The var statement declares a variable, optionally initializing it to a value.

var declarations, wherever they occur, are processed before any code is executed. This is called hoisting, and is discussed further below.

The scope of a variable declared with var is its current execution context, which is either the enclosing function or, for variables declared outside any function, global. If you re-declare a JavaScript variable, it will not lose its value.

Assigning a value to an undeclared variable implicitly creates it as a global variable (it becomes a property of the global object) when the assignment is executed.

More information:

ECMAScript reference

MDN documentation

WhileLoop(Box<Node>, Box<Node>)

The while statement creates a loop that executes a specified statement as long as the test condition evaluates to true.

The condition is evaluated before executing the statement.

More information:

ECMAScript reference

MDN documentation

Methods

impl Node

pub fn array_decl<N>(nodes: N) -> Self where
N: Into<Vec<Self>>,

Creates an ArrayDecl AST node.

pub fn arrow_function_decl<P, B>(params: P, body: B) -> Self where
P: Into<Vec<FormalParameter>>,
B: Into<Box<Self>>,

Creates an ArraowFunctionDecl AST node.

pub fn assign<L, R>(lhs: L, rhs: R) -> Self where
L: Into<Box<Self>>,
R: Into<Box<Self>>,

Creates an Assign AST node.

pub fn bin_op<O, L, R>(op: O, lhs: L, rhs: R) -> Self where
O: Into<BinOp>,
L: Into<Box<Self>>,
R: Into<Box<Self>>,

Creates a BinOp AST node.

pub fn block<N>(nodes: N) -> Self where
N: Into<Vec<Self>>,

Creates a Block AST node.

pub fn break_node<OL, L>(label: OL) -> Self where
L: Into<String>,
OL: Into<Option<L>>,

Creates a Break AST node.

pub fn call<F, P>(function: F, params: P) -> Self where
F: Into<Box<Self>>,
P: Into<Vec<Self>>,

Creates a Call AST node.

pub fn conditional_op<C, T, F>(condition: C, if_true: T, if_false: F) -> Self where
C: Into<Box<Self>>,
T: Into<Box<Self>>,
F: Into<Box<Self>>,

Creates a ConditionalOp AST node.

pub fn const_node<C>(node: C) -> Self where
C: Into<Const>,

Creates a Const AST node.

pub fn const_decl<D>(decl: D) -> Self where
D: Into<Vec<(String, Self)>>,

Creates a ConstDecl AST node.

pub fn continue_node<OL, L>(label: OL) -> Self where
L: Into<String>,
OL: Into<Option<L>>,

Creates a Continue AST node.

pub fn do_while_loop<B, C>(body: B, condition: C) -> Self where
B: Into<Box<Self>>,
C: Into<Box<Self>>,

Creates a DoWhileLoop AST node.

pub fn function_decl<ON, N, P, B>(name: ON, params: P, body: B) -> Self where
N: Into<String>,
ON: Into<Option<N>>,
P: Into<Vec<FormalParameter>>,
B: Into<Box<Self>>,

Creates a FunctionDecl AST node.

pub fn get_const_field<V, L>(value: V, label: L) -> Self where
V: Into<Box<Self>>,
L: Into<String>,

Creates a GetConstField AST node.

pub fn get_field<V, F>(value: V, field: F) -> Self where
V: Into<Box<Self>>,
F: Into<Box<Self>>,

Creates a GetField AST node.

pub fn for_loop<OI, OC, OS, I, C, S, B>(
init: OI,
condition: OC,
step: OS,
body: B
) -> Self where
OI: Into<Option<I>>,
OC: Into<Option<C>>,
OS: Into<Option<S>>,
I: Into<Box<Self>>,
C: Into<Box<Self>>,
S: Into<Box<Self>>,
B: Into<Box<Self>>,

Creates a ForLoop AST node.

pub fn if_node<C, B, E, OE>(condition: C, body: B, else_node: OE) -> Self where
C: Into<Box<Self>>,
B: Into<Box<Self>>,
E: Into<Box<Self>>,
OE: Into<Option<E>>,

Creates an If AST node.

pub fn let_decl<I>(init: I) -> Self where
I: Into<Vec<(String, Option<Self>)>>,

Creates a LetDecl AST node.

pub fn local<N>(name: N) -> Self where
N: Into<String>,

Creates a Local AST node.

pub fn new<N>(node: N) -> Self where
N: Into<Box<Self>>,

Creates a New AST node.

pub fn object<D>(def: D) -> Self where
D: Into<Vec<PropertyDefinition>>,

Creates an Object AST node.

pub fn return_node<E, OE>(expr: OE) -> Self where
E: Into<Box<Self>>,
OE: Into<Option<E>>,

Creates a Return AST node.

pub fn switch<V, C, OD, D>(val: V, cases: C, default: OD) -> Self where
V: Into<Box<Self>>,
C: Into<Vec<(Self, Vec<Self>)>>,
OD: Into<Option<D>>,
D: Into<Box<Self>>,

Creates a Switch AST node.

pub fn spread<V>(val: V) -> Self where
V: Into<Box<Self>>,

Creates a Spread AST node.

pub fn statement_list<L>(list: L) -> Self where
L: Into<Vec<Self>>,

Creates a StatementList AST node.

pub fn throw<V>(val: V) -> Self where
V: Into<Box<Self>>,

Creates a Throw AST node.

pub fn type_of<E>(expr: E) -> Self where
E: Into<Box<Self>>,

Creates a TypeOf AST node.

pub fn try_node<T, OC, OP, OF, C, P, F>(
try_node: T,
catch: OC,
param: OP,
finally: OF
) -> Self where
T: Into<Box<Self>>,
OC: Into<Option<C>>,
OP: Into<Option<P>>,
OF: Into<Option<F>>,
C: Into<Box<Self>>,
P: Into<Box<Self>>,
F: Into<Box<Self>>,

Creates a Try AST node.

pub fn this() -> Self

Creates a This AST node.

pub fn unary_op<V>(op: UnaryOp, val: V) -> Self where
V: Into<Box<Self>>,

Creates a UnaryOp AST node.

pub fn var_decl<I>(init: I) -> Self where
I: Into<Vec<(String, Option<Self>)>>,

Creates a VarDecl AST node.

pub fn while_loop<C, B>(condition: C, body: B) -> Self where
C: Into<Box<Self>>,
B: Into<Box<Self>>,

Creates a WhileLoop AST node.

fn display(&self, f: &mut Formatter, indentation: usize) -> Result

Implements the display formatting with indentation.