Enum boa::syntax::ast::node::Node

pub enum Node {
    ArrayDecl(ArrayDecl),
    ArrowFunctionDecl(ArrowFunctionDecl),
    Assign(Assign),
    BinOp(BinOp),
    Block(Block),
    Break(Break),
    Call(Call),
    ConditionalOp(ConditionalOp),
    Const(Const),
    ConstDeclList(ConstDeclList),
    Continue(Continue),
    DoWhileLoop(DoWhileLoop),
    FunctionDecl(FunctionDecl),
    FunctionExpr(FunctionExpr),
    GetConstField(GetConstField),
    GetField(GetField),
    ForLoop(ForLoop),
    If(If),
    LetDeclList(LetDeclList),
    Identifier(Identifier),
    New(New),
    Object(Object),
    Return(Return),
    Switch(Switch),
    Spread(Spread),
    Throw(Throw),
    Try(Try),
    This,
    UnaryOp(UnaryOp),
    VarDeclList(VarDeclList),
    WhileLoop(WhileLoop),
}

Variants

ArrayDecl(ArrayDecl)

Array declaration node. More information.

ArrowFunctionDecl(ArrowFunctionDecl)

An arrow function expression node. More information.

Assign(Assign)

An assignment operator node. More information.

BinOp(BinOp)

A binary operator node. More information.

Block(Block)

A Block node. More information.

Break(Break)

A break node. More information.

Call(Call)

A function call. More information.

ConditionalOp(ConditionalOp)

A javascript conditional operand ( x ? y : z ). More information.

Const(Const)

Literals represent values in JavaScript.

These are fixed values not variables that you literally provide in your script.

More information:

ECMAScript reference

MDN documentation

ConstDeclList(ConstDeclList)

A constant declaration list. More information.

Continue(Continue)

A continue statement. More information.

DoWhileLoop(DoWhileLoop)

A do ... while statement. More information.

FunctionDecl(FunctionDecl)

A function declaration node. More information.

FunctionExpr(FunctionExpr)

A function expressino node. More information.

GetConstField(GetConstField)

Provides access to an object types' constant properties. More information.

GetField(GetField)

Provides access to object fields. More information.

ForLoop(ForLoop)

A for statement. More information.

If(If)

An 'if' statement. More information.

LetDeclList(LetDeclList)

A let declaration list. More information.

Identifier(Identifier)

A local identifier node. More information.

New(New)

A new expression. More information.

Object(Object)

An object. More information.

Return(Return)

A return statement. More information.

Switch(Switch)

A switch {case} statement. More information.

Spread(Spread)

A spread (...x) statement. More information.

Throw(Throw)

A throw statement. More information.

Try(Try)

A try...catch node. More information.

This

The JavaScript this keyword refers to the object it belongs to.

A property of an execution context (global, function or eval) that, in non–strict mode, is always a reference to an object and in strict mode can be any value.

More information:

ECMAScript reference

MDN documentation

UnaryOp(UnaryOp)

Unary operation node. More information

VarDeclList(VarDeclList)

Array declaration node. More information.

WhileLoop(WhileLoop)

A 'while {...}' node. More information.

Implementations

impl Node

pub(crate) fn hoistable_order(a: &Node, b: &Node) -> Ordering

Returns a node ordering based on the hoistability of each node.

pub fn this() -> Self

Creates a This AST node.

fn display(&self, f: &mut Formatter, indentation: usize) -> Result

Implements the display formatting with indentation.