Struct boa::syntax::lexer::Lexer

pub struct Lexer<R> {
    cursor: Cursor<R>,
    goal_symbol: InputElement,
}

Lexer or tokenizer for the Boa JavaScript Engine.

Fields
cursor: Cursor<R>goal_symbol: InputElement
Implementations

impl<R> Lexer<R>

fn is_whitespace(ch: u32) -> bool

Checks if a character is whitespace as per ECMAScript standards.

The Rust char::is_whitespace function and the ECMAScript standard use different sets of characters as whitespaces:

Rust uses \p{White_Space},

ECMAScript standard uses \{Space_Separator} + \u{0009}, \u{000B}, \u{000C}, \u{FEFF}

More information

pub(crate) fn set_goal(&mut self, elm: InputElement)

Sets the goal symbol for the lexer.

pub(crate) fn get_goal(&self) -> InputElement

Gets the goal symbol the lexer is currently using.

pub(super) fn strict_mode(&self) -> bool

pub(super) fn set_strict_mode(&mut self, strict_mode: bool)

pub fn new(reader: R) -> Self where
R: Read,

Creates a new lexer.

pub(crate) fn lex_slash_token(
&mut self,
start: Position
) -> Result<Token, Error> where
R: Read,

pub fn next(&mut self) -> Result<Option<Token>, Error> where
R: Read,

Retrieves the next token from the lexer.

pub(crate) fn lex_template(&mut self, start: Position) -> Result<Token, Error> where
R: Read,