Trait boa::syntax::parser::TokenParser

trait TokenParser<R>: Sized where
    R: Read, {
    type Output;
    fn parse(self, cursor: &mut Cursor<R>) -> Result<Self::Output, ParseError>;
}

Trait implemented by parsers.

This makes it possible to abstract over the underlying implementation of a parser.

Associated Types

type Output

Output type for the parser.

Required methods

fn parse(self, cursor: &mut Cursor<R>) -> Result<Self::Output, ParseError>

Parses the token stream using the current parser.

This method needs to be provided by the implementor type.